% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{add_derived_matrix}
\alias{add_derived_matrix}
\alias{add_derived_matrices}
\title{Add Derived Matrices}
\usage{
add_derived_matrix(model, mat)

add_derived_matrices(model, mat_list)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{mat}{object of \code{EpiMatrixDerived-class}}

\item{mat_list}{object of class \code{EpiMatrixList}, containing at least
some objects of \code{EpiMatrixDerived-class}}
}
\value{
\code{\link{flexmodel}} object with intermediate computations
that could be used in calculating rates and/or including in simulation
output
}
\description{
Add Derived Matrices
}
