% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_factr}
\alias{add_factr}
\title{Intermediate Factors}
\usage{
add_factr(model, factr_nm, formula)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{factr_nm}{name of the new intermediate factor}

\item{formula}{formula (or string) that follows
this spec, \url{https://canmod.net/misc/flex_specs#v0.1.0},
or 1-by-1 \code{\link{struc}} object describing the
intermediate factor}
}
\value{
\code{\link{flexmodel}} object
}
\description{
Save intermediate computations so that they can be used
as factors in rate expressions and returned as part
of the simulation history. The definition of a factor
is given here: \url{https://canmod.net/misc/flex_specs#v0.1.0}.
}
\seealso{
See \code{\link{vec_factr}} to add more than
one intermediate factor at the same time.

Other flexmodel_definition_functions: 
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{flexmodel}()},
\code{\link{initial_population}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{update_disease_free_state}()},
\code{\link{update_linearized_params}()},
\code{\link{vec_rate}()}
}
\concept{flexmodel_definition_functions}
