% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_lag_diff}
\alias{add_lag_diff}
\alias{add_lag_diff_uneven}
\title{Add Variable by Lag Differencing}
\usage{
add_lag_diff(model, var_pattern, delay_n = 1)

add_lag_diff_uneven(
  model,
  input_names,
  output_names,
  lag_dates,
  method = c("uneven", "lag_one_day")
)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{var_pattern}{regular expression used to identify variables for
differencing. See \code{\link{avail_for_lag}} for a function that will
return the names of all variables that are available for differencing.}

\item{delay_n}{Delay in days for determining the lag in the differences.}

\item{input_names}{names of variables to be differenced}

\item{output_names}{names of the result of differencing}

\item{lag_dates}{Dates between which differences should be taken.}

\item{method}{method of producing lags using generic
\code{add_lag_diff_uneven}. The default, \code{"uneven"} requires
a vector dates at which to difference, and the alternative allows
for automatic generation of this list of dates for creating the
standard difference lag of one day, \code{"lag_one_day"}.}
}
\description{
Add a variable that will be available for comparisons with observed data
that is created by processing an existing variable in the simulation
history by lagged differencing.
}
