% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_opt_params}
\alias{update_opt_params}
\alias{add_opt_params}
\title{Update Optimization Parameters}
\usage{
update_opt_params(model, ...)

add_opt_params(model, ...)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{...}{a list of formulas for describing what parameters should
be optimized, whether/how they should be transformed before being
passed to the objective function, and what prior distribution (or
regularization function should be used -- see the section on
the formula syntax for more details.}
}
\description{
Add or update parameters to be optimized/calibrated in a
\code{flexmodel} object.
}
\section{Formula Syntax}{

The left-hand-side of the formula describes the parameters and how they
should be transformed, and the right-hand-side describes the associated
prior distribution.

The simplest approach is to specify one parameter at a time with the
following syntax:

\code{trans_param ~ trans_prior(hyperparameters ...)}

On the left-hand-side the optional transformation, \code{trans}, is
separated by an underscore from the parameter name, \code{param}. For
example one might optimize the transmission rate, \code{beta}, on the
log scale by specifying the left-hand-side as \code{log_beta}. To
optimize \code{beta} on the untransformed scale the left-hand-side
would be simply be \code{beta}. The currently available transformations
are \code{log}, \code{log10}, \code{logit}, \code{cloglog}, \code{inverse}.

On the right-hand-side the optional transformation, \code{trans}, is
separated from the name of the prior family by an underscore. This
transformation defines the scale on which the prior distribution is over.
Currently the transformation scale passed to the objective function
(on the left-hand-side) must match the transformation scale of the
priot (on the right-hand-side) -- perhaps this restriction will be
lifted one day. The currently available prior families are \code{flat}
(for no regularization) and \code{normal}. These families are
expressed as functions of hyperparameters where the first argument
is a location parameter, which also defines the starting point of the
optimizer.

\describe{
  \item{\code{flat}}{
    \describe{
      \item{initial}{the 'peak' of a flat function -- important because it sets the initial value of the optimizer.}
    }
  }
  \item{\code{normal}}{
      \describe{
        \item{mean}{mean of the normal distribution, also used as the initial value of the optimizer}
        \item{standard deviation}{standard deviation of the normal distribution}
      }
    }
}

One may also specify several parameters with the same
}

