% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_outflow}
\alias{add_outflow}
\title{Add Outflows}
\usage{
add_outflow(model, state_patterns = ".+", flow_state_patterns = ".+")
}
\description{
Add outflows corresponding to inflows specified by
\code{add_rate}, \code{rep_rate}, and \code{vec_rate}.
}
\details{
By default all inflows will have corresponding outflows.
To define outflows from specific states, pass a regular
expression to \code{state_patterns} that identifies these
states. To restrict outflows to those going to specific
states, use regular expressions to define these states.
}
\seealso{
Other flexmodels: 
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_tmb_indices}()},
\code{\link{init_model}()},
\code{\link{initial_population}()},
\code{\link{make_base_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{state_param_sum}()},
\code{\link{tmb_fun}()},
\code{\link{vec_rate}()}
}
\concept{flexmodels}
