% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_parallel_accumulators}
\alias{add_parallel_accumulators}
\title{Add Parallel Accumulators}
\usage{
add_parallel_accumulators(model, state_patterns)
}
\arguments{
\item{model}{TODO}

\item{state_patterns}{regular expressions for identifying states as
parallel accumulators}
}
\value{
another compartmental model with parallel accumulators specified
}
\description{
Add parallel accumulators to a compartmental model.
}
\seealso{
Other flexmodels: 
\code{\link{add_outflow}()},
\code{\link{add_rate}()},
\code{\link{add_tmb_indices}()},
\code{\link{init_model}()},
\code{\link{initial_population}()},
\code{\link{make_base_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{state_param_sum}()},
\code{\link{tmb_fun}()},
\code{\link{vec_rate}()}
}
\concept{flexmodels}
