% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_rate}
\alias{add_rate}
\title{Rate Structure}
\usage{
add_rate(model, from, to, formula)
}
\arguments{
\item{model}{compartmental model}

\item{from}{Name of state where flow is happening from}

\item{to}{Name of state where flow is happening to}

\item{formula}{Model formula defining dependence of the rate on
parameters and state variables}
}
\value{
another compartmental model with an additional non-zero rate matrix
element specified
}
\description{
Define how the rate of flow from one compartment to another
depends on the parameters and state variables.
}
