% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_sim_report_expr}
\alias{add_sim_report_expr}
\title{Add Expression to the Simulation History}
\usage{
add_sim_report_expr(model, expr_nm, formula)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{expr_nm}{Name of the new variable}

\item{formula}{formula describing the expression. See
\code{\link{avail_for_expr}} for
a function that will return the names of all variables that are available
for use in these formulas.}
}
\value{
updated \code{\link{flexmodel}} object
}
\description{
Create a new variable in the \code{\link{simulation_history}}
by taking sums and products of existing variables (and their
complements and inverses). These new variables can be compared
with observed data streams.
}
