% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_tmb_indices}
\alias{add_tmb_indices}
\title{Add TMB Indices}
\usage{
add_tmb_indices(model)
}
\arguments{
\item{model}{compartmental model}

\item{another}{compartmental model with indices for TMB}
}
\description{
Add, to a compartmental model, indices used to access appropriate values
during simulation and calibration using TMB
}
