% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{add_updated_vaxrate}
\alias{add_updated_vaxrate}
\title{compute and add vaxrates to ratemat
(returns whole ratemat because the update part is non-trivial)}
\usage{
add_updated_vaxrate(state, params, ratemat)
}
\arguments{
\item{state}{state vector (an object of class `state_pansim`)}

\item{params}{model parameters (an object of class `params_pansim`)}

\item{ratemat}{model rate matrix}
}
\description{
compute and add vaxrates to ratemat
(returns whole ratemat because the update part is non-trivial)
}
