% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{aggregate.pansim}
\alias{aggregate.pansim}
\title{Collapse columns (infected, ICU, hospitalized) in a pansim output}
\usage{
\method{aggregate}{pansim}(x)
}
\arguments{
\item{x}{a pansim object}
}
\description{
Collapse columns (infected, ICU, hospitalized) in a pansim output
}
