% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{aggregate_agecats}
\alias{aggregate_agecats}
\title{Aggregate age categories into larger age categories}
\usage{
aggregate_agecats(age, bin_edges_lower)
}
\arguments{
\item{age}{Numeric or factor vector of ages}

\item{bin_edges_lower}{list of numeric values for the lower edges of desired
age group bins in panel 2. If ages are of type factor, all user-specified
lower bin edges must match lower bin edges in the age data.}
}
\description{
A function that takes a vector of ages or age groups (either numeric or
factor) and returns assigned age categories based on desired lower bin edges
provided by the user.
}
\examples{
aggregate_agecats(seq(0, 90, by = 1), c(25, 45))
}
