% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{as.character.struc}
\alias{as.character.struc}
\title{Convert struc Object to a character vector}
\usage{
\method{as.character}{struc}(x, ...)
}
\arguments{
\item{x}{\code{\link{struc-class}} object}

\item{...}{for S3 method consistency}
}
\value{
character vector
}
\description{
Convert struc Object to a character vector
}
