% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmodel_convert.R
\name{as.flexmodel}
\alias{as.flexmodel}
\alias{as.flexmodel_to_calibrate}
\alias{as.flexmodel_obs_error}
\alias{as.flexmodel_one_step}
\title{Convert between Types of Flexmodels}
\usage{
as.flexmodel(model)

as.flexmodel_to_calibrate(model)

as.flexmodel_obs_error(model)

as.flexmodel_one_step(model)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}
}
\value{
another \code{\link{flexmodel}} object of a different type
}
\description{
Convert between Types of Flexmodels
}
