% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{as.matrix.struc}
\alias{as.matrix.struc}
\title{Convert struc Object to a matrix}
\usage{
\method{as.matrix}{struc}(x)
}
\arguments{
\item{x}{struc object}
}
\value{
matrix
}
\description{
Convert struc Object to a matrix
}
