% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{avail_for_rate}
\alias{avail_for_rate}
\alias{avail_for_sum}
\alias{avail_for_factr}
\alias{avail_for_expr}
\alias{avail_for_conv}
\alias{avail_for_lag}
\title{What Variables are Available?}
\usage{
avail_for_rate(model)

avail_for_sum(model)

avail_for_factr(model)

avail_for_expr(model)

avail_for_conv(model)

avail_for_lag(model)
}
\arguments{
\item{model}{a \code{\link{flexmodel}} object}
}
\value{
character vector of available variable names
}
\description{
Find out what variables are available for particular functions that
add variables to a model, based on existing variables.
}
\details{
\describe{
  \item{\code{avail_for_rate}}{\code{\link{add_rate}},\code{\link{rep_rate}},\code{\link{vec_rate}}}
  \item{\code{avail_for_sum}}{\code{\link{add_state_param_sum}}}
  \item{\code{avail_for_factr}}{\code{\link{add_factr}},\code{\link{vec_factr}}}
  \item{\code{avail_for_expr}}{\code{\link{add_sim_report_expr}}}
  \item{\code{avail_for_lag}}{\code{\link{add_lag_diff}}}
  \item{\code{avail_for_conv}}{\code{\link{add_conv}}}
}
}
