% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{browserManager}
\alias{browserManager}
\title{Handle external browser options for the shiny.}
\usage{
browserManager(openinBrowser)
}
\arguments{
\item{openinBrowser}{a logical indicating whether the shiny app should be opened in the browser or not}
}
\description{
Handle external browser options for the shiny.
}
