% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calc_conv}
\alias{calc_conv}
\title{calculate convolution}
\usage{
calc_conv(i, params)
}
\arguments{
\item{i}{an incidence time series}

\item{params}{a list or vector containing elements \code{c_prop}, \code{c_delay_mean}, \code{c_delay_cv}}
}
\description{
calculate convolution
}
