% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calib2}
\alias{calib2}
\title{Calibrate parameters}
\usage{
calib2(
  base_params,
  target = c(Gbar = 6, r = 0.23, X0 = 200),
  start_date,
  reg_date,
  init_var = "H",
  sim_args = NULL
)
}
\arguments{
\item{base_params}{base parameters}

\item{target}{calibration values: mean generation interval, initial growth rate, initial value of calibration variable (\code{X0})}

\item{start_date}{starting date for simulations (should be far enough before the start of the data to allow sorting-out of initial conditions)}

\item{reg_date}{first day of calibration data used in log-linear regression}

\item{init_var}{variable to use in initial-condition calibration matching}

\item{sim_args}{list of additional arguments to pass to \code{\link{run_sim}}}
}
\description{
Calibrate parameters
}
