% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{given a log-linear intercept and slope and target values for R0 or GI,
return initial conditions and parameters that will (approximately?)
recapitulate the observed log-linear dynamics}
\usage{
calibrate(
  int,
  slope,
  pop,
  params,
  R0 = 3,
  GI = NULL,
  date0 = ldmy("1-Mar-2020"),
  date1 = ldmy("25-Mar-2020"),
  var = "H"
)
}
\arguments{
\item{int}{log-linear intercept (expected value on date1)}

\item{slope}{log-linear slope (growth rate)}

\item{R0}{desired/calibrated R0}

\item{GI}{desired/calibrated mean generation interval (stub)}

\item{date0}{date for initial conditions}

\item{date1}{initial date for regression}

\item{var}{variable for regression (H, D, ICU)}
}
\description{
given a log-linear intercept and slope and target values for R0 or GI,
return initial conditions and parameters that will (approximately?)
recapitulate the observed log-linear dynamics
}
