% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_comb}
\alias{calibrate_comb}
\title{top-level calibration based on mobility, splines, phenom het}
\usage{
calibrate_comb(
  data,
  params,
  opt_pars = NULL,
  mob_data = NULL,
  mob_breaks = NULL,
  spline_days = 14,
  spline_df = NA,
  knot_quantile_var = NA,
  spline_pen = 0,
  maxit = 10000,
  use_DEoptim = TRUE,
  DE_cores = 1,
  use_mobility = FALSE,
  use_phenomhet = FALSE,
  use_spline = TRUE,
  vars = NULL,
  debug_plot = interactive(),
  debug = FALSE,
  return_X = FALSE,
  ...
)
}
\arguments{
\item{data}{a data set to compare to, containing date/var/value (current version assumes that only a single state var is included)}

\item{params}{parameters}

\item{opt_pars}{starting parameters (and structure).  Parameters that are part of the \code{params_pansim} parameter vector can be specified within the \code{params} element (with prefixes if they are transformed); other parameters can include distributional parameters or time-varying parameters}

\item{mob_data}{mobility data}

\item{spline_days}{days between spline knots}

\item{spline_df}{overall spline degrees of freedom}

\item{knot_quantile_var}{variable to use cum dist for knotspacing}

\item{spline_pen}{penalization for spline}

\item{maxit}{maximum iterations for Nelder-Mead/optimization step}

\item{use_DEoptim}{use differential evolution as first stage?}

\item{DE_cores}{number of parallel workers for DE}

\item{use_mobility}{include mobility as a covariate in the model?}

\item{use_phenomhet}{include phenomenological heterogeneity?}

\item{use_spline}{include spline?}

\item{vars}{which vars to use? (default is all in data)}

\item{debug_plot}{plot debugging curves? (doesn't work with parallel DEoptim)}

\item{debug}{print debugging messages?}

\item{return_X}{short-circuit/return model matrix?}

\item{...}{extra args}

\item{mob_brks}{vector of breakpoints for piecewise mobility model}
}
\description{
top-level calibration based on mobility, splines, phenom het
}
