% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_flexmodel}
\alias{calibrate_flexmodel}
\title{Calibrate Flexmodel}
\usage{
calibrate_flexmodel(
  model,
  .optimizer_wrapper = c("bbmle", "nlminb", "optim"),
  ...
)
}
\arguments{
\item{model}{\code{flexmodel_to_calibrate} object, which is a
\code{\link{flexmodel}} object that contains observed data added either
throught the \code{data} argument in \code{\link{flexmodel}} or with the
\code{\link{update_observed}} function}

\item{.optimizer_wrapper}{name of a valid optimizer -- currently \code{bbmle} is
recommended (see details for more information about the choices)}

\item{...}{additional arguments to pass on to the optimizer}
}
\value{
\code{flexmodel_calibrated} object
}
\description{
Calibrate Flexmodel
}
\details{
Currently there are the following three optimizer wrappers.

\describe{
  \item{\code{\link{bbmle_flexmodel}}}{
    Wraps the \code{mle2} function in the \code{bbmle} package. This is
    the recommended optimizer because it provides the ability to create
    ensemble forecasts through \code{\link{simulate_ensemble}}. Such
    forecasts are possible because of the functionality provided in
    \code{bbmle} to generate from the sampling distribution of the
    parameters. The variation induced by this sampling distribution is
    used to create a distribution of simulation trajectories and therefore
    confidence envelopes around forecasts or fits. The main limitation
    of \code{\link{bbmle_flexmodel}} is that it doesn't make use of
    the Hessian matrix that is provided by \code{TMB} through
    auto-differentiation, although it does make use of the gradient.
  }
  \item{\code{\link{nlminb_flexmodel}}}{
    Wraps the \code{\link{nlminb_flexmodel}} function. The main advantage
    of this optimizer is that it allows one to make use of the Hessian
    matrix (contrast with \code{\link{bbmle_flexmodel}}). The main
    disadvantage is that \code{\link{nlminb_flexmodel}} does not allow
    generating samples from the sampling distribution of the parameters,
    and therefore does not allow for ensemble simulation.
  }
  \item{\code{\link{optim_flexmodel}}}{
    Wraps the standard optimizer in R, but neither allows the use of
    Hessians nor ensemble forecasts.
  }
}
}
\seealso{
\code{\link{bbmle_flexmodel}}, \code{\link{nlminb_flexmodel}},
\code{\link{optim_flexmodel}}
}
