% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_old.R
\name{calibrate_slopeint}
\alias{calibrate_slopeint}
\title{given a log-linear intercept and slope and target values for R0 or GI,
return initial conditions and parameters that will (approximately?)
recapitulate the observed log-linear dynamics}
\usage{
calibrate_slopeint(
  int,
  slope,
  pop,
  params,
  target = c(Gbar = 6),
  init_target = NULL,
  date0 = "1-Mar-2020",
  date1 = "25-Mar-2020",
  var = "H",
  sim_args = NULL
)
}
\arguments{
\item{int}{log-linear intercept (expected value on date1)}

\item{slope}{log-linear slope (growth rate)}

\item{pop}{population size}

\item{params}{a parameter set}

\item{target}{vector of target statistics (allowed: R0, Gbar (mean generation interval), or kappa (CV^2 of generation interval)}

\item{init_target}{value of focal variable to hit on date1}

\item{date0}{date for initial conditions}

\item{date1}{initial date for regression}

\item{var}{variable for regression (H, D, ICU)}

\item{sim_args}{additional parameters to pass to \code{run_sim} if shooting}
}
\description{
given a log-linear intercept and slope and target values for R0 or GI,
return initial conditions and parameters that will (approximately?)
recapitulate the observed log-linear dynamics
}
\examples{
params <- read_params("ICU1.csv")
cc1 <- calibrate_slopeint(int=log(200),slope=0.23,pop=1e6,params)
summary(cc1$params)
cc2 <- calibrate_slopeint(int=log(200),slope=0.23,pop=1e6,params,target=c(R0=3))
summary(cc2$params)

}
