% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{check_age_cat_compatibility}
\alias{check_age_cat_compatibility}
\title{Check compatibility of user-specified age categories with age categories in
Mistry data}
\usage{
check_age_cat_compatibility(age_cat)
}
\arguments{
\item{age_cat}{age categories (e.g. generated by `mk_agecats()`)}
}
\description{
Check compatibility of user-specified age categories with age categories in
Mistry data
}
\examples{
check_age_cat_compatibility(mk_agecats(min = 0, max = 80, da = 2))
}
