% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{check_contact_rate_setting}
\alias{check_contact_rate_setting}
\title{Helper function to check setting-specific contact rate list initialization}
\usage{
check_contact_rate_setting(contact_rate_setting)
}
\arguments{
\item{contact_rate_setting}{named list containing setting-specific contact rates (in units of average contacts in the given setting per individual of age i with individuals of age j per day)}
}
\description{
Helper function to check setting-specific contact rate list initialization
}
\examples{
check_contact_rate_setting(mk_contact_rate_setting())
}
