% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{color_list}
\alias{color_list}
\title{Simulate ggplot picking colours for a plot.}
\usage{
color_list(n)
}
\arguments{
\item{n}{the number of variables in the simulated plot. Should match the number of variables in the simulation for this to be useful.}
}
\value{
hues, a list of the colours (specified in hex) that ggplot2 will use.
}
\description{
Obtain a list of the colours that ggplot will use for a plot with n variable.
}
