% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{colour_curve}
\alias{colour_curve}
\title{Write checkbox colouring HTML}
\usage{
colour_curve(curve)
}
\arguments{
\item{curves}{a list of input IDs of the checkboxes to modify the fill of.}

\item{colourList}{a list of hex obects specifiying the colour to change corresponding checkboxes in curves to.}
}
\value{
a tagsList containing the tags
}
\description{
Generate the necessary HTML/CSS tags to change the colour of a checkbox.
}
