% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{compare_sims}
\alias{compare_sims}
\alias{compare_sims_cbind}
\alias{compare_sims_plot}
\alias{compare_sims_diffmat}
\alias{compare_sims_unlist}
\title{Compare TMB-based and classic MacPan Simulations}
\usage{
compare_sims(
  classic_sim,
  tmb_sim,
  tolerance = NULL,
  compare_attr = TRUE,
  na_is_zero = FALSE
)

compare_sims_cbind(classic_sim, tmb_sim, index)

compare_sims_plot(...)

compare_sims_diffmat(classic_sim, tmb_sim, state, op = `-`)

compare_sims_unlist(classic_sim, tmb_sim, tolerance = testthat_tolerance())
}
\arguments{
\item{classic_sim}{result of `run_sim` without using TMB}

\item{tmb_sim}{result of `run_sim` using TMB}

\item{tolerance}{numerical tolerance}

\item{compare_attr}{compare attributes or just the simulations themselves}

\item{na_is_zero}{should NAs be replaced with zeros?}

\item{index}{length-one vector for extracting the simulated variable to be
compared}

\item{...}{
  Arguments passed on to \code{\link[=compare_sims_cbind]{compare_sims_cbind}}
  \describe{
    \item{\code{}}{}
  }}

\item{state}{vector of state names to compare}

\item{op}{binary comparison operator}
}
\description{
Exceptions to drop-in replacement:
1. don't require that the attributes are in the same order
2. don't require that the r version returns everything that the tmb
   version does (e.g. flexmodel)
3. don't require that the row.names are identical (is this ok?
   the r version counts iterations with skips, but is this informative?)
4. don't require that the call is identical (obvious i guess, but
   being exhaustive)
5. don't look for parameter called S0 in classic simulation results
   (FIXME: this one might not be necessary anymore)
6. don't require the ordering of params_timevar rows to be the same
   (FIXME: this should probably be handled in run_sim itself)
}
\seealso{
Other compare: 
\code{\link{compare_grads}()},
\code{\link{time_wrap}()}
}
\concept{compare}
