% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{complement}
\alias{complement}
\alias{inverse}
\title{Symbolic Complement and Inverse}
\usage{
complement(x)

inverse(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Symbolic Complement and Inverse
}
\examples{
complement('p')
inverse('N')
}
