% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{condense.pansim}
\alias{condense.pansim}
\title{Condense columns (infected, ICU, hospitalized) in a pansim output}
\usage{
\method{condense}{pansim}(
  object,
  add_reports = TRUE,
  diff_deaths = TRUE,
  keep_all = FALSE,
  params = attr(object, "params"),
  ...
)
}
\arguments{
\item{object}{a pansim object}

\item{add_reports}{add incidence and case reports?}

\item{diff_deaths}{compute first differences of death series to get daily deaths?}

\item{keep_all}{keep unaggregated variables in data frame as well?}

\item{params}{parameters (for defining convolution kernel for reports)}

\item{...}{additional args}
}
\description{
Condense columns (infected, ICU, hospitalized) in a pansim output
}
