% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{condense}
\alias{condense}
\title{condense an object}
\usage{
condense(object, ...)
}
\arguments{
\item{object}{an object to condense}

\item{...}{additional arguments}
}
\description{
condense an object
}
