% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{condense_age}
\alias{condense_age}
\title{collapse age/testing/etc. categories (only; don't do other condensation)}
\usage{
condense_age(x, levels = unique(epi_cat))
}
\arguments{
\item{x}{a state vector}

\item{levels}{levels/sort order}
}
\description{
collapse age/testing/etc. categories (only; don't do other condensation)
}
