% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{condense_params_vax}
\alias{condense_params_vax}
\title{condense vaxified parameters to base parameters}
\usage{
condense_params_vax(params)
}
\arguments{
\item{params}{vaxified parameters (e.g. generated with `expand_params_vax()`)}
}
\value{
an object of class `params_pansim`
}
\description{
condense vaxified parameters to base parameters
}
\examples{
condense_params_vax(expand_params_vax(read_params("PHAC.csv")))
}
