% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{condense_vax}
\alias{condense_vax}
\title{collapse vaccination categories (only; don't do other condensation)}
\usage{
condense_vax(x)
}
\arguments{
\item{x}{a state vector or simulation result df (an object of class `state_pansim` or `pansim`)}
}
\description{
collapse vaccination categories (only; don't do other condensation)
}
