% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmodel_kernel.R
\name{configure_epi_summaries}
\alias{configure_epi_summaries}
\alias{epi_kernel}
\alias{epi_pars}
\title{Configure Epidemiological Summaries}
\usage{
configure_epi_summaries(
  model,
  exposed_state_nm,
  foi_nm,
  prop_susceptible_nm,
  trans_rate_nm,
  ...
)

epi_kernel(model)

epi_pars(model)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{exposed_state_nm}{name of the exposed state}

\item{foi_nm}{name of the force of infection associated with
the exposed state}

\item{prop_susceptible_nm}{name of the variable that contains the
user-defined effective proportion of individuals who are susceptible}

\item{trans_rate_nm}{name of the parameter giving the baseline transmission
rate}

\item{...}{named parameter values to update for the cohort
model -- a common example is \code{N = 1}, which allows
certain models to track relative population sizes}
}
\description{
Configure Epidemiological Summaries
}
\section{Functions}{
\itemize{
\item \code{epi_kernel()}: obtain the kernel from \code{model}

\item \code{epi_pars()}: compute epidemic growth parameters for \code{model}

}}
