% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructing_vectors.R
\name{const_named_vector}
\alias{const_named_vector}
\title{Constant Named Vector}
\usage{
const_named_vector(nms, cnst)
}
\arguments{
\item{nms}{names of the output vector}

\item{cnst}{single numeric value to be used in every element of the
output vector}
}
\description{
Constant Named Vector
}
