% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{cross}
\alias{cross}
\title{Cross Two Sets of Names in Matrix Form}
\usage{
cross(from, to, mat)
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}
}
\value{

}
\description{
Cross Two Sets of Names in Matrix Form
}
\seealso{
Other struc_functions: 
\code{\link{struc}}
}
\concept{struc_functions}
