% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{cross_mat}
\alias{cross_mat}
\title{Cross Two Sets of Names in Matrix Form}
\usage{
cross_mat(x, y, sep = "_")
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}
}
\value{

}
\description{
Cross Two Sets of Names in Matrix Form
}
\seealso{
Other struc_functions: 
\code{\link{struc_block}()},
\code{\link{struc_stretch}()},
\code{\link{struc}}
}
\concept{struc_functions}
