% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmodel_simulate.R
\name{deprecated_simulation_functions}
\alias{deprecated_simulation_functions}
\alias{changing_ratemat_elements}
\alias{simulate_changing_ratemat_elements}
\alias{concatenated_state_vector}
\alias{simulate_state_vector}
\alias{initial_state_vector}
\alias{final_state_vector}
\alias{penultimate_state_vector}
\alias{final_state_ratio}
\alias{initial_ratemat}
\alias{simulation_condensed}
\alias{condense_flexmodel}
\alias{simulate.flexmodel}
\alias{simulation_fitted}
\title{Deprecated Simulation Functions}
\usage{
changing_ratemat_elements(model, sim_params = NULL)

simulate_changing_ratemat_elements(model, sim_params = NULL, add_dates = FALSE)

concatenated_state_vector(model, sim_params = NULL)

simulate_state_vector(
  model,
  sim_params = NULL,
  add_dates = FALSE,
  format = c("wide", "long", "ggplot")
)

initial_state_vector(model, sim_params = NULL)

final_state_vector(model, sim_params = NULL)

penultimate_state_vector(model, sim_params = NULL)

final_state_ratio(model, sim_params = NULL)

initial_ratemat(model, sim_params = NULL)

simulation_condensed(model, add_dates = TRUE, sim_params = NULL)

condense_flexmodel(model)

\method{simulate}{flexmodel}(
  object,
  nsim = 1,
  seed = NULL,
  do_condensation = FALSE,
  format = c("long", "wide"),
  add_dates = TRUE,
  sim_params = NULL,
  ...
)

simulation_fitted(model)
}
\description{
Deprecated Simulation Functions
}
\seealso{
Other simulation: 
\code{\link{simulate_ensemble}()},
\code{\link{simulation_history}()}
}
\concept{simulation}
