% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{derive}
\alias{derive}
\alias{derive.EpiMatrixList}
\title{Derive}
\usage{
derive(mat_list, ...)

\method{derive}{EpiMatrixList}(mat_list, ...)
}
\arguments{
\item{mat_list}{object containing a list of matrix-like objects}
}
\description{
Derive symbolic expressions of existing epidemiological variables, and
add them to the list. See \code{\link{epi_mat_list}} for more details
on usage.
}
