% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{describe_params}
\alias{describe_params}
\title{Describe parameters}
\usage{
describe_params(x, stop_missing_names = FALSE)
}
\arguments{
\item{x}{a \code{params_pansim} object}

\item{stop_missing_names}{stop if names are missing descriptions? (warn by default)}
}
\description{
Create a data frame with symbols, values and meanings of parameters
}
