% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dev_is_tikz}
\alias{dev_is_tikz}
\title{Is \code{\link[tikzDevice]{tikzDevice}} currently in use?}
\usage{
dev_is_tikz()
}
\value{
logical
}
\description{
Convenient for dealing with text in graphics,
which can be rendered much more professionally with \code{\link[tikzDevice]{tikz}}.
}
