% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{dimnames_from_template}
\alias{dimnames_from_template}
\title{Dimnames from Template}
\usage{
dimnames_from_template(x, template)
}
\arguments{
\item{x}{object with or without dimnames of the same dimensions as
\code{template}}

\item{template}{object with dimnames to apply to \code{x}}
}
\description{
Make the dimnames of \code{x} be the same as \code{template}
}
