% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{dims}
\alias{dims}
\alias{dim,struc-method}
\alias{dim,EpiMatrix-method}
\alias{dimnames,struc_dimnames-method}
\alias{names,EpiMatrix-method}
\alias{dimnames<-,struc,list-method}
\alias{dimnames,EpiMatrix-method}
\alias{dimnames_from_template,struc,struc_dimnames-method}
\title{Dimensions for Symbolic Objects}
\usage{
\S4method{dim}{struc}(x)

\S4method{dim}{EpiMatrix}(x)

\S4method{dimnames}{struc_dimnames}(x)

\S4method{names}{EpiMatrix}(x)

\S4method{dimnames}{struc,list}(x) <- value

\S4method{dimnames}{EpiMatrix}(x)

\S4method{dimnames_from_template}{struc,struc_dimnames}(x, template)
}
\description{
Dimensions for Symbolic Objects
}
\concept{symbolic}
