% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{distribute_counts}
\alias{distribute_counts}
\title{distribute counts given a desired distribution (with smart rounding)}
\usage{
distribute_counts(total, dist)
}
\arguments{
\item{total}{total count to distribute}

\item{dist}{target distribution (a vector that sums to 1)}
}
\description{
distribute counts given a desired distribution (with smart rounding)
}
