% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{elementwise}
\alias{elementwise}
\alias{*,struc,struc-method}
\alias{*,struc_dimnames,struc_dimnames-method}
\alias{*,struc,struc_dimnames-method}
\alias{*,struc_dimnames,struc-method}
\alias{+,struc,struc-method}
\alias{+,struc_dimnames,struc_dimnames-method}
\alias{+,struc,struc_dimnames-method}
\alias{+,struc_dimnames,struc-method}
\title{Symbolic Elementwise Operators}
\usage{
\S4method{*}{struc,struc}(e1, e2)

\S4method{*}{struc_dimnames,struc_dimnames}(e1, e2)

\S4method{*}{struc,struc_dimnames}(e1, e2)

\S4method{*}{struc_dimnames,struc}(e1, e2)

\S4method{+}{struc,struc}(e1, e2)

\S4method{+}{struc_dimnames,struc_dimnames}(e1, e2)

\S4method{+}{struc,struc_dimnames}(e1, e2)

\S4method{+}{struc_dimnames,struc}(e1, e2)
}
\arguments{
\item{e1}{Operand as a \code{\link{struc}} object}

\item{e2}{Operand as a \code{\link{struc}} object}
}
\description{
Symbolic Elementwise Operators
}
\concept{symbolic}
