% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{epi_mat}
\alias{epi_mat}
\title{Construct an Epidemiological Matrix}
\usage{
epi_mat(nm, x)
}
\arguments{
\item{nm}{Length-one character vector giving the name of the matrix}

\item{x}{numeric or character vector or matrix with either names
(for vectors) or dimnames (for matrices)}
}
\description{
Construct a single object of class \code{\link{EpiMatrix-class}}. Typically,
the \code{\link{epi_mat_list}} function is used to construct several
such epidemiological matrices.
}
