% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{epi_mat_names}
\alias{epi_mat_names}
\alias{fix_epi_names}
\alias{force_numeric_epi_names}
\alias{set_fixed_epi_names}
\alias{set_fixed_epi_dimnames}
\alias{set_numeric_epi_names}
\alias{set_numeric_epi_dimnames}
\alias{sort_epi_dims}
\title{Helpers for Making Consistent Names and Dimension Names}
\usage{
fix_epi_names(x)

force_numeric_epi_names(x)

set_fixed_epi_names(x)

set_fixed_epi_dimnames(x)

set_numeric_epi_names(x)

set_numeric_epi_dimnames(x)

sort_epi_dims(x)
}
\arguments{
\item{x}{a character vector of names or a numeric or character vector
or matrix}
}
\description{
Helpers for Making Consistent Names and Dimension Names
}
\section{Functions}{
\itemize{
\item \code{fix_epi_names()}: create names that include letters,
numbers, and underscores only -- this is the recommended format
for naming the components of an \code{epi_mat_list}

\item \code{force_numeric_epi_names()}: force names to include only digits

\item \code{set_fixed_epi_names()}: fix the names of an object to be passed
to \code{epi_mat_list}

\item \code{set_fixed_epi_dimnames()}: fix the dimension names of an object to be
passed to \code{epi_mat_list}

\item \code{set_numeric_epi_names()}: fix the names to contain only digits of an
object to be passed to \code{epi_mat_list}

\item \code{set_numeric_epi_dimnames()}: fix the dimension names to contain only digits of an
object to be passed to \code{epi_mat_list}

\item \code{sort_epi_dims()}: sort the elements of a matrix or vector
in the order of their dimnames or names

}}
