% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\docType{class}
\name{EpiMatrix-class}
\alias{EpiMatrix-class}
\alias{EpiMatrixInput-class}
\alias{EpiMatrixDerived-class}
\title{Class to Represent Matrices for use in Compartmental Models}
\description{
This is a new class that is intended to eventually replace
\code{\link{struc-class}}. The benefit of this new class is
that it is not neccesary to construct special matrices and
vectors (see \code{\link{struc-class}}) in order to define matrix
algebraic expressions for use in compartmental models. Instead,
the user simply supplies plain numerical R vectors and matrices
with consistent names and dimnames, and if special symbolic
objects are required by the engine they are created automatically
without the user needing to be aware of this. As the engine
evolves to require less symbolic manipulation of matrices, this
new class will work just as well and so it serves to soften the
transition to the new engine (\url{https://canmod.net/misc/cpp_side}).
}
