% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{expand_names}
\alias{expand_names}
\title{Expand Names}
\usage{
expand_names(x, y, sep = "_")
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}

\item{sep}{character separating the \code{x} and \code{y} components
of the output names}
}
\description{
To expand states with subcategories (age, vaccine status)
x_y, with x varying faster
}
