% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{expand_params_age}
\alias{expand_params_age}
\title{Expand parameter list to include age structure}
\usage{
expand_params_age(
  params,
  age_cat = mk_agecats(),
  beta0 = NULL,
  pmat = mk_pmat(),
  Nvec = mk_Nvec()
)
}
\arguments{
\item{params}{parameter list (e.g. read in with `read_params()`)}

\item{age_cat}{vector of age categories}

\item{beta0}{vector of age-specific beta0 values (if NULL, assume same beta0 for all age groups as already provided in params)}

\item{pmat}{contact matrix; default is uniform}

\item{Nvec}{population distribution (as counts); default is uniform}
}
\description{
Expand parameter list to include age structure
}
\examples{
params <- read_params("PHAC_testify.csv")
params_age <- expand_params_age(params)
}
