% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{expand_params_desc_age}
\alias{expand_params_desc_age}
\title{Edit parameter description attribute to include age-specific definitions
(helper function for `expand_params_age()`)}
\usage{
expand_params_desc_age(params_desc)
}
\arguments{
\item{params_desc}{parameter descriptions, as initialized in `read_params()`}
}
\description{
Edit parameter description attribute to include age-specific definitions
(helper function for `expand_params_age()`)
}
\examples{
params <- read_params("PHAC_testify.csv")
params_desc <- attr(params, "description")
expand_params_desc_age(params_desc)
}
