% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{expand_params_desc_variant}
\alias{expand_params_desc_variant}
\title{Edit parameter description attribute to include variant-specific definitions
(helper function for `expand_params_variant()`)}
\usage{
expand_params_desc_variant(params_desc)
}
\arguments{
\item{params_desc}{parameter descriptions, as initialized in `read_params()`}
}
\value{

}
\description{
Edit parameter description attribute to include variant-specific definitions
(helper function for `expand_params_variant()`)
}
