% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{expand_params_desc_vax}
\alias{expand_params_desc_vax}
\title{Edit parameter description attribute to include vax-specific definitions
(helper function for `expand_params_vax()`)}
\usage{
expand_params_desc_vax(params_desc)
}
\arguments{
\item{params_desc}{parameter descriptions, as initialized in `read_params()`}
}
\value{

}
\description{
Edit parameter description attribute to include vax-specific definitions
(helper function for `expand_params_vax()`)
}
