% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{expand_params_S0}
\alias{expand_params_S0}
\title{Expand Parameter List to Include Initial Proportion of Susceptibles}
\usage{
expand_params_S0(params, S0)
}
\arguments{
\item{params}{parameter list (e.g. read in with \code{read_params()})}

\item{S0}{initial proportion of individuals in susceptible compartments}
}
\description{
TODO: deprecate this
}
