% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{expand_params_variant}
\alias{expand_params_variant}
\title{Expand parameter list to include variant strain}
\usage{
expand_params_variant(
  params,
  variant_prop = 0.01,
  variant_advantage = 1.5,
  variant_vax_efficacy_dose1 = 0.3,
  variant_vax_efficacy_dose2 = 0.8
)
}
\arguments{
\item{params}{parameter list (e.g. read in with \code{read_params()})}

\item{variant_prop}{current proportion of infections caused by the variant}

\item{variant_advantage}{transmissibility advantage of variant compared to current dominant strain (as a multiplicative factor), e.g., a 50\% more transmissible variant would have \code{variant_advantage = 1.5}.}

\item{variant_vax_efficacy_dose1}{one-dose vaccine efficacy against variant (only used in vaxified model)}

\item{variant_vax_efficacy_dose2}{two-dose vaccine efficacy against variant (only used in vaxified model)}
}
\description{
Expand parameter list to include variant strain
}
\examples{
params <- read_params("PHAC.csv")
params_variant <- expand_params_variant(params)
}
