% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{expand_params_vax}
\alias{expand_params_vax}
\title{Expand parameter list to include vaccination}
\usage{
expand_params_vax(params, vax_cat = mk_vaxcats(), doses_per_day = NULL)
}
\arguments{
\item{params}{parameter list (e.g. read in with `read_params()`)}

\item{vax_cat}{vector of vaccination categories}
}
\description{
Expand parameter list to include vaccination
}
\examples{
params <- read_params("PHAC_testify.csv")
params_age <- expand_params_age(params)
}
