% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{expand_state_vax}
\alias{expand_state_vax}
\title{expand state vector by vaccination status}
\usage{
expand_state_vax(x, vax_cat = mk_vaxcats())
}
\arguments{
\item{x}{state vector}

\item{vax_cat}{vaccine status categories}
}
\description{
by default, everyone starts with an unvaccinated status
}
\examples{
params <- read_params("PHAC_testify.csv")
ss <- make_state(params=params)
ss2 <- expand_state_vax(ss)
}
