% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testify.R
\name{expand_stateval}
\alias{expand_stateval}
\title{expand states and values to include}
\usage{
expand_stateval(
  x,
  method = c("untested", "spread"),
  add_accum = TRUE,
  non_expanded = c("R", "D", "X")
)
}
\arguments{
\item{x}{state vector}

\item{method}{method for distributing values across new (expanded) states}

\item{add_accum}{add N and P (neg/pos test) accumulator categories?}

\item{non_expanded}{states that should \emph{not} be expanded into testing categorieso}
}
\description{
expand states and values to include
}
