% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{expand_stateval_age}
\alias{expand_stateval_age}
\alias{ageify}
\title{expand state vector and rate matrix by age classes}
\usage{
expand_stateval_age(x, age_cat = mk_agecats())

ageify(ratemat, params, age_cat = mk_agecats())
}
\arguments{
\item{x}{state vector}

\item{age_cat}{vector of age categories}

\item{ratemat}{rate matrix}

\item{params}{parameter vector}
}
\description{
expand state vector and rate matrix by age classes
}
\examples{
pp <- read_params("PHAC_testify.csv")
ss <- make_state(params=pp)
ss2 <- expand_stateval_age(ss)
pp <- read_params("PHAC_testify.csv")
state <- make_state(params=pp)
M <- make_ratemat(state,pp, sparse=TRUE)
Ma <- ageify(M, pp)
library(Matrix)
Matrix::image(Ma)
Mta <- ageify(testify(M,pp),pp)
Matrix::image(Mta)
}
