% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testify.R
\name{expand_stateval_testing}
\alias{expand_stateval_testing}
\title{expand states and values to include}
\usage{
expand_stateval_testing(x, method = c("untested", "spread"), add_accum = TRUE)
}
\arguments{
\item{x}{state vector}

\item{method}{method for distributing values across new (expanded) states}

\item{add_accum}{add N and P (neg/pos test) accumulator categories?}
}
\description{
expand states and values to include
}
