% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructing_vectors.R
\name{expand_strain_frame}
\alias{expand_strain_frame}
\alias{expand_strain_names}
\title{Expand Strain Names}
\usage{
expand_strain_frame(
  base_states = c("S", "I", "R"),
  constrained_states = c(),
  constraint_counts = 0:1,
  n_strains = 2,
  strain_name_prefix = "strain"
)

expand_strain_names(..., sep = "")
}
\arguments{
\item{base_states}{character vector giving the states of the base model}

\item{n_strains}{number of strains}

\item{strain_name_prefix}{prefix for names of strains (they are numbered)}

\item{...}{
  Arguments passed on to \code{\link[=expand_strain_frame]{expand_strain_frame}}
  \describe{
    \item{\code{}}{}
  }}

\item{sep}{separator between strains}

\item{infected_states}{character vector giving the infected states}
}
\value{
\code{expand_strain_frame} returns a data frame with one column
for each strain and one row for each expanded state.
\code{expand_strain_names} returns a vector with the full state names for
each state in the expanded model.
}
\description{
This is similar to \code{\link{expand_names}} but for multi-strain models.
The expansion also allows setting constraints on the numbers of particular
\code{base_states} in the final constrained set of states.
}
