% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{extend_end_date}
\alias{extend_end_date}
\title{Extend End Date}
\usage{
extend_end_date(model, days_to_extend)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{days_to_extend}{number of days to extend the end date}
}
\description{
Extend the final simulation date a number of days
in the future.
}
