% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{feature_check}
\alias{feature_check}
\title{Feature Check}
\usage{
feature_check(introduced_version, feature)
}
\description{
Throw an error if a feature is not being used, even though it is required
by the spec being used.
}
