% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_stored}
\alias{fix_stored}
\title{convert 'old' forecast args to 'new' format}
\usage{
fix_stored(x, sim_fun = run_sim_break)
}
\arguments{
\item{x}{fitted model (\code{fit_pansim})}

\item{sim_fun}{simulation function}
}
\description{
This takes care of a few backward incompatibilities
\itemize{
\item converts dates (could previously be in a wide range
of formats, are now required to be in as.Date()-transformable
format so that we don't depend on anytime)
\item inserts the \emph{current} version of \code{run_sim_break}
(or whatever \code{sim_fun} is specified)
\item converts params_timevar to new format
}
}
