% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{flexmodel}
\alias{flexmodel}
\alias{init_model}
\title{Initialize Flexible Compartmental Model}
\usage{
flexmodel(
  params,
  state = NULL,
  start_date = NULL,
  end_date = NULL,
  params_timevar = NULL,
  do_hazard = getOption("MP_default_do_hazard"),
  do_make_state = getOption("MP_default_do_make_state"),
  do_hazard_lin = getOption("MP_default_do_hazard_lin"),
  do_approx_hazard = getOption("MP_default_do_approx_hazard"),
  do_approx_hazard_lin = getOption("MP_default_do_approx_hazard_lin"),
  do_sim_constraint = getOption("MP_default_do_sim_constraint"),
  sim_lower_bound = getOption("MP_default_sim_lower_bound"),
  max_iters_eig_pow_meth = 8000,
  tol_eig_pow_meth = 1e-06,
  haz_eps = 1e-06,
  data = NULL,
  ...
)

init_model(...)
}
\arguments{
\item{params}{a \code{param_pansim} object}

\item{state}{a \code{state_pansim} object}

\item{start_date}{simulation start date}

\item{end_date}{simulation end date}

\item{params_timevar}{data frame with scheduling for piece-wise
constant parameter variation (TODO: direct to other help pages)}

\item{do_hazard}{should hazard simulation steps be used?
(https://canmod.net/misc/flex_specs#v0.0.5) -- only used
if \code{spec_ver_gt('0.0.4')}}

\item{do_make_state}{should state be remade on the c++ size?
(https://canmod.net/misc/flex_specs#v0.1.1) -- only used
if \code{spec_ver_gt('0.1.0')}}

\item{do_hazard_lin}{like \code{do_hazard} but for the
linearized model that is used to construct the initial state
variable -- only used when \code{do_make_state == TRUE}}

\item{do_approx_hazard}{like \code{do_approx_hazard} but for
the linearized model that is used to construct the initial
state (experimental)}

\item{do_sim_constraint}{should simulated values be smoothly
constrained to be above \code{sim_lower_bound} when computing
negative binomial (maybe others in the future?) loss functions?
the smooth constraint function is
\eqn{y = x + \epsilon * \exp(-x / \epsilon)}, where \eqn{\epsilon}
is \code{sim_lower_bound} and \eqn{x} is the simulated value.}

\item{sim_lower_bound}{optional lower bound on the simulated values
when computing negative binomial loss functions (only applicable
when \code{do_sim_constraint} is \code{TRUE}.}

\item{max_iters_eig_pow_meth}{maximum number of iterations
to use in computing the eigenvector for initial state
construction}

\item{tol_eig_pow_meth}{tolerance for determining convergence
of the power method used in initial state construction}

\item{data}{optional observed data frame in long format to
compare with simulated trajectories. must have the following
columns: \code{date}, \code{var}, \code{value}.}
}
\value{
flexmodel object representing a compartmental model
}
\description{
https://canmod.net/misc/flex_specs
}
\seealso{
Other flexmodels: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{initial_population}()},
\code{\link{make_base_model}()},
\code{\link{make_omicron_model}()},
\code{\link{make_sir_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{tmb_fun}()},
\code{\link{update_tmb_indices}()},
\code{\link{vec_rate}()}
}
\concept{flexmodels}
