% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{*,struc_expanded,struc-method}
\alias{*,struc_expanded,struc-method}
\alias{*,struc_expanded,struc_expanded-method}
\alias{dim,struc_expanded-method}
\title{Functions for Developers}
\usage{
\S4method{*}{struc_expanded,struc}(e1, e2)

\S4method{*}{struc_expanded,struc_expanded}(e1, e2)

\S4method{dim}{struc_expanded}(x)
}
\description{
Functions for Developers

Functions for Developers

Functions for Developers
}
\keyword{internal}
