% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{forecast_ensemble}
\alias{forecast_ensemble}
\title{find confidence envelopes by simulation}
\usage{
forecast_ensemble(
  fit,
  nsim = 200,
  forecast_args = fit$forecast_args,
  qvec = c(0.05, 0.5, 0.95),
  qnames = c("lwr", "value", "upr"),
  seed = NULL,
  imp_wts = FALSE,
  Sigma = bbmle::vcov(fit$mle2),
  scale_Sigma = 1,
  calc_Rt = FALSE,
  fix_pars_re = "nb_disp",
  raw_ensembles = FALSE,
  .progress = if (interactive()) "text" else "none"
)
}
\arguments{
\item{fit}{output from \code{calibrate}}

\item{nsim}{number of simulations}

\item{forecast_args}{arguments to pass to \code{forecast_sim}}

\item{qvec}{vector of quantiles: NULL to return an array (nt x nvars x nsims) instead of a tibble with date/var+ quantiles}

\item{qnames}{quantile names}

\item{seed}{random-number seed}

\item{imp_wts}{use importance weighting, i.e. weight ensemble based on log-likelihood?}

\item{Sigma}{covariance matrix to pass to \code{pop_pred_samp}}

\item{scale_Sigma}{multiplier for covariance matrix}

\item{calc_Rt}{calculate and include R(t) in prediction/forecast?}

\item{fix_pars_re}{a regular expression specifying the names of parameters that should be treated as fixed when constructing the parameter ensemble}

\item{raw_ensembles}{(logical) return ensembles (FIXME: should implement return_type=c("array","quantiles","raw") (but not really raw, "dataframe"??))}

\item{.progress}{progress bar?}
}
\description{
find confidence envelopes by simulation
}
