% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{forecast_sim}
\alias{forecast_sim}
\title{simulate/forecast a single trajectory}
\usage{
forecast_sim(
  p,
  opt_pars,
  base_params,
  start_date,
  end_date,
  time_args = NULL,
  fixed_pars = NULL,
  stoch = NULL,
  stoch_start = NULL,
  sim_args = list(),
  aggregate_args = NULL,
  condense_args = NULL,
  return_val = c("aggsim", "vals_only"),
  sim_fun = run_sim_break,
  calc_Rt = FALSE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{p}{vector of parameters - on the link (log/logit) scale as appropriate; these are the parameters that will be adjusted during calibration}

\item{opt_pars}{starting parameters (and structure).  Parameters
that are part of the \code{params_pansim} parameter vector can
be specified within the \code{params} element (with prefixes
if they are transformed); other parameters can include
distributional parameters or time-varying parameters}

\item{base_params}{baseline parameters (an object (vector?) of
type \code{params_pansim} containing all of the parameters
needed for a simulation; some may be overwritten during the
calibration process)}

\item{start_date}{starting date for sims (far enough back to allow
states to sort themselves out)}

\item{end_date}{ending date}

\item{time_args}{arguments passed to \code{sim_fun}}

\item{fixed_pars}{parameters to fix}

\item{stoch}{stochastic settings (see \code{\link{run_sim}})}

\item{stoch_start}{dates on which to enable stochasticity (vector of dates with names 'proc' and 'obs')}

\item{sim_args}{additional arguments to pass to
\code{\link{run_sim}}}

\item{aggregate_args}{arguments passed to
\code{\link{aggregate.pansim}}}

\item{condense_args}{arguments to pass to \code{\link{condense}} (via \code{\link{run_sim}})}

\item{return_val}{specify values to return (aggregated simulation, or just the values?)}

\item{sim_fun}{function for simulating a single run
(e.g. \code{\link{run_sim_break}},
\code{\link{run_sim_mobility}})}

\item{calc_Rt}{calculate and include R(t) in prediction/forecast?}

\item{debug}{print debugging messages?}

\item{...}{extra args (ignored)}
}
\description{
simulate based on a vector of parameters (including both time-varying change parameters, initial conditions, and other dynamical parameters), for fitting or forecasting
}
\examples{
ff <- ont_cal1$forecast_args
op <- ff$opt_pars
p <- unlist(op)
params <- fix_pars(read_params("ICU1.csv"))
forecast_sim(p, op, base_params=params,ff$start_date, ff$end_date,
    time_args=ff$time_args)
}
\seealso{
Other classic_macpan: 
\code{\link{add_d_log}()},
\code{\link{add_updated_vaxrate}()},
\code{\link{aggregate_agecats}()},
\code{\link{calibrate_comb}()},
\code{\link{calibrate}()},
\code{\link{check_age_cat_compatibility}()},
\code{\link{check_contact_rate_setting}()},
\code{\link{col_multiply}()},
\code{\link{condense_age}()},
\code{\link{condense_params_vax}()},
\code{\link{condense_state}()},
\code{\link{condense_vax}()},
\code{\link{dev_is_tikz}()},
\code{\link{do_step}()},
\code{\link{expand_params_age}()},
\code{\link{expand_params_desc_age}()},
\code{\link{expand_params_desc_variant}()},
\code{\link{expand_params_desc_vax}()},
\code{\link{expand_params_mistry}()},
\code{\link{expand_params_variant}()},
\code{\link{expand_params_vax}()},
\code{\link{expand_state_age}()},
\code{\link{expand_state_vax}()},
\code{\link{expand_stateval_testing}()},
\code{\link{fix_pars}()},
\code{\link{fix_stored}()},
\code{\link{forecast_ensemble}()},
\code{\link{getData}()},
\code{\link{get_GI_moments}()},
\code{\link{get_Gbar}()},
\code{\link{get_R0}()},
\code{\link{get_doses_per_day}()},
\code{\link{get_evec}()},
\code{\link{get_kernel_moments}()},
\code{\link{get_opt_pars}()},
\code{\link{get_r}()},
\code{\link{invlink_trans}()},
\code{\link{make_betavec}()},
\code{\link{make_beta}()},
\code{\link{make_jac}()},
\code{\link{make_ratemat}()},
\code{\link{make_state}()},
\code{\link{make_test_wtsvec}()},
\code{\link{make_vaxrate}()},
\code{\link{mk_Nvec}()},
\code{\link{mk_agecats}()},
\code{\link{mk_contact_rate_setting}()},
\code{\link{mk_mistry_Nvec}()},
\code{\link{mk_pmat}()},
\code{\link{mk_vaxcats}()},
\code{\link{mle_fun}()},
\code{\link{non_expanded_states}},
\code{\link{rExp}()},
\code{\link{read_params}()},
\code{\link{repair_names_age}()},
\code{\link{restore}()},
\code{\link{run_sim_ageify}()},
\code{\link{run_sim_break}()},
\code{\link{run_sim_loglin}()},
\code{\link{run_sim_mobility}()},
\code{\link{run_sim_range}()},
\code{\link{run_sim}()},
\code{\link{show_ratemat}()},
\code{\link{testify}()},
\code{\link{texify}()},
\code{\link{trans_state_vars}()},
\code{\link{update_contact_rate_setting}()},
\code{\link{update_foi}()},
\code{\link{update_params_mistry}()},
\code{\link{vis_model}()},
\code{\link{write_params}()}
}
\concept{classic_macpan}
