% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmodel_parameters.R
\name{get_and_set_model_parameters}
\alias{get_and_set_model_parameters}
\alias{pars_base_sim}
\alias{pars_base_init}
\alias{pars_base_opt}
\alias{pars_time_sim}
\alias{pars_time_spec}
\alias{pars_time_opt}
\alias{pars_time_series}
\alias{pars_infer_init}
\alias{pars_infer_opt}
\alias{pars_infer_conf}
\alias{pars_base_sim<-}
\title{Get and Set Model Parameters}
\usage{
pars_base_sim(model)

pars_base_init(model)

pars_base_opt(model)

pars_time_sim(model)

pars_time_spec(model)

pars_time_opt(model)

pars_time_series(model)

pars_infer_init(model)

pars_infer_opt(model)

pars_infer_conf(model)

pars_base_sim(model) <- value
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{value}{named vector of parameters}
}
\value{
parameter vector or data frame of time-variation of parameters
}
\description{
Get and Set Model Parameters
}
