% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmodel_state.R
\name{get_and_set_model_state}
\alias{get_and_set_model_state}
\alias{state_default}
\alias{state_init}
\alias{state_final}
\alias{state_final_rand}
\alias{state_final_cond}
\alias{state_default<-}
\alias{state_init<-}
\alias{state_default.flexmodel}
\title{Get and Set Model State}
\usage{
state_default(model)

state_init(model)

state_final(model)

state_final_rand(model)

state_final_cond(model)

state_default(model) <- value

state_init(model) <- value

\method{state_default}{flexmodel}(model)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{value}{named vector of states}
}
\value{
state vector or data frame of time-variation of parameters
}
\description{
Get and Set Model State
}
\section{Functions}{
\itemize{
\item \code{state_default}: State vector used to define the model and used as the initial state if \code{do_make_state == FALSE}.

\item \code{state_init}: Initial state vector at \code{model$start_date}

\item \code{state_final}: Deterministic value of the final state vector at \code{model$end_date}

\item \code{state_final_rand}: Random value of the final state vector at \code{model$end_date}

\item \code{state_final_cond}: Condensed state vector at \code{model$end_date}.
}}

