% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{get_break}
\alias{get_break}
\title{estimate magnitude of drop in beta0 at known dates}
\usage{
get_break(
  date0 = ldmy("1-Mar-2020"),
  end_date = ldmy("8-Apr-2020"),
  break_dates = c("20-Mar-2020"),
  params,
  data,
  var = "H",
  sim_args = NULL,
  optim_args = list()
)
}
\arguments{
\item{date0}{date for initial conditions}

\item{end_date}{ending date of sim}

\item{break_dates}{dates of breakpoints in transmission}

\item{params}{a parameter set}

\item{data}{data to match}

\item{var}{variable for regression (H, D, ICU)}

\item{sim_args}{additional parameters to pass to \code{run_sim} if shooting}

\item{optim_args}{additional args to \code{optim}}
}
\description{
estimate magnitude of drop in beta0 at known dates
}
