% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{get_doses_per_day}
\alias{get_doses_per_day}
\title{Compute total number of vaccine doses actually administered per day in a simulation}
\usage{
get_doses_per_day(res, dose = c(1, 2))
}
\arguments{
\item{res}{simulation result from model with vaccination (generated using `run_sim`)}

\item{dose}{the number of doses to consider, vector. (note that Moderna/Pfizer both recommend 2 doses) (??)}
}
\value{
`tibble` with columns for date and total vaccine doses actually administered each day (as captured in the simulation)
}
\description{
Compute total number of vaccine doses actually administered per day in a simulation
}
