% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afuns.R
\name{get_evec}
\alias{get_evec}
\title{get dominant eigenvector}
\usage{
get_evec(p, method = c("expsim", "analytical"), ...)
}
\arguments{
\item{p}{parameters}

\item{method}{computational method}

\item{...}{args passed through to rExp (esp testify flag!)}
}
\description{
get dominant eigenvector
}
