% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afuns.R
\name{get_Gbar}
\alias{get_Gbar}
\title{compute mean generation interval from parameters}
\usage{
get_Gbar(p, method = c("analytical", "kernel"))
}
\arguments{
\item{p}{a set of parameters}

\item{method}{computational method}
}
\description{
compute mean generation interval from parameters
}
