% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afuns.R
\name{get_GI_moments}
\alias{get_GI_moments}
\title{compute moments of generation interval (mean and CV^2)}
\usage{
get_GI_moments(params)
}
\arguments{
\item{params}{parameters}
}
\description{
compute moments of generation interval (mean and CV^2)
}
