% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{get_init}
\alias{get_init}
\title{Back-calculate initial conditions
Suppose we have a slope and an intercept for some variable
(e.g. hospitalizations) starting
from day t1.  We want an initial condition for day t0.
(Don't go too far back or rounding will kill us)#}
\usage{
get_init(
  date0 = anydate("1-Mar-2020"),
  date1 = anydate("25-Mar-2020"),
  params,
  int = NULL,
  slope = NULL,
  var = "H",
  init_target = NULL,
  sim_args = NULL
)
}
\arguments{
\item{date0}{date to get initial conditions for}

\item{date1}{initial date for regressions (t=0)}

\item{params}{a parameter set}

\item{int}{regression intercept}

\item{slope}{regression slope}

\item{var}{regression variable}

\item{init_target}{value of focal variable to hit on date1}

\item{sim_args}{additional parameters to pass to \code{run_sim} if shooting}
}
\description{
Back-calculate initial conditions
Suppose we have a slope and an intercept for some variable
(e.g. hospitalizations) starting
from day t1.  We want an initial condition for day t0.
(Don't go too far back or rounding will kill us)#
}
\examples{
params <- read_params(system.file("params","ICU1.csv",package="McMasterPandemic"))
## eigenvector projection
get_init(var="H",int=2.85,slope=0.176, param=params)
## shooting
get_init(var="H",init_target=50, param=params)
}
