% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afuns.R
\name{get_kernel_moments}
\alias{get_kernel_moments}
\title{compute R0, r, etc. based on kernel computation}
\usage{
get_kernel_moments(params)
}
\arguments{
\item{params}{parameter vector}
}
\description{
compute R0, r, etc. based on kernel computation
}
