% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_opt_pars}
\alias{get_opt_pars}
\title{generate reasonable default opt_pars}
\usage{
get_opt_pars(params, vars = c("hosp", "death", "report"))
}
\arguments{
\item{params}{baseline parameters}

\item{vars}{which variables are being fitted to?}
}
\description{
generate reasonable default opt_pars
}
\examples{
params <- read_params("ICU1.csv")
get_opt_pars(params)
get_opt_pars(params,vars="hosp")
}
