% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afuns.R
\name{get_r}
\alias{get_r}
\title{return growth rate (from Jacobian)}
\usage{
get_r(p, state = NULL, method = c("expsim", "kernel", "analytical"))
}
\arguments{
\item{p}{parameters}

\item{state}{state (needed for vaxified model)}

\item{method}{computation method}
}
\description{
return growth rate (from Jacobian)
}
