% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afuns.R
\name{get_R0}
\alias{get_R0}
\title{calculate R0 for a given set of parameters}
\usage{
get_R0(params, components = FALSE, method = c("analytical", "kernel"))
}
\arguments{
\item{params}{parameters}

\item{components}{report R0 component-by-component?}
}
\description{
calculate R0 for a given set of parameters
}
