% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{gradfun}
\alias{gradfun}
\title{gradient function for ODE runs}
\usage{
gradfun(t, y, parms, M)
}
\arguments{
\item{t}{time vector}

\item{y}{state vector}

\item{parms}{parameter vector}

\item{M}{rate matrix}
}
\description{
gradient function for ODE runs
}
