% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{\%_\%}
\alias{\%_\%}
\title{Paste with Underscore Separator}
\usage{
x \%_\% y
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}
}
\description{
Paste with an underscore separator, except for length-zero character
strings in the first vector
}
