% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_and_strings.R
\name{index_sep}
\alias{index_sep}
\title{Get Substrings by Indices and Separators}
\usage{
index_sep(x, i, sep = "_", complement = FALSE)
}
\arguments{
\item{x}{character vector}

\item{i}{integer vector without sign mixing}

\item{sep}{length-one character vector}

\item{complement}{if \code{TRUE} the indices in \code{i} that are not
matched are returned}
}
\description{
For example \code{index_sep('a_b_c', 2, '_')} equals \code{'b'}.
For example \code{index_sep('a_b_c', c(1, 3), '_')} equals \code{'a_c'}.
For example \code{index_sep('a_b_c', -2, '_')} equals \code{'a_c'}.
For example \code{index_sep('a_b_c', 4, '_')} equals \code{''}.
For example \code{index_sep('a', 1, '_')} equals \code{'a'}.
For example \code{index_sep('a', 2, '_')} equals \code{''}.
For example \code{index_sep(c('a_b', 'c'), 2, '_')} equals \code{c('b', '')}.
For example \code{index_sep('a_b_c', c(3, 1), '_')} equals \code{'c_a'}.
}
