% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{init_model}
\alias{init_model}
\title{Initialize Flexible Compartmental Model}
\usage{
init_model(
  params,
  state = NULL,
  start_date = NULL,
  end_date = NULL,
  params_timevar = NULL,
  do_hazard = TRUE,
  ...
)
}
\arguments{
\item{params}{a \code{param_pansim} object}

\item{state}{a \code{state_pansim} object}

\item{start_date}{simulation start date}

\item{end_date}{simulation end date}

\item{params_timevar}{data frame with scheduling for piece-wise
constant parameter variation (TODO: direct to other help pages)}

\item{do_hazard}{should hazard simulation steps be used?
(https://canmod.net/misc/flex_specs#v0.0.5) -- only used
if \code{spec_ver_gt('0.0.4')}}
}
\value{
flexmodel object representing a compartmental model
}
\description{
https://canmod.net/misc/flex_specs
}
\seealso{
Other flexmodels: 
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_tmb_indices}()},
\code{\link{initial_population}()},
\code{\link{make_base_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{state_param_sum}()},
\code{\link{tmb_fun}()},
\code{\link{vec_rate}()}
}
\concept{flexmodels}
