% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{initial_population}
\alias{initial_population}
\title{Initial Population}
\usage{
initial_population(model, total, infected)
}
\arguments{
\item{total}{name of a single parameter to represent the total size of the
population -- over all compartments}

\item{infected}{name of a single parameter to represent the initial total size of
the infected population -- over all infected compartments}
}
\description{
\code{infected} is the multiplier of the normalized infected components of the
eigenvector. \code{total - infected} is multiplied by
\code{1/length(initial_susceptible)}
}
\details{
In the future we might want the ability to specify a distribution
across susceptible compartments -- currently a uniform distribution
is assumed..
}
\seealso{
Other flexmodel_definition_functions: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{flexmodel}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{update_disease_free_state}()},
\code{\link{update_linearized_params}()},
\code{\link{vec_rate}()}
}
\concept{flexmodel_definition_functions}
