% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{inner}
\alias{inner}
\title{Inner Product}
\usage{
inner(x, y, ...)
}
\arguments{
\item{x}{first operand to an inner product}

\item{y}{second operand to an inner product}
}
\description{
Compute the inner product between two objects.
}
