% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invlink_trans}
\alias{invlink_trans}
\title{apply inverse-link functions to parameter vector or list based on names}
\usage{
invlink_trans(p)
}
\arguments{
\item{p}{a named vector or list of parameter values with names of the form <linkfun>_name}
}
\description{
apply inverse-link functions to parameter vector or list based on names
}
\examples{
invlink_trans(c(log_p1=0,logit_p2=0))
invlink_trans(list(log_p1=c(0,0),logit_p2=c(0,0,0)))
invlink_trans(list(p1=c(log_a=0,log_b=0),p2=4))
}
