% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{is_1by1}
\alias{is_1by1}
\alias{is_1byn}
\alias{is_nby1}
\title{Test if 1-by-1}
\usage{
is_1by1(x)

is_1byn(x)

is_nby1(x)
}
\arguments{
\item{x}{struc object}
}
\value{
TRUE or FALSE
}
\description{
Test if a \code{\link{struc-class}} object is a scalar
}
\seealso{
Other struc_info_functions: 
\code{\link{same_dims}()}
}
\concept{struc_info_functions}
