% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{justValues_f}
\alias{justValues_f}
\title{Parse inputs from strings.}
\usage{
justValues_f(valuesString, mode)
}
\arguments{
\item{mode}{the type of data encoded in the string: either dates, symbols, or values}

\item{valueString}{the string we'd like to extract values from.}
}
\description{
Handle inputs from the time-varying transmission option, which are recieved as strings.
}
