% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{kronecker,struc,struc-method}
\alias{kronecker,struc,struc-method}
\title{describeIn `struc-class` Kronecker product}
\usage{
\S4method{kronecker}{struc,struc}(X, Y)
}
\description{
describeIn `struc-class` Kronecker product
}
