% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructing_vectors.R
\name{layered_zero_state}
\alias{layered_zero_state}
\title{Layered Zero State Vector}
\usage{
layered_zero_state(...)
}
\arguments{
\item{...}{character vectors with sub-model state names}
}
\description{
Initialize a state vector with all zeros that is based on
a fully-crossed set of states in several sub-models.
}
