% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ldmy}
\alias{ldmy}
\title{attempt to convert x to a date unless it already is one}
\usage{
ldmy(x)
}
\arguments{
\item{x}{a date, or a character string in some day-month-year format that \code{lubridate::dmy} can handle}
}
\description{
attempt to convert x to a date unless it already is one
}
\keyword{internal}
