% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{lin_combin,struc,list-method}
\alias{lin_combin,struc,list-method}
\title{Symbolic Linear Combinations}
\usage{
\S4method{lin_combin}{struc,list}(w, l)
}
\arguments{
\item{w}{Vector of weights as a \code{\link{struc}} object}

\item{l}{List of \code{\link{struc}} objects of compatible dimensions
to be combined linearly}
}
\value{
\code{\link{struc}} object of the same dimensions as those in
the list given by the linear combination
\code{w[1] * l[[1]] + w[2] * l[[2]] ...}
}
\description{
Symbolic Linear Combinations
}
