% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{lin_combin}
\alias{lin_combin}
\title{Linear Combination}
\usage{
lin_combin(w, l)
}
\arguments{
\item{w}{object containing weights for the linear combination}

\item{l}{list of objects to linearly combine, with one object per
element in \code{w}}
}
\description{
Compute the linear combination of a list of objects
}
