% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_and_strings.R
\name{alt_group}
\alias{alt_group}
\alias{wrap_exact}
\alias{all_in}
\alias{any_var}
\alias{all_except}
\title{Regular Expression Conveniences}
\usage{
alt_group(x, exact = FALSE, negate = FALSE)

wrap_exact(x)

all_in(x)

any_var(x)

all_except(x)
}
\arguments{
\item{x}{character vector to convert into a length-1
character vector representing a regular expression
pattern}

\item{exact}{should the regular expression look for
exact matches?}

\item{negate}{should the regular expression look for
stings that do not match?}
}
\value{
a regular expression pattern to be used in functions
like \code{\link{grep}} and \code{\link{grepl}}, but also in
\code{\link{add_state_param_sum}}
}
\description{
Specifications of \code{\link{flexmodel}} objects often
involves string manipulations with regular expressions,
and these functions provide some convenience in this area.
}
