% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{make_age_groups}
\alias{make_age_groups}
\title{Assign age groups}
\usage{
make_age_groups(age, bin_edges_lower)
}
\arguments{
\item{age}{Numeric or factor vector of ages}

\item{bin_edges_lower}{list of numeric values for the lower edges
of desired age group bins in panel 2. If ages are of type factor,
all user-specified lower bin edges must match lower bin edges in
the age data.}
}
\value{

}
\description{
A function that takes a vector of ages (either numeric or factor)
and returns assigned age groups based on desired lower bin edges
provided by the user.
}
\details{
This is a helper function for \code{\link{plot.agetime}}.
}
\examples{
make_age_groups(seq(0, 90, by = 1), c(25, 45))
}
