% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{make_betavec}
\alias{make_betavec}
\title{construct vector of transmission multipliers}
\usage{
make_betavec(state, params, full = TRUE, testify = FALSE)
}
\arguments{
\item{state}{state vector}

\item{params}{parameter vector}

\item{full}{include non-infectious compartments (with transmission of 0) as well as infectious compartments?}

\item{testify}{expand to allow for testing status?}
}
\description{
construct vector of transmission multipliers
}
