% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{make_delay_kernel}
\alias{make_delay_kernel}
\title{construct a Gamma-distributed delay kernel}
\usage{
make_delay_kernel(prop, delay_mean, delay_cv, max_len = 10)
}
\arguments{
\item{prop}{area under the curve (proportion reported)}

\item{delay_mean}{mean value}

\item{delay_cv}{coeff of var}

\item{max_len}{maximum delay}
}
\description{
construct a Gamma-distributed delay kernel
}
