% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flowchart.R
\name{make_flowchart}
\alias{make_flowchart}
\title{display flow chart on current graphics device}
\usage{
make_flowchart(params = read_params("ICU1.csv"))
}
\arguments{
\item{params}{parameters (flowchart is determined by non-zero flows)}
}
\description{
display flow chart on current graphics device
}
