% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{make_hello_world_model}
\alias{make_hello_world_model}
\title{Hello World}
\usage{
make_hello_world_model()
}
\description{
Return a hello-world SIR model
}
\seealso{
\code{\link{make_sir_model}}

Other flexmodels: 
\code{\link{flexmodel}()},
\code{\link{make_ageified_model}()},
\code{\link{make_base_model}()},
\code{\link{make_omicron_model}()},
\code{\link{make_sir_model}()},
\code{\link{make_vaccination_model}()}

Other canned_models: 
\code{\link{make_ageified_model}()},
\code{\link{make_base_model}()},
\code{\link{make_omicron_model}()},
\code{\link{make_sir_model}()},
\code{\link{make_vaccination_model}()}
}
\concept{canned_models}
\concept{flexmodels}
