% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_IP.R
\name{make_omicron_model}
\alias{make_omicron_model}
\title{Make an Omicron invasion model}
\usage{
make_omicron_model(..., do_variant = FALSE)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=init_model]{init_model}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Make an Omicron invasion model
}
\seealso{
Other flexmodels: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{flexmodel}()},
\code{\link{initial_population}()},
\code{\link{make_base_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{tmb_fun}()},
\code{\link{update_tmb_indices}()},
\code{\link{vec_rate}()}

Other canned_models: 
\code{\link{make_base_model}()},
\code{\link{make_vaccination_model}()}
}
\concept{canned_models}
\concept{flexmodels}
