% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{make_ratemat}
\alias{make_ratemat}
\title{Create transition matrix}
\usage{
make_ratemat(
  state,
  params,
  do_ICU = TRUE,
  sparse = FALSE,
  symbols = FALSE,
  indices = FALSE
)
}
\arguments{
\item{state}{named vector of states}

\item{params}{named vector of parameters}

\item{do_ICU}{include additional health utilization compartments}

\item{sparse}{return sparse matrix?}

\item{symbols}{return character (symbol) form? (FIXME: call adjust_symbols here rather than in show_ratemat()?)}

\item{indices}{return indices for lower-level stuff?}
}
\value{
matrix of (daily) transition rates
}
\description{
Defines rates (per day) of flow \emph{from} compartment \code{i}
(row) \emph{to} compartment \code{j} (column).
}
\details{
The rates are as follows:

\eqn{ S to E:  - (\beta_0 / N) S (C_a I_a + C_p I_p + (1-iso_m)C_m I_m + (1-iso_s)C_s I_s) }
\eqn{ E to I_a: }
\eqn{ E to I_p: }
\eqn{ ... }

See \code{\link{read_params}} for parameter definitions.
}
\note{
Base version matches structure of Stanford/Georgia models
\itemize{
  \item flow diagram: see \url{http://covid-measures.stanford.edu/} 'model details' tab
        or \code{../pix/model_schematic.png}
  \item parameter definitions: see \code{params_CI_base.csv}, \code{params_ICU_diffs.csv}
}
}
\examples{
params <- read_params("ICU1.csv")
state <- make_state(params[["N"]],E0=params[["E0"]], use_eigvec=FALSE)
M <- make_ratemat(state,params)
if (require(Matrix)) {
   image(Matrix(M))
}
make_ratemat(state,params,symbols=TRUE)
}
\seealso{
Other classic_macpan: 
\code{\link{add_d_log}()},
\code{\link{add_updated_vaxrate}()},
\code{\link{aggregate_agecats}()},
\code{\link{calibrate_comb}()},
\code{\link{calibrate}()},
\code{\link{check_age_cat_compatibility}()},
\code{\link{check_contact_rate_setting}()},
\code{\link{col_multiply}()},
\code{\link{condense_age}()},
\code{\link{condense_params_vax}()},
\code{\link{condense_state}()},
\code{\link{condense_vax}()},
\code{\link{dev_is_tikz}()},
\code{\link{do_step}()},
\code{\link{expand_params_age}()},
\code{\link{expand_params_desc_age}()},
\code{\link{expand_params_desc_variant}()},
\code{\link{expand_params_desc_vax}()},
\code{\link{expand_params_mistry}()},
\code{\link{expand_params_variant}()},
\code{\link{expand_params_vax}()},
\code{\link{expand_state_age}()},
\code{\link{expand_state_vax}()},
\code{\link{expand_stateval_testing}()},
\code{\link{fix_pars}()},
\code{\link{fix_stored}()},
\code{\link{forecast_ensemble}()},
\code{\link{forecast_sim}()},
\code{\link{getData}()},
\code{\link{get_GI_moments}()},
\code{\link{get_Gbar}()},
\code{\link{get_R0}()},
\code{\link{get_doses_per_day}()},
\code{\link{get_evec}()},
\code{\link{get_kernel_moments}()},
\code{\link{get_opt_pars}()},
\code{\link{get_r}()},
\code{\link{invlink_trans}()},
\code{\link{make_betavec}()},
\code{\link{make_beta}()},
\code{\link{make_jac}()},
\code{\link{make_state}()},
\code{\link{make_test_wtsvec}()},
\code{\link{make_vaxrate}()},
\code{\link{mk_Nvec}()},
\code{\link{mk_agecats}()},
\code{\link{mk_contact_rate_setting}()},
\code{\link{mk_mistry_Nvec}()},
\code{\link{mk_pmat}()},
\code{\link{mk_vaxcats}()},
\code{\link{mle_fun}()},
\code{\link{non_expanded_states}},
\code{\link{rExp}()},
\code{\link{read_params}()},
\code{\link{repair_names_age}()},
\code{\link{restore}()},
\code{\link{run_sim_ageify}()},
\code{\link{run_sim_break}()},
\code{\link{run_sim_loglin}()},
\code{\link{run_sim_mobility}()},
\code{\link{run_sim_range}()},
\code{\link{run_sim}()},
\code{\link{show_ratemat}()},
\code{\link{testify}()},
\code{\link{texify}()},
\code{\link{trans_state_vars}()},
\code{\link{update_contact_rate_setting}()},
\code{\link{update_foi}()},
\code{\link{update_params_mistry}()},
\code{\link{vis_model}()},
\code{\link{write_params}()}
}
\concept{classic_macpan}
