% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmodel_sequential.R
\name{make_splitter}
\alias{make_splitter}
\alias{make_splitter_by_fac}
\alias{make_splitter_eq_rows}
\alias{make_splitter_eq_time}
\alias{make_splitter_date_breaks}
\title{Make Splitter}
\usage{
make_splitter_by_fac(fac)

make_splitter_eq_rows(n_rows)

make_splitter_eq_time(n_days)

make_splitter_date_breaks(date_breaks)
}
\arguments{
\item{fac}{Vector that can be coerced to a \code{\link{factor}} to be
used to split the \code{data} argument passed to the resulting
data splitter.}

\item{n_rows}{Target number of rows in each split of the data.}

\item{n_days}{Target number of days in each split of the data.}

\item{date_breaks}{Date vector giving the target first dates in each split
of the data.}
}
\value{
A function that takes a data frame with a \code{\link{Date}}
column called \code{date} and returns the output of a call to the
\code{\link{split}} function.
}
\description{
Construct a function that will take an observed dataset
(see \code{\link{update_observed}}) and return a list of observed
datasets. See below for descriptions of the different types of
constructors for different types of splitting.
}
\section{Functions}{
\itemize{
\item \code{make_splitter_by_fac()}: Split based on a custom factor vector. The
resulting function makes direct use of the \code{\link{split}} function.

\item \code{make_splitter_eq_rows()}: Split data into groups with an approximately
equal number of rows.

\item \code{make_splitter_eq_time()}: Split data into groups with an
approximately equal number of days.

\item \code{make_splitter_date_breaks()}: Split data into groups based on custom dates
to be used as cut-points

}}
