% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{make_state}
\alias{make_state}
\title{retrieve parameters from a CSV file}
\usage{
make_state(
  N = params[["N"]],
  E0 = params[["E0"]],
  type = "ICU1h",
  state_names = NULL,
  use_eigvec = NULL,
  params = NULL,
  x = NULL,
  normalize = FALSE,
  ageify = NULL,
  vaxify = NULL,
  testify = NULL
)
}
\arguments{
\item{N}{population size}

\item{E0}{initial number exposed}

\item{type}{(character) specify what model type this is intended
for (e.g., \code{"ICU1"}, \code{"CI"}); determines state names}

\item{state_names}{vector of state names, must include S and E}

\item{use_eigvec}{use dominant eigenvector to distribute non-Susc values
to zero: default is to set this to \code{TRUE} if \code{params} is non-NULL}

\item{params}{parameter vector (looked in for N and E0)}

\item{x}{proposed (named) state vector; missing values will be set}

\item{normalize}{(logical) should the state vector be normalized to sum to 1?}

\item{ageify}{expand state vector to include different age groups (??)}

\item{vaxify}{expand state vector to include groups that have 1-2 vaccine doses (??)}

\item{testify}{expand state vector to include testing compartments (untested, neg waiting, pos waiting, pos received) ?}
}
\description{
retrieve parameters from a CSV file
}
\details{
The parameters that must be set are:

\eqn{   N:  }  population size

\eqn{   \beta_0:  }  transmission rate

\eqn{   1/\sigma:  }  mean \emph{latent} period

\eqn{   1/\gamma_a:  }  mean \emph{infectious} period for asymptomatic individuals

\eqn{   ... }

generate initial state vector
}
\note{
\code{"CI"} refers to the Stanford group's
    "covid intervention" model.
}
\examples{
p <- read_params("ICU1.csv")
make_state(N=1e6,E0=1)
make_state(params=p)
}
\seealso{
Other classic_macpan: 
\code{\link{add_d_log}()},
\code{\link{add_updated_vaxrate}()},
\code{\link{aggregate_agecats}()},
\code{\link{calibrate_comb}()},
\code{\link{calibrate}()},
\code{\link{check_age_cat_compatibility}()},
\code{\link{check_contact_rate_setting}()},
\code{\link{col_multiply}()},
\code{\link{condense_age}()},
\code{\link{condense_params_vax}()},
\code{\link{condense_state}()},
\code{\link{condense_vax}()},
\code{\link{dev_is_tikz}()},
\code{\link{do_step}()},
\code{\link{expand_params_age}()},
\code{\link{expand_params_desc_age}()},
\code{\link{expand_params_desc_variant}()},
\code{\link{expand_params_desc_vax}()},
\code{\link{expand_params_mistry}()},
\code{\link{expand_params_variant}()},
\code{\link{expand_params_vax}()},
\code{\link{expand_state_age}()},
\code{\link{expand_state_vax}()},
\code{\link{expand_stateval_testing}()},
\code{\link{fix_pars}()},
\code{\link{fix_stored}()},
\code{\link{forecast_ensemble}()},
\code{\link{forecast_sim}()},
\code{\link{getData}()},
\code{\link{get_GI_moments}()},
\code{\link{get_Gbar}()},
\code{\link{get_R0}()},
\code{\link{get_doses_per_day}()},
\code{\link{get_evec}()},
\code{\link{get_kernel_moments}()},
\code{\link{get_opt_pars}()},
\code{\link{get_r}()},
\code{\link{invlink_trans}()},
\code{\link{make_betavec}()},
\code{\link{make_beta}()},
\code{\link{make_jac}()},
\code{\link{make_ratemat}()},
\code{\link{make_test_wtsvec}()},
\code{\link{make_vaxrate}()},
\code{\link{mk_Nvec}()},
\code{\link{mk_agecats}()},
\code{\link{mk_contact_rate_setting}()},
\code{\link{mk_mistry_Nvec}()},
\code{\link{mk_pmat}()},
\code{\link{mk_vaxcats}()},
\code{\link{mle_fun}()},
\code{\link{non_expanded_states}},
\code{\link{rExp}()},
\code{\link{read_params}()},
\code{\link{repair_names_age}()},
\code{\link{restore}()},
\code{\link{run_sim_ageify}()},
\code{\link{run_sim_break}()},
\code{\link{run_sim_loglin}()},
\code{\link{run_sim_mobility}()},
\code{\link{run_sim_range}()},
\code{\link{run_sim}()},
\code{\link{show_ratemat}()},
\code{\link{testify}()},
\code{\link{texify}()},
\code{\link{trans_state_vars}()},
\code{\link{update_contact_rate_setting}()},
\code{\link{update_foi}()},
\code{\link{update_params_mistry}()},
\code{\link{vis_model}()},
\code{\link{write_params}()}
}
\concept{classic_macpan}
