% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testify.R
\name{make_test_posvec}
\alias{make_test_posvec}
\title{make vector of test positivity}
\usage{
make_test_posvec(params)
}
\arguments{
\item{params}{parameter vector}
}
\description{
make vector of test positivity
}
