% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{make_unflexmodel}
\alias{make_unflexmodel}
\title{Represent a Standard Model as a flexmodel}
\usage{
make_unflexmodel(
  params,
  state = NULL,
  start_date = "2020-03-20",
  end_date = "2020-05-1",
  params_timevar = NULL,
  step_args = list()
)
}
\description{
Represent a Standard Model as a flexmodel
}
