% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{make_vaccination_model}
\alias{make_vaccination_model}
\title{Make a Two-Dose Vaccination Model}
\usage{
make_vaccination_model(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=init_model]{init_model}}
  \describe{
    \item{\code{params}}{a \code{param_pansim} object}
    \item{\code{state}}{a \code{state_pansim} object}
    \item{\code{struc}}{a \code{struc_pansim} object}
    \item{\code{start_date}}{simulation start date}
    \item{\code{end_date}}{simulation end date}
    \item{\code{params_timevar}}{data frame with scheduling for piece-wise
constant parameter variation (TODO: direct to other help pages)}
    \item{\code{do_hazard}}{should hazard simulation steps be used?
(https://canmod.net/misc/flex_specs#v0.0.5) -- only used
if \code{spec_ver_gt('0.0.4')}}
  }}
}
\description{
Make a Two-Dose Vaccination Model
}
\seealso{
Other flexmodels: 
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_tmb_indices}()},
\code{\link{init_model}()},
\code{\link{make_base_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{state_param_sum}()},
\code{\link{tmb_fun}()},
\code{\link{vec_rate}()}

Other canned_models: 
\code{\link{make_base_model}()}
}
\concept{canned_models}
\concept{flexmodels}
