% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{make_vaxrate}
\alias{make_vaxrate}
\title{generate per capita daily vaccination rates}
\usage{
make_vaxrate(state, params)
}
\arguments{
\item{state}{state vector (an object of class `state_pansim`)}

\item{params}{model parameters (an object of class `params_pansim`)}
}
\description{
generate per capita daily vaccination rates
}
