% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{mat_rate}
\alias{mat_rate}
\title{Specify Matrix of Rates}
\usage{
mat_rate()
}
\description{
Not implemented
}
\seealso{
Other flexmodel_definition_functions: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{flexmodel}()},
\code{\link{initial_population}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{update_disease_free_state}()},
\code{\link{update_linearized_params}()},
\code{\link{vec_rate}()}

Other rate_functions: 
\code{\link{add_rate}()},
\code{\link{rep_rate}()},
\code{\link{vec_rate}()}
}
\concept{flexmodel_definition_functions}
\concept{rate_functions}
