% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{matmult}
\alias{matmult}
\alias{\%*\%,struc,struc-method}
\alias{\%*\%,struc_dimnames,struc_dimnames-method}
\alias{\%*\%,struc,struc_dimnames-method}
\alias{\%*\%,struc_dimnames,struc-method}
\alias{kronecker,struc,struc-method}
\title{Symbolic Matrix Multiplication}
\usage{
\S4method{\%*\%}{struc,struc}(x, y)

\S4method{\%*\%}{struc_dimnames,struc_dimnames}(x, y)

\S4method{\%*\%}{struc,struc_dimnames}(x, y)

\S4method{\%*\%}{struc_dimnames,struc}(x, y)

\S4method{kronecker}{struc,struc}(X, Y)
}
\arguments{
\item{x}{Operand as a \code{\link{struc}} object}

\item{y}{Operand as a \code{\link{struc}} object}

\item{X}{Operand as a \code{\link{struc}} object}

\item{Y}{Operand as a \code{\link{struc}} object}
}
\description{
Symbolic Matrix Multiplication
}
\concept{symbolic}
