% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{McMasterPandemic-package}
\alias{McMasterPandemic}
\alias{McMasterPandemic-package}
\title{\pkg{McMasterPandemic}}
\description{
This R package provides compartmental epidemic models for
forecasting and analysis of infectious diseases
}
\details{
The \href{https://canmod.github.io/macpan-book/}{user guide} provides
extensive information on the
\href{https://canmod.github.io/macpan-book/index.html#history-and-motivation}{history}
of McMasterPandemic,
\href{https://canmod.github.io/macpan-book/index.html#installation}{installation},
building a
\href{https://canmod.github.io/macpan-book/model-initialization.html}{simple SIR model},
and a library of
\href{https://canmod.github.io/macpan-book/examples.html}{examples}.

If you clicked on the \href{https://canmod.github.io/macpan-book/index.html#history-and-motivation}{history}
link above, you will have read that \code{McMasterPandemic} started out as
a specific set of models for the COVID-19 pandemic in Canada.
To get started with this functionality for using these specific models,
read the introductory vignette, which can be viewed via
\code{vignette("getting_started",package="McMasterPandemic")}.
Use \code{browseVignettes("McMasterPandemic")} to find other vignettes
for this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mac-theobio/McMasterPandemic}
}

}
\author{
\strong{Maintainer}: Ben Bolker \email{bolker@mcmaster.ca}

Authors:
\itemize{
  \item Steve Walker
  \item David Earn
  \item Morgan Kain
  \item Mike Li
  \item Jonathan Dushoff
}

Other contributors:
\itemize{
  \item Zach Levine [contributor]
  \item Matthew So [contributor]
  \item Weiguang Guan [contributor]
  \item Darren Flynn-Primrose [contributor]
}

}
