% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructing_vectors.R
\name{merge_named_vec_attr}
\alias{merge_named_vec_attr}
\title{Merge Named Vector Attribute}
\usage{
merge_named_vec_attr(v, u, a)
}
\arguments{
\item{v}{vector with an attribute, \code{a}}

\item{u}{vector with attribute \code{a} that will be merged into
\code{a} in \code{v} using \code{\link{merge_named_vectors}}}

\item{a}{name of the atribute to merge}
}
\description{
Merge Named Vector Attribute
}
