% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{merge_named_vectors}
\alias{merge_named_vectors}
\title{Merge One Vector into Another by Name}
\usage{
merge_named_vectors(v, u)
}
\arguments{
\item{v}{named vector or list}

\item{u}{named vector of list}
}
\description{
If an item in \code{u} has the same name as an item
in \code{v} then replace the value in \code{v} with that
in \code{u}, otherwise create a new element in \code{v}.
}
