% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{mk_agecats}
\alias{mk_agecats}
\title{construct age categories}
\usage{
mk_agecats(min = 1, max = 100, da = 10)
}
\arguments{
\item{min}{minimum age}

\item{max}{maximum age}

\item{da}{age bin width}
}
\description{
construct age categories
}
