% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{mk_contact_rate_setting}
\alias{mk_contact_rate_setting}
\title{Helper function to generate a vector of setting-specific contact rates}
\usage{
mk_contact_rate_setting(values_to_update = NULL)
}
\arguments{
\item{values_to_update}{named list of values to update in default list of setting-specific contact rates}
}
\value{
named list of setting-specific contact rates
}
\description{
Helper function to generate a vector of setting-specific contact rates
}
\examples{
mk_contact_rate_setting()
mk_contact_rate_setting(list(school = 0))
}
