% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{mk_mistry_Cmat}
\alias{mk_mistry_Cmat}
\title{Make contact matrix using Mistry et al. approach}
\usage{
mk_mistry_Cmat(
  weights = c(household = 4.11, school = 11.41, work = 8.07, community = 2.79),
  province = "Ontario"
)
}
\arguments{
\item{weights}{named list containing setting-specific weights (in units of average contacts in the given setting per individual of age i with individuals of age j per day)}

\item{province}{province for which to construct the matrix (if NULL, make a Canada-wide contact matrix)}
}
\value{
matrix of average contacts between individuals of ages i and j per individual of age i per day
}
\description{
Make contact matrix using Mistry et al. approach
}
\examples{
mk_mistry_Cmat()
}
