% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{mk_mistry_pmat}
\alias{mk_mistry_pmat}
\title{Make contact matrix using Mistry et al. approach}
\usage{
mk_mistry_pmat(
  weights = c(household = 4.11, school = 11.41, work = 8.07, community = 2.79),
  province = "Ontario",
  age_cat = NULL
)
}
\arguments{
\item{weights}{named list containing setting-specific weights (in units of average contacts in the given setting per individual of age i with individuals of age j per day)}

\item{province}{province for which to construct the matrix (if NULL, make a Canada-wide contact matrix)}

\item{age_cat}{(optional) list of age groups to aggregate ages in; use mk_agecats() to generate (must start with 0 and end with 84); default is single ages starting with 0 and up to 83, then a single 84+ category}
}
\value{
matrix of average contacts between individuals of ages i and j per individual of age i per unit time (day)
}
\description{
Make contact matrix using Mistry et al. approach
}
\examples{
mk_mistry_pmat()
}
