% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{mk_Nvec}
\alias{mk_Nvec}
\title{generate a population distribution (as a vector of counts)}
\usage{
mk_Nvec(
  age_cat = mk_agecats(),
  Ntot = 1e+06,
  dist = c("unif", "rand"),
  names = FALSE
)
}
\arguments{
\item{age_cat}{age categories (made with `mk_agecats()`)}

\item{Ntot}{total population size}

\item{dist}{(character) either "unif" for uniform or "rand" for a random population distribution}

\item{names}{(logical) return a named vector? (names are given age categories)}
}
\description{
generate a population distribution (as a vector of counts)
}
