% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{mk_vaxrates}
\alias{mk_vaxrates}
\title{generate per capita daily vaccination rates}
\usage{
mk_vaxrates(params, doses_per_day = list(vax = 10000))
}
\arguments{
\item{params}{model parameters (an object of class `params_pansim`)}

\item{doses_per_day}{named list of doses per day in the region by vax category and any other subcat (e.g. age)}
}
\description{
generate per capita daily vaccination rates
}
