% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{names_from_values}
\alias{names_from_values}
\alias{names_from_values.EpiMatrixDerived}
\alias{names_from_values.struc}
\title{Names from Values}
\usage{
names_from_values(x)

\method{names_from_values}{EpiMatrixDerived}(x)

\method{names_from_values}{struc}(x)
}
\arguments{
\item{x}{object from which to create element names from its values}
}
\description{
Get element names of an array-like object by processing the
values of the object.
}
