% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_and_strings.R
\name{names_sim_hist}
\alias{names_sim_hist}
\alias{names_sim_cond}
\title{Simulation Variable Names}
\usage{
names_sim_hist(model)

names_sim_cond(model)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}
}
\value{
character vector of names
}
\description{
Return vector of names of simulation variables. The \code{names_sim_hist}
function returns the names of variables associated with the table returned
by \code{\link{simulation_history}}. The \code{names_sim_cond} function
returns the names of variables in the \code{simulation_history} if
the \code{condense} argument is \code{FALSE} or if
\code{\link{simulate_ensemble}} is used.
}
\section{Functions}{
\itemize{
\item \code{names_sim_cond()}: testing

}}
