% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{names_to_values}
\alias{names_to_values}
\title{Names to Values}
\usage{
names_to_values(x)
}
\arguments{
\item{x}{object to fill with its names}
}
\description{
Replace the values of an object with its names.
}
\details{
This function only makes sense for objects that have the same shape
as their names (e.g. vectors, EpiMatrix ... well that might be it actually)
}
