% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\docType{data}
\name{non_expanded_states}
\alias{non_expanded_states}
\alias{test_extensions}
\alias{test_accumulators}
\alias{asymp_cat}
\alias{severe_cat}
\alias{cryptic_cat}
\alias{vacc_extensions}
\alias{vacc_accumulators}
\title{global variables for testify expansion}
\format{
An object of class \code{character} of length 2.

An object of class \code{character} of length 4.

An object of class \code{character} of length 2.

An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.

An object of class \code{character} of length 2.
}
\usage{
non_expanded_states

test_extensions

test_accumulators

asymp_cat

severe_cat

cryptic_cat

vacc_extensions

vacc_accumulators
}
\description{
global variables for testify expansion

## FIXME: document these for real!
##' global variables for vaccify expansion
#' @export
non_expanded_states <- c()
}
\keyword{datasets}
