% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ont_all}
\alias{ont_all}
\alias{ont_cal1}
\alias{ont_2brks}
\alias{ont_cal_2brks}
\title{Sample data and fit from Ontario (up to 18 April 2020)}
\format{
a data frame ...

a calibrated object ...
Sample COVID19 calibration

a calibrated object ...
Sample COVID19 calibration - only two breakpoints
}
\source{
mike's repo
}
\usage{
ont_all

ont_cal1

ont_cal_2brks
}
\description{
Sample data and fit from Ontario (up to 18 April 2020)
}
\keyword{datasets}
