% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{opt_obj}
\alias{opt_obj}
\alias{convergence_info}
\title{Optimizer Object}
\usage{
opt_obj(model)

convergence_info(model)
}
\arguments{
\item{model}{object of class \code{flexmodel_calibrated}}
}
\description{
Get the object returned by the optimizer used to calibrate a
\code{flexmodel_to_calibrate} object. The \code{convergence_info}
function gets convergence information in case it is buried within
the \code{opt_obj}.
}
