% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmodel_calibrate.R
\name{optim_flexmodel}
\alias{optim_flexmodel}
\alias{nlminb_flexmodel}
\alias{bbmle_flexmodel}
\alias{fitted.flexmodel_calibrated}
\title{Wrapped Optimizers for Flexmodels}
\usage{
optim_flexmodel(model, ...)

nlminb_flexmodel(model, ...)

bbmle_flexmodel(model, ...)

\method{fitted}{flexmodel_calibrated}(object, ...)
}
\arguments{
\item{model}{a \code{\link{flexmodel}} object}

\item{...}{additional arguments to pass to the wrapped optimizer
to their calibrated values?}

\item{object}{a \code{\link{flexmodel}} object (for S3 method consistency)}
}
\description{
Currently there are \code{optim_flexmodel}, \code{nlminb_flexmodel},
and \code{bbmle_flexmodel}
}
