% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{parse_name_sum}
\alias{parse_name_sum}
\title{Recursive function to parse a formula containing sums of names}
\usage{
parse_name_sum(x)
}
\description{
Recursive function to parse a formula containing sums of names
}
