% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{parse_opt_form}
\alias{parse_opt_form}
\title{Recursive function to parse a formula containing parameter
optimization specifications}
\usage{
parse_opt_form(x, e = NULL)
}
\description{
trans1_param ~ trans2_prior(hyperparameters, ...)
}
\details{
trans1_param can be:
  (1) a symbol/name (or sum of symbols)
  (2) an expression that evaluates (in the environment of the formula)
      to a character vector
  (3) a literal length-1 character vector

if trans1_param is a symbol:
trans1 = name of a valid parameter transformation, giving the scale
         on which the optimizer treats the parameter(s)
param = name of a parameter in the parameter vector

if trans1_param is an expression
}
