% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{parse_opt_param}
\alias{parse_opt_param}
\title{Parse Fitted Parameter Formula}
\usage{
parse_opt_param(x, params, params_timevar = NULL)
}
\arguments{
\item{x}{opt_params formula}

\item{params}{parameter vector}

\item{regex_param_nm}{should the parameter name be treated as a regular
expression used to produce a vector of parameter names}
}
\description{
\code{trans1_parameter ~ trans2_prior(hyperparameter1, ..., laplace = FALSE)}
}
\details{
If \code{parameter}
}
