% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pfun}
\alias{pfun}
\title{identify locations within matrix}
\usage{
pfun(from, to, mat, value = FALSE, recycle = FALSE)
}
\arguments{
\item{from}{string identifying 'from' compartment(s) for flows}

\item{to}{string identifying 'to' compartment(s) for flows}

\item{mat}{flow matrix with appropriate dimnames}

\item{value}{(logical) return character (TRUE) or index (FALSE) ?}

\item{recycle}{(logical) if necessary, recycle vectors (only if length-1) as necessary}
}
\description{
checks global option "macpan_pfun_method" for method: "startsWith" (use efficient method, "grep" (use grep/regex), "both" (testing only! try both and compare)
}
