% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.fit_pansim}
\alias{plot.fit_pansim}
\title{plot forecasts from fits}
\usage{
\method{plot}{fit_pansim}(
  x,
  data = NULL,
  break_dates = NULL,
  dlspace = 1,
  limspace = 10,
  add_tests = FALSE,
  add_ICU_cap = FALSE,
  mult_var = NULL,
  predict_args = NULL,
  directlabels = FALSE,
  ...
)
}
\arguments{
\item{data}{original time series data}

\item{dlspace}{spacing for direct labels (not working)}

\item{limspace}{extra space (in days) to add to make room for direct labels}

\item{add_tests}{plot newTests/1000?}

\item{add_ICU_cap}{include horizontal lines showing ICU capacity?}

\item{mult_var}{variable in data set indicating multiple forecast types to compare}

\item{directlabels}{use direct labels?}

\item{forecast}{a forecast data frame}

\item{breaks}{breakpoints}

\item{add_data}{include data as points?}
}
\description{
plot forecasts from fits
}
\examples{
plot(ont_cal1)
ont_trans <- trans_state_vars(ont_all)
plot(ont_cal1,data=ont_trans)
plot(ont_cal1,data=ont_trans, add_tests=TRUE)
plot(ont_cal_2brks,predict_args=list(ensemble=TRUE))
}
