% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.pansim}
\alias{plot.pansim}
\title{plot method for simulations}
\usage{
\method{plot}{pansim}(
  x,
  drop_states = c("t", "S", "R", "E", "I", "X", "incidence"),
  keep_states = NULL,
  condense = FALSE,
  log = FALSE,
  show_times = TRUE,
  ...
)
}
\arguments{
\item{x}{fitted \code{pansim} object}

\item{drop_states}{states to \emph{exclude} from plot}

\item{keep_states}{states to \emph{include} in plot (overrides \code{drop_states})}

\item{condense}{condense states (e.g. all ICU states -> "ICU") before plotting?  See \code{\link{condense.pansim}}}

\item{log}{plot y-axis on log scale?}

\item{show_times}{indicate times when parameters changed?}

\item{...}{additional arguments to \code{\link{condense.pansim}}}
}
\description{
plot method for simulations
}
