% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.predict_pansim}
\alias{plot.predict_pansim}
\title{plot forecasts from fits}
\usage{
\method{plot}{predict_pansim}(
  x,
  data = NULL,
  break_dates = NULL,
  dlspace = 1,
  limspace = 10,
  add_tests = FALSE,
  add_ICU_cap = FALSE,
  mult_var = NULL,
  directlabels = TRUE,
  log = TRUE,
  log_lwr = 1,
  ...
)
}
\arguments{
\item{x}{a calibrated object (result from \code{\link{calibrate}}) or a prediction (from \code{\link{predict.fit_pansim}})}

\item{data}{original time series data}

\item{break_dates}{breakpoints}

\item{dlspace}{spacing for direct labels (not working)}

\item{limspace}{extra space (in days) to add to make room for direct labels}

\item{add_tests}{plot newTests/1000?}

\item{add_ICU_cap}{include horizontal lines showing ICU capacity?}

\item{mult_var}{variable in data set indicating multiple forecast types to compare}

\item{directlabels}{use direct labels?}

\item{log}{use a log10 scale for the y axis?}

\item{log_lwr}{lower limit when using log scale}

\item{...}{extra arguments (unused)}
}
\description{
plot forecasts from fits
}
\examples{
plot(ont_cal1)
ont_trans <- trans_state_vars(ont_all)
plot(ont_cal1,data=ont_trans)
plot(ont_cal1,data=ont_trans, add_tests=TRUE)
plot(ont_cal1,data=ont_trans, predict_args=list(end_date="2020-07-01"))
\donttest{
## FIXME: don't try these until we have an example where ensemble works
## pp <- predict(ont_cal_2brks, ensemble=TRUE)
## plot(pp)
## plot(pp, data=ont_trans)
}
}
