% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_res_by_age}
\alias{plot_res_by_age}
\title{Plot age-structured simulation result faceted by age categories}
\usage{
plot_res_by_age(res, drop_states = NULL, condense_I = FALSE)
}
\arguments{
\item{res}{age-structured simulation result}

\item{drop_states}{states to \emph{exclude} from plot}
}
\value{

}
\description{
Plot age-structured simulation result faceted by age categories
}
