% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{predfun}
\alias{predfun}
\title{run simulation with specified parameters; extract results
matching dates and variable order in data}
\usage{
predfun(
  beta0,
  E0,
  data,
  params,
  start_date = "10-Feb-2020",
  values_only = TRUE
)
}
\arguments{
\item{beta0}{baseline transmission}

\item{E0}{starting value}

\item{data}{data (for subsetting/matching)}

\item{params}{parameters}

\item{start_date}{start date}

\item{values_only}{return a vector rather than a data frame}
}
\description{
run simulation with specified parameters; extract results
matching dates and variable order in data
}
