% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{prep_res_for_plotting}
\alias{prep_res_for_plotting}
\title{Prepare age-structured simulation results data frame for plotting}
\usage{
prep_res_for_plotting(
  res,
  drop_states = NULL,
  condense_I = FALSE,
  split_age_vax = FALSE
)
}
\arguments{
\item{res}{age-structured simulation result}

\item{drop_states}{epidemiological states to \emph{exclude} from plot}

\item{condense_I}{condense all the different I (infectious) states to a single I state? (??)}

\item{split_age_vax}{should the age and vaccination category be split into different columns?}
}
\description{
Prepare age-structured simulation results data frame for plotting
}
