% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/david.R
\name{print.panparams}
\alias{print.panparams}
\title{Print method for \code{panpamams} objects}
\usage{
\method{print}{panparams}(x)
}
\arguments{
\item{x}{a \code{panparams} object}
}
\description{
Print method for \code{panpamams} objects
}
