% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{r_tmb_comparable}
\alias{r_tmb_comparable}
\title{Compare R and TMB Engines}
\usage{
r_tmb_comparable()
}
\description{
Set options so that the TMB and R engines
can be compared.
}
\details{
This is useful for testing.
}
