% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{rate}
\alias{rate}
\title{Define Rate for Single Element of Rate Matrix}
\usage{
rate(from, to, formula, state, params, ratemat)
}
\arguments{
\item{from}{from state}

\item{to}{to state}

\item{formula}{one-sided formula defining the rate with reference
to the parameters and state variables}

\item{state}{state_pansim object}

\item{params}{param_pansim object}

\item{ratemat}{rate matrix}
}
\description{
Define Rate for Single Element of Rate Matrix
}
