% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{rate_matrix_lookup}
\alias{rate_matrix_lookup}
\title{Rate Matrix Loopup Table}
\usage{
rate_matrix_lookup(ratemat)
}
\arguments{
\item{ratemat}{rate matrix}

\item{state}{state_pansim object}
}
\description{
Rate Matrix Loopup Table
}
