% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_params}
\alias{read_params}
\title{read simulation parameters}
\usage{
read_params(
  fn,
  value_col = "Value",
  symbol_col = "Symbol",
  desc_col = "Parameter"
)
}
\arguments{
\item{fn}{file name (CSV file containing at least value and symbol columns);  file should either be findable from current working directory or available in the built-in \code{params} directory}

\item{value_col}{name of column containing values}

\item{symbol_col}{name of column containing symbols}

\item{desc_col}{name of (optional) column containing descriptions}
}
\description{
read simulation parameters
}
