% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{rep_rate}
\alias{rep_rate}
\title{Repeat a Rate for Several Rate Matrix Elements}
\usage{
rep_rate(model, from, to, formula, mapping = c("pairwise", "blockwise"))
}
\arguments{
\item{model}{flexmodel}

\item{formula}{formula or length-1 character vector}

\item{indices}{two-column matrix of indices with column names "from_pos"
and "to_pos", locaing elements of the rate matrix in model}
}
\description{
Repeat a Rate for Several Rate Matrix Elements
}
