% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{repair_age_names}
\alias{repair_age_names}
\title{Repair ageified names}
\usage{
repair_age_names(x)
}
\arguments{
\item{x}{a named list or data.frame}
}
\value{
a named list or data.frame
}
\description{
sometimes age categories get changed from 0-10 to 0.10 and 60+ to 60.
replace each substituted period with the correct character
}
\examples{
state <- c(S_0.10 = 100, E_60. = 1)
repair_age_names(state)
}
