% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restore}
\alias{restore}
\title{convert a parameter vector back to a structured list
Like \code{\link{relist}}, but adds the capability to fill in
parameters from a "fixed-parameter" vector}
\usage{
restore(flesh, skeleton, fixed = NULL)
}
\arguments{
\item{flesh}{a vector to be restored}

\item{skeleton}{a list, the structure of which determines the structure of the result}

\item{fixed}{a list which determines extra components to fill in}
}
\description{
convert a parameter vector back to a structured list
Like \code{\link{relist}}, but adds the capability to fill in
parameters from a "fixed-parameter" vector
}
\note{
Depends at present on the names of the unlisted object; may be fragile
}
\examples{
opt_pars <- list(log_E0=4, log_beta0=-1, log_rel_beta0=c(-1,-1), log_nb_disp=0)
restore(unlist(opt_pars),opt_pars)
invlink_trans(restore(unlist(opt_pars),opt_pars))
}
