% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{rowcolsums}
\alias{rowcolsums}
\alias{rowSums,struc-method}
\alias{rowSums,struc_dimnames-method}
\alias{colSums,struc-method}
\alias{colSums,struc_dimnames-method}
\title{Symbolic Row and Column Sums}
\usage{
\S4method{rowSums}{struc}(x, na.rm = FALSE, dims = 1L)

\S4method{rowSums}{struc_dimnames}(x, na.rm = FALSE, dims = 1L)

\S4method{colSums}{struc}(x, na.rm = FALSE, dims = 1L)

\S4method{colSums}{struc_dimnames}(x, na.rm = FALSE, dims = 1L)
}
\arguments{
\item{x}{\code{\link{struc}} object}
}
\description{
Symbolic Row and Column Sums
}
\concept{symbolic}
