% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Run the McMasterPandemic Shiny}
\usage{
run_shiny(useBrowser = TRUE)
}
\arguments{
\item{useBrowser}{Open the shiny in the browser.}
}
\description{
Run the McMasterPandemic Shiny
}
