% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{run_sim_ageify}
\alias{run_sim_ageify}
\title{Agify a basic simulation}
\usage{
run_sim_ageify(
  base_params,
  base_state,
  age_cat = mk_agecats(),
  beta0vec = NULL,
  Cmat = mk_Cmat(),
  Nvec = mk_Nvec(),
  ...
)
}
\arguments{
\item{base_params}{base parameter set without age structure (e.g. generated by `read_params()`)}

\item{base_state}{base state vector without age structure (e.g. generated by `make_state()`)}

\item{age_cat}{age categories (e.g. generated by `mk_agecats()`)}

\item{beta0vec}{vector of beta0 values for each age group (e.g. generated by `mk_beta0vec()`)}

\item{Cmat}{contact matrix (e.g. generated by `mk_Cmat()`)}

\item{Nvec}{population distribution (e.g. generated by `mk_Nvec()`)}

\item{...}{additional arguments to `run_sim()`}
}
\value{
simulation results (an object of class `pansim`)
}
\description{
Wrapper function for `run_sim()` that incorporates existing ageify tools
}
\examples{
params <- update(read_params("PHAC_testify.csv"), testing_intensity=0)
state <- make_state(params=params)
run_sum_ageify(base_params = params, base_state = state)
}
