% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{run_sim_loglin}
\alias{run_sim_loglin}
\title{run with log-linear model applied to beta}
\usage{
run_sim_loglin(
  params,
  extra_pars = NULL,
  time_args = NULL,
  sim_args = list(),
  return_timevar = FALSE,
  ...
)

run_sim_loglin(
  params,
  extra_pars = NULL,
  time_args = NULL,
  sim_args = list(),
  return_timevar = FALSE,
  ...
)
}
\arguments{
\item{params}{parameters}

\item{extra_pars}{??}

\item{time_args}{list containing \code{break_dates}}

\item{sim_args}{parameters to pass to \code{\link{run_sim}}}

\item{return_timevar}{return data frame of beta by time?}

\item{...}{additional arguments to \code{run_sim}}
}
\description{
run with log-linear model applied to beta

run with transmission propto relative mobility
}
