% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{select_epi_mats}
\alias{select_epi_mats}
\alias{select_derived_mats}
\alias{select_input_mats}
\title{Select Epidemiological Matrices}
\usage{
select_derived_mats(mat_list)

select_input_mats(mat_list)
}
\arguments{
\item{mat_list}{an \code{\link{EpiMatrixList}} object.}
}
\description{
Select different types of epidemiological matrices from an
\code{\link{EpiMatrixList}} object.
}
\section{Functions}{
\itemize{
\item \code{select_derived_mats()}: Select \code{\link{EpiMatrixDerived-class}}
objects

\item \code{select_input_mats()}: Select \code{\link{EpiMatrixInput-class}}
objects

}}
