% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmodel_sequential.R
\name{sequential}
\alias{sequential}
\title{Fitting Models Sequentially}
\description{
Inputs:
1. observed data splitting function (see \code{\link{make_splitter}})
2. a start date offset
3. time varying parameters
}
