% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_old.R
\name{sim_fun}
\alias{sim_fun}
\title{calibrate and run simulation
obsolete, replaced by forecast_sim?}
\usage{
sim_fun(
  target,
  base_params,
  start_date = "1-Mar-2020",
  reg_date,
  end_date = "1-May-2020",
  init_var = "H",
  sim_args = NULL,
  return_val = c("sim", "aggsim", "vals_only")
)
}
\arguments{
\item{target}{calibration values: mean generation interval, initial growth rate, initial value of calibration variable (\code{X0})}

\item{base_params}{base parameters}

\item{start_date}{starting date for simulations (should be far enough before the start of the data to allow sorting-out of initial conditions)}

\item{reg_date}{first day of calibration data used in log-linear regression}

\item{end_date}{ending date for simulation}

\item{init_var}{variable to use in initial-condition calibration matching}

\item{sim_args}{list of additional arguments to pass to \code{\link{run_sim}}}

\item{return_val}{return values: "sim"=full simulation (wide format); "aggsim"=aggregated and pivoted simulation output, "vals_only" = vector containing only values output from simulation}
}
\description{
calibrate and run simulation
obsolete, replaced by forecast_sim?
}
