% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{sim_range}
\alias{sim_range}
\title{Run simulation across a range of times
Possible refactorins step for run_sim}
\usage{
sim_range(
  params,
  state,
  steps = 100,
  dt = 1,
  stoch = c(obs = FALSE, proc = FALSE),
  ratemat_args = NULL,
  step_args = NULL
)
}
\arguments{
\item{params}{named vector of parameters}

\item{state}{named vector of states}

\item{steps}{(number of steps to take)}

\item{dt}{time step (days)}

\item{stoch}{stochastic simulation? logical vector for observation and process noise}

\item{ratemat_args}{additional arguments to pass to \code{make_ratemat}}

\item{step_args}{additional arguments to pass to \code{do_step}}
}
\description{
Run simulation across a range of times
Possible refactorins step for run_sim
}
