% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{simulate_ensemble}
\alias{simulate_ensemble}
\title{Ensemble Simulation}
\usage{
simulate_ensemble(
  model,
  sim_params_matrix = NULL,
  qvec = c(lwr = 0.025, value = 0.5, upr = 0.975),
  ...
)
}
\description{
Given a sample of parameter vectors, simulate the condensed state simulation
history for each ... TODO
}
\seealso{
Other simulation: 
\code{\link{simulation_history}()}
}
\concept{simulation}
