% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{simulation_condensed}
\alias{simulation_condensed}
\title{Condensed set of Simulated Variables}
\usage{
simulation_condensed(model, add_dates = TRUE, sim_params = NULL)
}
\description{
Condensed set of Simulated Variables
}
