% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{sir_sims_for_tests}
\alias{sir_sims_for_tests}
\alias{read_sir}
\alias{read_sir_sim}
\alias{list_sir}
\alias{list_sir_sim}
\title{SIR Simulations for Tests}
\usage{
read_sir(sir_name)

read_sir_sim(sir_sim_name)

list_sir()

list_sir_sim()
}
\arguments{
\item{sir_name}{Name of test SIR model}

\item{sir_sim_name}{Name of simulations from a test SIR model}
}
\description{
SIR Simulations for Tests
}
\section{Functions}{
\itemize{
\item \code{read_sir()}: Read a test SIR model

\item \code{read_sir_sim()}: Read simulations from a test SIR model

\item \code{list_sir()}: List available test SIR models

\item \code{list_sir_sim()}: List available simulations from test SIR
models

}}
