% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{spec_check}
\alias{spec_check}
\alias{feature_check}
\title{Spec and Feature Checks}
\usage{
spec_check(
  introduced_version,
  feature,
  exception_type = c("stop", "warning", "message")
)

feature_check(introduced_version, feature)
}
\arguments{
\item{introduced_version}{string with the version of the spec
(e.g. \code{"0.0.5"})}

\item{feature}{free-form text describing the feature}

\item{exception_type}{what type of exception to throw if the check
is not passed}
}
\value{
No return value. Called to get specific error messages when required.
}
\description{
\code{spec_check} will throw an error if a feature is being used that is
not supported yet. \code{feature_check} will throw an error if a feature
is not being used, even though it is required by the spec version being used.
}
