% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{spec_version}
\alias{spec_version}
\alias{spec_ver_eq}
\alias{spec_ver_gt}
\alias{spec_ver_lt}
\alias{spec_ver_btwn}
\title{Spec Version}
\usage{
spec_version()

spec_ver_eq(version)

spec_ver_gt(version)

spec_ver_lt(version)

spec_ver_btwn(version_left, version_right)
}
\arguments{
\item{version}{string with the version of the spec
(e.g. \code{"0.0.5"})}
}
\value{
\code{svlist} object with the specification version.
}
\description{
Return the specification version being assumed by the package.
Equivalent to \code{parse_version(getOption('MP_flex_spec_version'))}.
}
\details{
Specification Document: https://canmod.net/misc/flex_specs.

\code{spec_ver_eq}, \code{spec_ver_gt}, \code{spec_ver_lt}, and
\code{spec_ver_btwn} return logical values indicating whether or not
the \code{spec_version} is equal to, greater than, less than or between
a user-specified version
}
