% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructing_vectors.R
\name{special_vector_construction}
\alias{special_vector_construction}
\alias{layered_zero_state}
\alias{layered_const_state}
\alias{const_named_vector}
\alias{self_named_vector}
\title{Special Vector Construction}
\usage{
layered_zero_state(...)

layered_const_state(..., cnst)

const_named_vector(nms, cnst)

self_named_vector(nms)
}
\arguments{
\item{...}{character vectors with sub-model state names}

\item{nms}{}
}
\description{
Special Vector Construction
}
\section{Functions}{
\itemize{
\item \code{layered_zero_state}: Initialize a state (or other) vector with all zeros that is based on
a fully-crossed set of states in several sub-models.

\item \code{layered_const_state}: Initialize a state (or other) vector with a constant value that is
based on a fully-crossed set of states in several sub-models

output vectos

\item \code{const_named_vector}: Initialize a state (or other) vector with a constant value

\item \code{self_named_vector}: Construct a character vector with names given by its values
}}

