% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_and_strings.R
\name{\%_\%}
\alias{\%_\%}
\alias{\%~\%}
\alias{\%+\%}
\title{String Concatenation Operators}
\usage{
x \%_\% y

x \%~\% y

x \%+\% y
}
\arguments{
\item{x}{character vector}

\item{y}{character vector}
}
\description{
Paste with an underscore separator, tilde, or blank space,
except for length-zero character strings in the first vector
}
