% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{struc}
\alias{struc}
\alias{as.struc}
\title{Construct a struc Object}
\usage{
struc(...)

as.struc(x)
}
\arguments{
\item{...}{character vectors or matrices}

\item{x}{object to convert to a struc object}
}
\value{
struc object
}
\description{
Construct a struc Object
}
