% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{struc_bdiag}
\alias{struc_bdiag}
\title{Block Diagonal Matrix}
\usage{
struc_bdiag(x)
}
\arguments{
\item{x}{list of \code{struc} objects}
}
\value{
block diagonal struc object
}
\description{
Block Diagonal Matrix
}
