% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{struc_block}
\alias{struc_block}
\title{Repeat a Block}
\usage{
struc_block(x, row_times, col_times)
}
\arguments{
\item{x}{struc object representing the block}

\item{row_times}{number of times to replicate the block downwards}

\item{col_times}{number of times to replicate the block to the right}
}
\value{
struc object
}
\description{
Repeat a Block
}
