% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{struc_eval}
\alias{struc_eval}
\title{Numerically Evaluate Struc Object}
\usage{
struc_eval(x, values)
}
\arguments{
\item{x}{\code{\link{struc-class}} object}

\item{values}{object coercible to a named list}
}
\description{
Numerically Evaluate Struc Object
}
