% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\docType{class}
\name{struc_expanded-class}
\alias{struc_expanded-class}
\title{Class to Represent an Expanded Matrix Structure Object}
\description{
Class to Represent an Expanded Matrix Structure Object
}
\section{Slots}{

\describe{
\item{\code{l}}{list of struc objects}

\item{\code{dims}}{numeric vector giving the dimensions of the matrix
structure object}
}}

