% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{struc_expanded}
\alias{struc_expanded}
\title{Construct a struc_expanded Object}
\usage{
struc_expanded(l, d)
}
\arguments{
\item{l}{list of struc objects}

\item{d}{dimensions of the resulting object}
}
\value{
struc_expanded object
}
\description{
Construct a struc_expanded Object
}
