% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{struc_stretch}
\alias{struc_stretch}
\title{Stretch a Block}
\usage{
struc_stretch(x, row_each, col_each)
}
\arguments{
\item{x}{struc object representing the block}

\item{row_each}{number of times to replicate each element of the block downwards}

\item{col_each}{number of times to replicate each element of the block to the right}
}
\value{
struc object
}
\description{
Stretch a Block
}
