% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{subset}
\alias{subset}
\alias{[.struc}
\alias{[.EpiMatrix}
\alias{[<-.EpiMatrixInput}
\alias{[<-.EpiMatrixDerived}
\alias{[.EpiMatrixList}
\alias{diagonal,struc-method}
\title{Symbolic Matrix Subsetting}
\usage{
\method{`[`}{struc}(x, i, j, ..., drop = FALSE)

\method{[}{EpiMatrix}(x, i, j, ..., drop = FALSE)

\method{[}{EpiMatrixInput}(x, i, j, ...) <- value

\method{[}{EpiMatrixDerived}(x, i, j, ...) <- value

\method{[}{EpiMatrixList}(x, i, j, ..., drop = FALSE)

\S4method{diagonal}{struc}(x)
}
\description{
Symbolic Matrix Subsetting
}
\concept{symbolic}
