% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.params_pansim}
\alias{summary.params_pansim}
\title{summary method for \code{pansim} parameter objects}
\usage{
\method{summary}{params_pansim}(object, ...)
}
\arguments{
\item{object}{a params vector}

\item{...}{unused}
}
\description{
summary method for \code{pansim} parameter objects
}
