% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{sumprod}
\alias{sumprod}
\alias{sum,struc-method}
\alias{sum,struc_dimnames-method}
\alias{prod,struc-method}
\alias{prod,struc_dimnames-method}
\alias{inner,struc,struc-method}
\title{Summarise Symbolic Expressions as Scalars}
\usage{
\S4method{sum}{struc}(x, ..., na.rm = FALSE)

\S4method{sum}{struc_dimnames}(x, ..., na.rm = FALSE)

\S4method{prod}{struc}(x, ..., na.rm = FALSE)

\S4method{prod}{struc_dimnames}(x, ..., na.rm = FALSE)

\S4method{inner}{struc,struc}(x, y)
}
\description{
Summarise Symbolic Expressions as Scalars
}
\concept{symbolic}
