% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{symbtrans}
\alias{symbtrans}
\alias{t,struc_dimnames-method}
\title{Symbolic Matrix Transpose}
\usage{
\S4method{t}{struc}(x)

\S4method{t}{struc_dimnames}(x)
}
\arguments{
\item{x}{\code{\link{struc}} object}
}
\description{
Symbolic Matrix Transpose
}
\concept{symbolic}
