% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{texify}
\alias{texify}
\title{Convert expression string to TeX format}
\usage{
texify(x, dollars = TRUE, force = dev_is_tikz())
}
\arguments{
\item{x}{character string to be interpreted}

\item{dollars}{if \code{TRUE} then surround output string in \code{$} signs}

\item{force}{if \code{TRUE} then convert to TeX even if \code{\link{dev_is_tikz}} returns \code{FALSE}}
}
\description{
This has very limited capabilities and is intended
mainly to convert parameter names in mathematical expression strings
to associated TeX symbols so they look good when \code{\link[tikzDevice]{tikz}}
is in use.
}
\examples{
texify("R0 = beta/gamma")
texify("R0 = beta/gamma", dollars=FALSE, force=TRUE)

}
\seealso{
\code{\link[Hmisc]{latexTranslate}}
}
