% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tibble_row_to_named_vec}
\alias{tibble_row_to_named_vec}
\title{convert a tibble row (with all entries of the same data type) to a named vector (used in `condense_state()`)}
\usage{
tibble_row_to_named_vec(x)
}
\arguments{
\item{x}{a tibble with at most one row}
}
\value{
a named vector
}
\description{
convert a tibble row (with all entries of the same data type) to a named vector (used in `condense_state()`)
}
