% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{time_wrap}
\alias{time_wrap}
\title{Time Two Expressions}
\usage{
time_wrap(expr1, expr2, units = "secs")
}
\arguments{
\item{expr1}{first expression}

\item{expr2}{second expression}

\item{units}{time units with which to measure the system time
of each expression (see \code{\link{difftime}})}
}
\value{
list with three components: (1) time to evaluate
the first expression, (2) time to evaluate the second expression,
and (3) the time for the first expression divided by the time
for the second
}
\description{
Evaluate two expressions in the environment in which
\code{time_wrap} is called, and record the system
time associated with each expression.
}
\seealso{
Other compare: 
\code{\link{compare_grads}()},
\code{\link{compare_sims}()}
}
\concept{compare}
