% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{tmb_fun}
\alias{tmb_fun}
\title{Make Objective Function with TMB}
\usage{
tmb_fun(model, DLL = "McMasterPandemic")
}
\arguments{
\item{model}{object of class \code{flexmodel}}

\item{DLL}{argument to pass to the \code{DLL} argument of
\code{TMB::MakeADFun}. TODO: make this optional, and choose the main
package DLL once it exists
(https://github.com/mac-theobio/McMasterPandemic/issues/96)}
}
\description{
Construct an objective function in TMB from a \code{flexmodel}
object. The behaviour of \code{tmb_fun} depends on \code{spec_version()}
}
