% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{tmb_params}
\alias{tmb_params}
\title{Extract Parameter Vector to Pass to a TMB Function}
\usage{
tmb_params(model)
}
\arguments{
\item{model}{flexmodel}
}
\description{
Get all of the parameters from a flexmodel in a vector that is
ready to be passed to a TMB AD objective function, gradient function,
hessian function, simulate function, report function.
}
\details{
Currently this includes \code{params} and if \code{spec_ver_gt('0.1.0')}
it also includes time-varying multipliers in
\code{model$timevar$piece_wise$schedule$last_tv_mult}.
}
