% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{tmb_params_trans}
\alias{tmb_params_trans}
\title{Parameter Vectors Transformed for TMB}
\usage{
tmb_params_trans(model, vec_type = c("tmb_fun_arg", "params", "tv_mult"))
}
\arguments{
\item{model}{flexmodel}

\item{vec_type}{type of vector to return:
(1) tmb_fun_arg = same length as the argument for TMB objective function,
(2) params = same length as the params element in \code{model}
(3) tv_mult = same length as the \code{model$timevar$piece_wise$schedule$init_tv_mult}}
}
\description{
Parameter Vectors Transformed for TMB
}
