% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{togglePanelColourManager}
\alias{togglePanelColourManager}
\title{Colour checkboxes checkboxes.}
\usage{
togglePanelColourManager(curves, colourList)
}
\arguments{
\item{curves}{a list of input IDs of the checkboxes to modify the fill of.}

\item{colourList}{a list of hex obects specifiying the colour to change corresponding checkboxes in curves to.}
}
\value{
a tagsList containing the tags
}
\description{
Generate the necessary HTML/CSS tags to change the colour of a checkbox. Colour curve toggle buttons to match corresponding curves on the graph.
}
