% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_sort.R
\name{topological_sort}
\alias{topological_sort}
\title{Topological Sort}
\usage{
topological_sort(model)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}
}
\value{
character vector of state names in topological order
}
\description{
Topologically sort the state names so that states earlier in the flow
of individuals come earlier in sorted order.
}
