% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trans_state_vars}
\alias{trans_state_vars}
\title{Translate state variables}
\usage{
trans_state_vars(x)
}
\arguments{
\item{x}{a vector of state variables (factor or character)}
}
\description{
Translate state variables
}
