% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unpack}
\alias{unpack}
\title{unpack a list into the current (calling) environment
this is a replacement for \code{with()} (which is hard to debug) and \code{attach()} (which is frowned upon by CRAN/triggers package-check warnings)
(for more traditional Python tuple unpacking see the \code{zeallot} package)}
\usage{
unpack(x)
}
\arguments{
\item{x}{a named vector or a list}
}
\description{
unpack a list into the current (calling) environment
this is a replacement for \code{with()} (which is hard to debug) and \code{attach()} (which is frowned upon by CRAN/triggers package-check warnings)
(for more traditional Python tuple unpacking see the \code{zeallot} package)
}
