% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{update.params_pansim}
\alias{update.params_pansim}
\title{update parameters within a list of parameters}
\usage{
\method{update}{params_pansim}(object, ..., .list = FALSE)
}
\arguments{
\item{object}{a \code{params_pansim} object}

\item{...}{a list of named elements, or a single named list if \code{.list} is specified}

\item{.list}{treat the first argument as a named list?}
}
\description{
update parameters within a list of parameters
}
\examples{
pp <- list(params = c(a=2,b=1))
object <- c(a=1, b=0)
class(object) <- "params_pansim"
update(object,pp,.list=TRUE)
pp2 <- list(params = c(a=2))
update(object,pp, .list=TRUE)
update(object, a=2)
update(object, a=2, b=1)
}
