% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_condense_map}
\alias{update_condense_map}
\alias{add_condense_map}
\title{Specify and Name Variables for Condensation}
\usage{
update_condense_map(model, map = NULL)

add_condense_map(model, map_update)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{map}{named vector with names that are a subset of
the variables in the simulation model. if \code{NULL} the
identity map is used that makes all simulation history
variables available without name changes.
(\code{final_sim_report_names(model)}) and values that
give the names of the variables in the condensed data set}

\item{map_update}{map items to add to existing map}
}
\description{
The condensed simulation is a (possibly renamed) subset
of the variables in the simulation history. This function
allows one to define this subset and how it is renamed,
by creating a map with the following structure:
\code{c(orig_var_i = "cond_var_1", ..., orig_var_j = "cond_var_n")}.
}
\section{Functions}{
\itemize{
\item \code{add_condense_map()}: append new items to the condensation map

}}
