% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{update_contact_rate_setting}
\alias{update_contact_rate_setting}
\title{Helper function to update setting-specific contact rate list}
\usage{
update_contact_rate_setting(contact_rate_setting, values_to_update)
}
\arguments{
\item{contact_rate_setting}{named list containing setting-specific contact rates (in units of average contacts in the given setting per individual of age i with individuals of age j per day)}

\item{values_to_update}{named list of values to update in default list of setting-specific contact rates}
}
\value{
named list of setting-specific contact rates
}
\description{
Helper function to update setting-specific contact rate list
}
\examples{
update_contact_rate_setting(mk_contact_rate_setting(), list(school = 0))
}
