% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_error_dist}
\alias{update_error_dist}
\title{Update Error Distribution}
\usage{
update_error_dist(model, ...)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{...}{error distribution formulas}
}
\description{
Update Error Distribution
}
