% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_initial_state}
\alias{update_initial_state}
\title{Update Initial State}
\usage{
update_initial_state(model, silent = FALSE)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{silent}{warn if the model is not properly defined for
updating the initial state?}
}
\description{
Update the initial state of a \code{\link{flexmodel}} object
using an eigenvector-based approach to finding a near-disease-free
state
}
