% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_linearized_params}
\alias{update_linearized_params}
\title{Update Linearized Model Parameters}
\usage{
update_linearized_params(model, param_pattern, value)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{param_pattern}{regular expression identifying parameters
that require updating before they can be used in linearized
model simulations}

\item{value}{numeric value required to update}
}
\description{
Specify how to update model parameters for use with
linearized models during initial state vector construction.
}
\seealso{
Other flexmodel_definition_functions: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{initial_population}()},
\code{\link{mat_rate}()},
\code{\link{rep_rate}()},
\code{\link{update_disease_free_state}()},
\code{\link{vec_rate}()}
}
\concept{flexmodel_definition_functions}
