% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_loss_params}
\alias{update_loss_params}
\title{Update Observation Error}
\usage{
update_loss_params(model, loss_params, regenerate_rates = TRUE)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{loss_params}{TODO}

\item{regenerate_rates}{TODO}
}
\description{
Update Observation Error
}
