% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_observed}
\alias{update_observed}
\title{Update Observed Data}
\usage{
update_observed(model, data, loss_params = NULL, regenerate_rates = TRUE)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{data}{observed data frame in long format to
compare with simulated trajectories. must have the following
columns: \code{date}, \code{var}, \code{value}.}

\item{loss_params}{TODO}

\item{regenerate_rates}{should rates be regenerated to sort out any
index misalignment that may be introduced by editing the model?}
}
\value{
a \code{\link{flexmodel}} object with an \code{observed} element,
which is a list with two elements: (1) \code{data} (the attached data)
and (2) \code{loss} (a data frame describing the additional parameters
that are required to fit the model to data -- these parameters are added
to the \code{params} of the model)
}
\description{
Attach a data set to a \code{\link{flexmodel}}
object. Any existing data will be removed.
}
