% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_opt_params}
\alias{update_opt_params}
\title{Update Optimization Parameters}
\usage{
update_opt_params(model, ...)
}
\description{
\code{params ~ value}
\code{trans_param ~ value}
\code{trans_param ~ prior(hyperparameters ...)}
\code{trans_param ~ prior(hyperparameters ..., laplace = TRUE)}
}
