% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{update_params_mistry}
\alias{update_params_mistry}
\title{Update Mistry-based parameters}
\usage{
update_params_mistry(
  params,
  contact_rate_setting = NULL,
  transmissibility = NULL
)
}
\arguments{
\item{params}{parameter list initialized using `expand_params_mistry()`}

\item{contact_rate_setting}{named list containing setting-specific contact rates (in units of average contacts in the given setting per individual of age i with individuals of age j per day)}

\item{transmissibility}{probability of transmission upon contact with an infected (beta0 = transmissibility * contact_rate)}
}
\value{
an object of class `params_pansim`
}
\description{
Recalculate beta0 and/or pmat if setting-specific contact rates and/or
overall transmissiblity are changed
}
\details{
For `contact_rate_setting`, a full list (initialized with `mk_contact_rate_setting()`) or a partial list (e.g. `list(community = 0)`) can be provided. If a partial list is provided, default values from Mistry et al. 2021 are assumed for the contact rates that have not been specified (see `mk_contact_rate_setting()` for values).
}
\examples{
params <- read_params("PHAC_testify.csv")
age_cat <- mk_agecats(min = 0, max = 80, da = 10)
params_mistry <- expand_params_mistry(params = params, province = "ON", age_cat = age_cat)
params_mistry <- update_params_mistry(params_mistry,
 contact_rate_setting = list(school = 0))
}
\seealso{
Other classic_macpan: 
\code{\link{add_d_log}()},
\code{\link{add_updated_vaxrate}()},
\code{\link{aggregate_agecats}()},
\code{\link{calibrate_comb}()},
\code{\link{calibrate}()},
\code{\link{check_age_cat_compatibility}()},
\code{\link{check_contact_rate_setting}()},
\code{\link{col_multiply}()},
\code{\link{condense_age}()},
\code{\link{condense_params_vax}()},
\code{\link{condense_state}()},
\code{\link{condense_vax}()},
\code{\link{dev_is_tikz}()},
\code{\link{do_step}()},
\code{\link{expand_params_age}()},
\code{\link{expand_params_desc_age}()},
\code{\link{expand_params_desc_variant}()},
\code{\link{expand_params_desc_vax}()},
\code{\link{expand_params_mistry}()},
\code{\link{expand_params_variant}()},
\code{\link{expand_params_vax}()},
\code{\link{expand_state_age}()},
\code{\link{expand_state_vax}()},
\code{\link{expand_stateval_testing}()},
\code{\link{fix_pars}()},
\code{\link{fix_stored}()},
\code{\link{forecast_ensemble}()},
\code{\link{forecast_sim}()},
\code{\link{getData}()},
\code{\link{get_GI_moments}()},
\code{\link{get_Gbar}()},
\code{\link{get_R0}()},
\code{\link{get_doses_per_day}()},
\code{\link{get_evec}()},
\code{\link{get_kernel_moments}()},
\code{\link{get_opt_pars}()},
\code{\link{get_r}()},
\code{\link{invlink_trans}()},
\code{\link{make_betavec}()},
\code{\link{make_beta}()},
\code{\link{make_jac}()},
\code{\link{make_ratemat}()},
\code{\link{make_state}()},
\code{\link{make_test_wtsvec}()},
\code{\link{make_vaxrate}()},
\code{\link{mk_Nvec}()},
\code{\link{mk_agecats}()},
\code{\link{mk_contact_rate_setting}()},
\code{\link{mk_mistry_Nvec}()},
\code{\link{mk_pmat}()},
\code{\link{mk_vaxcats}()},
\code{\link{mle_fun}()},
\code{\link{non_expanded_states}},
\code{\link{rExp}()},
\code{\link{read_params}()},
\code{\link{repair_names_age}()},
\code{\link{restore}()},
\code{\link{run_sim_ageify}()},
\code{\link{run_sim_break}()},
\code{\link{run_sim_loglin}()},
\code{\link{run_sim_mobility}()},
\code{\link{run_sim_range}()},
\code{\link{run_sim}()},
\code{\link{show_ratemat}()},
\code{\link{testify}()},
\code{\link{texify}()},
\code{\link{trans_state_vars}()},
\code{\link{update_contact_rate_setting}()},
\code{\link{update_foi}()},
\code{\link{vis_model}()},
\code{\link{write_params}()}
}
\concept{classic_macpan}
