% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_ranef_params}
\alias{update_ranef_params}
\title{Update Random Effect Parameters}
\usage{
update_ranef_params(model, ...)
}
\arguments{
\item{model}{\code{flexmodel_to_calibrate} object}

\item{...}{formulas for specifying parameters as random effects -- see
\code{\link{update_opt_params}} for more detail}
}
\description{
Specify parameters as random effects
}
