% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_simulation_bounds}
\alias{update_simulation_bounds}
\title{Update Simulation Bounds}
\usage{
update_simulation_bounds(model, start_date = NULL, end_date = NULL)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{start_date}{optional new start date for simulations}

\item{end_date}{optional new end date for simulations}
}
\description{
Update Simulation Bounds
}
