% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_tmb_indices}
\alias{update_tmb_indices}
\alias{add_tmb_indices}
\title{Update TMB Indices}
\usage{
update_tmb_indices(model)

add_tmb_indices(model)
}
\arguments{
\item{model}{compartmental model}

\item{another}{compartmental model with indices for TMB}
}
\description{
Add or update indices used to access appropriate values
during simulation and calibration using TMB
}
\details{
\describe{
  \item{\code{make_ratemat_indices$from}}{}
  \item{\code{make_ratemat_indices$to}}{}
  \item{\code{make_ratemat_indices$count}}{}
  \item{\code{updateidx}}{indices into the \code{to}, \code{from}, and \code{count}
  vectors, identifying rates that need (or at least will) be updated at
  every simulation step}
}
}
\seealso{
Other flexmodels: 
\code{\link{add_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{init_model}()},
\code{\link{initial_population}()},
\code{\link{make_base_model}()},
\code{\link{make_omicron_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{state_param_sum}()},
\code{\link{tmb_fun}()},
\code{\link{vec_rate}()}
}
\concept{flexmodels}
