% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{v2d}
\alias{v2d}
\title{Vector to Data Frame}
\usage{
v2d(x, values_col = NULL, names_col = "names")
}
\arguments{
\item{x}{named vector}

\item{values_col}{name of the values column -- the default takes a guess}

\item{names_col}{name of the names column}
}
\description{
Convert a named vector to a data frame with two columns, one for the
names and one for the values.
}
