% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccify.R
\name{vaccify}
\alias{vaccify}
\title{expand rate matrix for testing status}
\usage{
vaccify(ratemat, params, debug = FALSE, testing_time = NULL)
}
\arguments{
\item{ratemat}{original rate matrix}

\item{params}{parameters}

\item{debug}{what it sounds like}

\item{testing_time}{"report" (N and P are counted at the time when individuals move from _n, _p to _u, _t compartments) or "sample" (N and P are counted when individuals move from _u to _n or _p)}
}
\description{
expand rate matrix for testing status
}
\examples{
params <- read_params("PHAC_testify.csv")
state <- make_state(params[["N"]],E0=params[["E0"]])
M <- make_ratemat(state,params)
image(M)
}
