% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{values}
\alias{values}
\alias{values,EpiMatrix-method}
\title{Extract Values}
\usage{
values(x)

\S4method{values}{EpiMatrix}(x)
}
\arguments{
\item{x}{an object from which to extract values}
}
\description{
Extract the values of an object with structure.
}
