% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{vec_factr}
\alias{vec_factr}
\title{Vectors of Intermediate Factors}
\usage{
vec_factr(model, factr_nms, formula)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{factr_nms}{vector of names of the new intermediate factors}

\item{formula}{\code{\link{struc}} object describing the
vector of intermediate factors}
}
\value{
\code{\link{flexmodel}} object
}
\description{
Vectors of Intermediate Factors
}
\seealso{
See \code{\link{add_factr}} to add a single
scalar-valued intermediate factor and \code{\link{vec}}
to create a vector-valued \code{\link{struc}} object.
}
