% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{vec_rate}
\alias{vec_rate}
\title{Specify Vector of Rates}
\usage{
vec_rate(model, from, to, formula, mapping = c("pairwise", "blockwise"))
}
\description{
Specify Vector of Rates
}
\seealso{
Other flexmodels: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{flexmodel}()},
\code{\link{initial_population}()},
\code{\link{make_base_model}()},
\code{\link{make_omicron_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{tmb_fun}()},
\code{\link{update_tmb_indices}()}
}
\concept{flexmodels}
