% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vis_model}
\alias{vis_model}
\title{display flow chart on current graphics device}
\usage{
vis_model(
  params = read_params("PHAC_testify.csv"),
  testify = FALSE,
  ageify = FALSE,
  method = c("Matrix", "diagram", "igraph"),
  ...
)
}
\arguments{
\item{params}{parameters (flowchart is determined by non-zero flows)}

\item{testify}{add testing flows?}

\item{ageify}{add age structure?}

\item{method}{visualization method}

\item{...}{parameters to pass to \code{\link{show_ratemat}}}
}
\description{
display flow chart on current graphics device
}
\examples{
vis_model()
vis_model(testify=TRUE)
vis_model(method="diagram")
}
