% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{write_params}
\alias{write_params}
\title{write parameters to CSV file}
\usage{
write_params(params, fn, label)
}
\arguments{
\item{params}{a params object}

\item{fn}{file name}

\item{label}{a label for the parameters}
}
\description{
write parameters to CSV file
}
