% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_lag_diff}
\alias{add_lag_diff}
\title{Add Variable by Lag Differencing}
\usage{
add_lag_diff(model, var_pattern, delay_n = 1)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{var_pattern}{regular expression used to identify variables for
differencing. See \code{\link{avail_for_lag}} for a function that will
return the names of all variables that are available for differencing.}

\item{delay_n}{Delay in days for determining the lag in the differences.}
}
\description{
Add a variable that will be available for comparisons with observed data
that is created by processing an existing variable in the simulation
history by lagged differencing.
}
