% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_linearized_outflow}
\alias{add_linearized_outflow}
\alias{add_outflow}
\title{Add Outflows}
\usage{
add_linearized_outflow(model, from, to)

add_outflow(model, from = ".+", to = ".+")
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{from}{string giving a regular expression for identifying
states from which individuals are flowing}

\item{to}{string giving a regular expression for identifying
states to which individuals are flowing}
}
\value{
updated \code{\link{flexmodel}} object
}
\description{
Add outflows corresponding to inflows specified by
\code{add_rate}, \code{rep_rate}, and \code{vec_rate}.
}
\details{
By default, this function will set outflows for all
corresponding inflows. To define outflows \code{from}
specific states \code{to} other specific states, pass
regular expressions to the \code{from} and \code{to}
arguments to identify these states.

\code{add_linearized_outflow} is used to specify
outflows in linearized models
(\url{https://canmod.net/misc/flex_specs#v0.1.1}).
}
\seealso{
Other flexmodel_definition_functions: 
\code{\link{add_factr}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{initial_population}()},
\code{\link{mat_rate}()},
\code{\link{rep_rate}()},
\code{\link{update_disease_free_state}()},
\code{\link{update_linearized_params}()},
\code{\link{vec_rate}()}
}
\concept{flexmodel_definition_functions}
