% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_parallel_accumulators}
\alias{add_parallel_accumulators}
\title{Add Parallel Accumulators}
\usage{
add_parallel_accumulators(model, state_patterns)
}
\arguments{
\item{model}{TODO}

\item{state_patterns}{regular expressions for identifying states as
parallel accumulators}
}
\value{
another compartmental model with parallel accumulators specified
}
\description{
Add parallel accumulators to a compartmental model.
}
