% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_parallel_accumulators}
\alias{add_parallel_accumulators}
\title{Add Parallel Accumulators}
\usage{
add_parallel_accumulators(model, state_patterns)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{state_patterns}{regular expressions for identifying states as
parallel accumulators}
}
\value{

}
\description{
[deprecated] Add parallel accumulators to a compartmental model.
}
\seealso{
Other flexmodel_definition_functions: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{initial_population}()},
\code{\link{mat_rate}()},
\code{\link{rep_rate}()},
\code{\link{update_disease_free_state}()},
\code{\link{update_linearized_params}()},
\code{\link{vec_rate}()}
}
\concept{flexmodel_definition_functions}
