% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_pow}
\alias{add_pow}
\title{Add Power Law}
\usage{
add_pow(model, pow_nms, pow_arg1_nms, pow_arg2_nms, pow_const_nms)
}
\description{
Compute and save the intermediate result of applying a
power law to the state variables, parameters, sums of these
things, and factr expressions of them as well.
}
