% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_rate}
\alias{add_rate}
\title{Add Rate}
\usage{
add_rate(model, from, to, formula)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{from}{Name of state from which flow is coming}

\item{to}{Name of state to which flow is going}

\item{formula}{Model formula defining dependence of the rate on
parameters and state variables}
}
\value{
another compartmental model with an additional non-zero rate matrix
element specified
}
\description{
Define how the rate of flow from one compartment to another
depends on parameters, state variables, .
}
\seealso{
Other flexmodel_definition_functions: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{initial_population}()},
\code{\link{mat_rate}()},
\code{\link{rep_rate}()},
\code{\link{update_disease_free_state}()},
\code{\link{update_linearized_params}()},
\code{\link{vec_rate}()}

Other rate_functions: 
\code{\link{mat_rate}()},
\code{\link{rep_rate}()},
\code{\link{vec_rate}()}
}
\concept{flexmodel_definition_functions}
\concept{rate_functions}
