% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_state_mappings}
\alias{add_state_mappings}
\title{Add State Mappings}
\usage{
add_state_mappings(
  model,
  eigen_drop_pattern,
  infected_drop_pattern,
  initial_susceptible_pattern
)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{eigen_drop_pattern}{regular expression for identifying states
to be dropped before computing the eigenvector of the Jacobian
of the linearized model}

\item{infected_drop_pattern}{regular expression for identifying states
to be dropped from the eigenvector so that only 'infected' states
remain}

\item{initial_susceptible_pattern}{regular expression for
identifying states associated with susceptible classes}
}
\value{
\code{\link{flexmodel}} object
}
\description{
Add regular expressions for identifying what states
are included in various stages of the eigenvector-based
method for constructing initial states
(\url{https://canmod.net/misc/flex_specs#v0.1.1})
}
\seealso{
Other flexmodels: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_param_sum}()},
\code{\link{flexmodel}()},
\code{\link{initial_population}()},
\code{\link{make_base_model}()},
\code{\link{make_omicron_model}()},
\code{\link{make_sir_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{tmb_fun}()},
\code{\link{update_tmb_indices}()},
\code{\link{vec_rate}()}
}
\concept{flexmodels}
