% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{add_state_param_sum}
\alias{add_state_param_sum}
\title{Sums of States and Parameters}
\usage{
add_state_param_sum(model, sum_name, summands)
}
\arguments{
\item{model}{\code{\link{flexmodel}}}

\item{sum_name}{name of sum of state variables and parameters}

\item{summands}{character vector of regular expressions for identifying
state variables and parameters to sum together}
}
\value{
\code{\link{flexmodel}}
}
\description{
Save intermediate sums of states and parameters so that
they can be used as factors in rate expressions and
retured as part of the simulation history.
}
\seealso{
Other flexmodels: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{flexmodel}()},
\code{\link{initial_population}()},
\code{\link{make_base_model}()},
\code{\link{make_omicron_model}()},
\code{\link{make_sir_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{tmb_fun}()},
\code{\link{update_tmb_indices}()},
\code{\link{vec_rate}()}
}
\concept{flexmodels}
