% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{aggregate.pansim}
\alias{aggregate.pansim}
\title{Temporal aggregation of pansim objects}
\usage{
\method{aggregate}{pansim}(
  x,
  start = NULL,
  period = NULL,
  FUN = mean,
  fixed_vars = TRUE,
  extend = 30,
  ...
)
}
\arguments{
\item{x}{a pansim object}

\item{start}{starting date for temporal aggregation}

\item{period}{time period for temporal aggregation (e.g. "7 days", see \code{\link{seq.Date}})}

\item{FUN}{temporal aggregation function (applied across all variables) \emph{or} a list of the form \code{list(FUN1=c('var1','var2'), FUN2=c('var3', 'var4'))} (temporal aggregation is done after state aggregation, so variable names specified should be adjusted appropriately)}

\item{fixed_vars}{treat names in FUN as fixed strings rather than regular expressions?}

\item{extend}{how far to extend time series for aggregation purposes}

\item{...}{unused, for generic consistency}
}
\description{
Temporal aggregation of pansim objects
}
\examples{
params <- read_params("ICU1.csv")
state <- make_state(params=params)
sdate <- "10-Feb-2020" ## arbitrary!
res <- run_sim(params,state,start_date=sdate,end_date="1-Jun-2020")
a1 <- aggregate(res, start="12-Feb-2020",period="7 days", FUN=sum)
## column-specific aggregation
first <- dplyr::first
agg_funs <- list(mean=c("H","ICU","I"), sum=c("report","death"))
a2 <- aggregate(condense(res), start="12-Feb-2020",period="7 days", FUN=agg_funs)
}
