% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{calibrate via negative binomial MLE, simultaneously fitting initial conditions, initial growth rate, time-changes in growth rate, and dispersion parameters}
\usage{
calibrate(
  start_date = min(data$date) - start_date_offset,
  start_date_offset = 15,
  end_date = max(data$date),
  break_dates = c("23-Mar-2020", "30-Mar-2020"),
  base_params,
  data,
  opt_pars = list(params = c(log_E0 = 4, log_beta0 = -1), log_rel_beta0 = c(-1, -1),
    log_nb_disp = 0),
  fixed_pars = NULL,
  sim_args = NULL,
  aggregate_args = NULL,
  debug = FALSE,
  debug_plot = FALSE,
  mle2_method = "Nelder-Mead",
  mle2_control = list(maxit = 10000),
  mle2_args = list()
)
}
\arguments{
\item{start_date}{starting date for sims (far enough back to allow states to sort themselves out)}

\item{start_date_offset}{days to go back before first data value}

\item{end_date}{ending date}

\item{break_dates}{specified breakpoints in beta0}

\item{base_params}{baseline parameters (an object (vector?) of type \code{params_pansim} containing all of the parameters needed for a simulation; some may be overwritten during the calibration process)}

\item{data}{a data set to compare to, containing date/var/value (current version assumes that only a single state var is included)}

\item{opt_pars}{starting parameters (and structure).  Parameters that are part of the \code{params_pansim} parameter vector can be specified within the \code{params} element (with prefixes if they are transformed); other parameters can include distributional parameters or time-varying parameters}

\item{fixed_pars}{parameters to fix}

\item{sim_args}{additional arguments to pass to \code{\link{run_sim}}}

\item{aggregate_args}{arguments passed to \code{\link{aggregate.pansim}}}

\item{debug}{print debugging messages?}

\item{debug_plot}{plot debugging curves?}

\item{mle2_method}{method arg for mle2}

\item{mle2_control}{control args for mle2}

\item{mle2_args}{additional arguments for mle2}
}
\description{
calibrate via negative binomial MLE, simultaneously fitting initial conditions, initial growth rate, time-changes in growth rate, and dispersion parameters
}
