% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_comb}
\alias{calibrate_comb}
\title{Combined calibration of model to multiple types of data}
\usage{
calibrate_comb(
  data,
  params,
  opt_pars = NULL,
  mob_data = NULL,
  mob_breaks = NULL,
  mob_breaks_int = FALSE,
  mob_logist_scale = NA,
  spline_days = 14,
  spline_setback = 0,
  spline_extrap = c("linear", "constant"),
  spline_df = NA,
  knot_quantile_var = NA,
  spline_pen = 0,
  spline_type = "bs",
  spline_int = FALSE,
  testing_data = NULL,
  maxit = 10000,
  skip.hessian = FALSE,
  use_DEoptim = TRUE,
  DE_cores = 1,
  beta_break_dates = NULL,
  use_mobility = FALSE,
  use_phenomhet = FALSE,
  use_spline = FALSE,
  use_testing = FALSE,
  vars = NULL,
  debug_plot = interactive(),
  debug = FALSE,
  debug_hist = FALSE,
  return_val = c("fit", "X", "formula", "args", "time_args"),
  start_date = NULL,
  priors = list(),
  ...
)
}
\arguments{
\item{data}{a data set to compare to, containing date/var/value
(current version assumes that only a single state var is
included)}

\item{params}{parameters}

\item{opt_pars}{starting parameters (and structure).  Parameters
that are part of the \code{params_pansim} parameter vector can
be specified within the \code{params} element (with prefixes
if they are transformed); other parameters can include
distributional parameters or time-varying parameters}

\item{mob_data}{mobility data}

\item{mob_breaks}{vector of breakpoints for piecewise mobility
model}

\item{mob_breaks_int}{(logical) specifies whether the intercept of
the mobility/transmission relationship changes at each
mobility breakpoint. The default (\code{FALSE}) specifies that
mobility in 'mobility period' \code{i} is
\code{beta0*(rel_mobility)^p_i}; \code{TRUE} would specify
\code{beta0*(a_i)*(rel_mobility)^p_i}, with \code{a_1==1}.}

\item{mob_logist_scale}{specifies the scale of the smooth
(logistic) transition between mobility periods, in days: if it
is \code{NA} (default), the model uses sharp breaks for a
piecewise-constant model.  Otherwise, the parameter used makes
a logistic transition between the two periods with the
specified scale.}

\item{spline_days}{days between spline knots}

\item{spline_setback}{days before end of time series to set
boundary knots for spline (this implies \emph{linear}
extrapolation after knots if \code{spline_type="ns"} is
specified, which is probably wise)}

\item{spline_extrap}{spline extrapolation model ("linear" or
"constant")}

\item{spline_df}{overall spline degrees of freedom}

\item{knot_quantile_var}{variable to use cum dist for knotspacing}

\item{spline_pen}{penalization for spline}

\item{spline_type}{spline type ("ns" for natural spline or "bs"
for b-spline)}

\item{spline_int}{spline intercept (??)}

\item{testing_data}{data frame with columns containing dates
(\code{Date}) and testing intensity (\code{intensity}) (=
tests per capita per day)}

\item{maxit}{maximum iterations for Nelder-Mead/optimization step}

\item{skip.hessian}{skip Hessian calculation?}

\item{use_DEoptim}{use differential evolution as first stage?}

\item{DE_cores}{number of parallel workers for DE}

\item{beta_break_dates}{discrete break dates for transmission changes}

\item{use_mobility}{include mobility as a covariate in the model?}

\item{use_phenomhet}{include phenomenological heterogeneity?}

\item{use_spline}{include spline?}

\item{use_testing}{include variation in testing intensity?}

\item{vars}{which vars to use? (default is all in data)}

\item{debug_plot}{plot debugging curves? (doesn't work with
parallel DEoptim)}

\item{debug}{print debugging messages?}

\item{debug_hist}{keep information on parameter history?}

\item{return_val}{"fit" (return calibrated value); "X"
(short-circuit/return model matrix?); "formula" (return
log-linear formula for time-varying beta)}

\item{start_date}{start date}

\item{priors}{a list of tilde-delimited expressions giving prior
distributions expressed in terms of the elements of
\code{opt_pars},
e.g. \code{list(~dlnorm(rel_beta0[1],meanlog=-1,sd=0.5))}}

\item{...}{extra args}
}
\description{
Top-level calibration based on mobility, splines, and
phenomenological heterogeneity.  This function is a wrapper for
\code{\link{calibrate}}, which first sets up the model matrix for
time-dependent transmission rates \eqn{\beta(t)}.  For example,
for splines, it sets up the spline basis and the coefficients of
each basis function.  For mobility, it sets up the slope and
intercept parameters (since mobility is treated as a log-linear
function between break points).
}
\examples{
if (require(dplyr)) {
  dd <- ont_all \%>\% trans_state_vars() \%>\%
       filter(var \%in\% c("H","report"))
 params <- read_params("ICU1.csv")
 ## quick and dirty example (maximize speed)
\dontrun{
 calibrate_comb(data=dd, params=params,
               use_spline=TRUE,
              maxit=10, skip.hessian=TRUE, use_DEoptim =FALSE)
X <-  calibrate_comb(data=dd, params=params,
              use_spline=TRUE,
              spline_type="ns",
              spline_setback=1,
              spline_extrap="constant",
              return_val="X")
matplot(X, ylab="")
form <-  calibrate_comb(data=dd, params=params,
              use_spline=TRUE,
              spline_type="ns",
              spline_setback=1,
              spline_extrap="constant",
              return_val="formula")
print(form)
}
}
}
\seealso{
Other classic_macpan: 
\code{\link{add_d_log}()},
\code{\link{add_updated_vaxrate}()},
\code{\link{aggregate_agecats}()},
\code{\link{calibrate}()},
\code{\link{check_age_cat_compatibility}()},
\code{\link{check_contact_rate_setting}()},
\code{\link{col_multiply}()},
\code{\link{condense_age}()},
\code{\link{condense_params_vax}()},
\code{\link{condense_state}()},
\code{\link{condense_vax}()},
\code{\link{dev_is_tikz}()},
\code{\link{do_step}()},
\code{\link{expand_params_age}()},
\code{\link{expand_params_desc_age}()},
\code{\link{expand_params_desc_variant}()},
\code{\link{expand_params_desc_vax}()},
\code{\link{expand_params_mistry}()},
\code{\link{expand_params_variant}()},
\code{\link{expand_params_vax}()},
\code{\link{expand_state_age}()},
\code{\link{expand_state_vax}()},
\code{\link{expand_stateval_testing}()},
\code{\link{fix_pars}()},
\code{\link{fix_stored}()},
\code{\link{forecast_ensemble}()},
\code{\link{forecast_sim}()},
\code{\link{getData}()},
\code{\link{get_GI_moments}()},
\code{\link{get_Gbar}()},
\code{\link{get_R0}()},
\code{\link{get_doses_per_day}()},
\code{\link{get_evec}()},
\code{\link{get_kernel_moments}()},
\code{\link{get_opt_pars}()},
\code{\link{get_r}()},
\code{\link{invlink_trans}()},
\code{\link{make_betavec}()},
\code{\link{make_beta}()},
\code{\link{make_jac}()},
\code{\link{make_ratemat}()},
\code{\link{make_state}()},
\code{\link{make_test_wtsvec}()},
\code{\link{make_vaxrate}()},
\code{\link{mk_Nvec}()},
\code{\link{mk_agecats}()},
\code{\link{mk_contact_rate_setting}()},
\code{\link{mk_mistry_Nvec}()},
\code{\link{mk_pmat}()},
\code{\link{mk_vaxcats}()},
\code{\link{mle_fun}()},
\code{\link{non_expanded_states}},
\code{\link{rExp}()},
\code{\link{read_params}()},
\code{\link{repair_names_age}()},
\code{\link{restore}()},
\code{\link{run_sim_ageify}()},
\code{\link{run_sim_break}()},
\code{\link{run_sim_loglin}()},
\code{\link{run_sim_mobility}()},
\code{\link{run_sim_range}()},
\code{\link{run_sim}()},
\code{\link{show_ratemat}()},
\code{\link{testify}()},
\code{\link{texify}()},
\code{\link{trans_state_vars}()},
\code{\link{update_contact_rate_setting}()},
\code{\link{update_foi}()},
\code{\link{update_params_mistry}()},
\code{\link{vis_model}()},
\code{\link{write_params}()}
}
\concept{classic_macpan}
