% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{check_contact_rate_setting}
\alias{check_contact_rate_setting}
\title{Helper function to check setting-specific contact rate list initialization}
\usage{
check_contact_rate_setting(contact_rate_setting)
}
\description{
Helper function to check setting-specific contact rate list initialization
}
\examples{
check_contact_rate_setting(mk_contact_rate_setting)
}
