% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{col_multiply}
\alias{col_multiply}
\title{Compute elementwise multiplication of a vector by each column in a matrix
This is implemented for performance reasons, as it is much faster than sweep.}
\usage{
col_multiply(M, v)
}
\arguments{
\item{M}{matrix}

\item{v}{vector}
}
\description{
Compute elementwise multiplication of a vector by each column in a matrix
This is implemented for performance reasons, as it is much faster than sweep.
}
