% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{compare_sims}
\alias{compare_sims}
\title{Compare TMB-based and classic MacPan Simulations}
\usage{
compare_sims(
  classic_sim,
  tmb_sim,
  tolerance = testthat_tolerance(),
  compare_attr = TRUE
)
}
\arguments{
\item{classic_sim}{result of `run_sim` without using TMB}

\item{tmb_sim}{result of `run_sim` using TMB}

\item{tolerance}{numerical tolerance}

\item{compare_attr}{compare attributes or just the simulations themselves}
}
\description{
Exceptions to drop-in replacement:
1. don't require that the attributes are in the same order
2. don't require that the r version returns everything that the tmb
   version does (e.g. flexmodel)
3. don't require that the row.names are identical (is this ok?
   the r version counts iterations with skips, but is this informative?)
4. don't require that the call is identical (obvious i guess, but
   being exhaustive)
5. don't look for parameter called S0 in classic simulation results
   (FIXME: this one might not be necessary anymore)
6. don't require the ordering of params_timevar rows to be the same
   (FIXME: this should probably be handled in run_sim itself)
}
