% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{complement}
\alias{complement}
\alias{inverse}
\alias{complement,character-method}
\alias{inverse,character-method}
\alias{complement,struc-method}
\alias{inverse,struc-method}
\title{Symbolic Complement and Inverse}
\usage{
complement(x)

inverse(x)

\S4method{complement}{character}(x)

\S4method{inverse}{character}(x)

\S4method{complement}{struc}(x)

\S4method{inverse}{struc}(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Symbolic Complement and Inverse
}
\examples{
complement('p')
inverse('N')
}
