% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{condense.pansim}
\alias{condense.pansim}
\title{Condense columns (infected, ICU, hospitalized) in a pansim output}
\usage{
\method{condense}{pansim}(object, add_reports = TRUE, diff_deaths = TRUE, keep_all = FALSE, ...)
}
\arguments{
\item{object}{a pansim object}

\item{add_reports}{add incidence and case reports?}

\item{keep_all}{keep unaggregated variables in data frame as well?}

\item{...}{additional args}
}
\description{
Condense columns (infected, ICU, hospitalized) in a pansim output
}
