% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{condense_state}
\alias{condense_state}
\title{collapse (non-accumulator) states into subcategories (ages, vax status)}
\usage{
condense_state(x, return_type = c("tibble", "named_vector", "unnamed_vector"))
}
\arguments{
\item{x}{a state vector or data frame (each row is a different time point)}

\item{values_only}{just return values (unlisted and unnamed?)}
}
\value{
a tibble of counts aggregated across epidemiological states
}
\description{
collapse (non-accumulator) states into subcategories (ages, vax status)
}
