% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{derive}
\alias{derive}
\title{#' @export
epi_model_init = structure(
  list(
    const_list = epi_const_list(),
    mat_list = epi_mat_list(),
    param_mat_nms = character(0L),
    state_mat_nms = character(0L),
    rate_list = epi_rate_list()
  ),
  class = "EpiModelInit"
)
Derive}
\usage{
derive(mat_list, ...)
}
\description{
Derive symbolic expressions of existing epidemiological variables, and
add them to the list
}
