% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{describe_params}
\alias{describe_params}
\title{Describe parameters}
\usage{
describe_params(x)
}
\arguments{
\item{x}{a \code{params_pansim} object}
}
\description{
Create a data frame with symbols, values and meanings of parameters
}
