% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{do_step}
\alias{do_step}
\title{Take a single simulation time step}
\usage{
do_step(
  state,
  params,
  transmat,
  dt = 1,
  do_hazard = FALSE,
  stoch = c(obs = FALSE, proc = FALSE)
)
}
\arguments{
\item{state}{named vector of states}

\item{params}{named vector of parameters}

\item{transmat}{transition matrix}

\item{dt}{time step (days)}

\item{do_hazard}{use hazard calculation?}

\item{stoch}{stochastic simulation? logical vector for observation and process noise}
}
\description{
Take a single simulation time step
}
