% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{do_step}
\alias{do_step}
\title{Take a single simulation time step}
\usage{
do_step(
  state,
  params,
  ratemat,
  dt = 1,
  do_hazard = FALSE,
  stoch_proc = FALSE,
  do_exponential = FALSE
)
}
\arguments{
\item{state}{named vector of states}

\item{params}{named vector of parameters}

\item{ratemat}{transition matrix}

\item{dt}{time step (days)}

\item{do_hazard}{use hazard calculation?}

\item{stoch_proc}{stochastic process error?}

\item{do_exponential}{prevent outflow of susceptibles, to create a pure-exponential process?}
}
\description{
Take a single simulation time step
}
\examples{
params1 <- read_params("ICU1.csv")
state1 <- make_state(params=params1)
M <- make_ratemat(params=params1, state=state1)
s1A <- do_step(state1,params1, M, stoch_proc=TRUE)
}
