% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{expand_params_age}
\alias{expand_params_age}
\title{expand parameter list to include age structure}
\usage{
expand_params_age(
  pp,
  age_cat = mk_agecats(),
  Cmat = mk_Cmat(),
  Nvec = mk_Nvec(),
  beta0vec = NULL
)
}
\arguments{
\item{pp}{parameter list (e.g. read in with `read_params()`)}

\item{age_cat}{vector of age categories}

\item{Cmat}{contact matrix; default is uniform}

\item{Nvec}{population distribution (as counts); default is uniform}

\item{beta0vec}{vector of beta0 values; default is all the same value}
}
\description{
expand parameter list to include age structure
}
\examples{
pp <- read_params("PHAC_testify.csv")
ppa <- expand_params_age(pp)
}
