% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{expand_params_desc_vax}
\alias{expand_params_desc_vax}
\title{Edit parameter description attribute to include vax-specific definitions
(helper function for `expand_params_vax()`)}
\usage{
expand_params_desc_vax(
  params_desc,
  model_type = c("onedose", "twodose", "twodosewane")
)
}
\arguments{
\item{params_desc}{parameter descriptions, as initialized in `read_params()`}

\item{model_type}{choose the one-dose, the two-dose model, or the two-dose model with waning}
}
\description{
Edit parameter description attribute to include vax-specific definitions
(helper function for `expand_params_vax()`)
}
