% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageify.R
\name{expand_params_mistry}
\alias{expand_params_mistry}
\title{Make contact matrix using Mistry et al. approach}
\usage{
expand_params_mistry(
  params,
  transmissibility = 1,
  province = "Ontario",
  weights = list(household = 4.11, school = 11.41, work = 8.07, community = 2.79),
  age_cat = NULL
)
}
\arguments{
\item{province}{province for which to construct the matrix (if NULL, make a Canada-wide contact matrix)}

\item{weights}{named list containing setting-specific weights (in units of average contacts in the given setting per individual of age i with individuals of age j per day)}

\item{age_cat}{(optional) list of age groups to aggregate ages in; use mk_agecats() to generate (must start with 0 and end with 84); default is single ages starting with 0 and up to 83, then a single 84+ category}
}
\value{
matrix of average contacts between individuals of ages i and j per individual of age i per unit time (day)
}
\description{
Make contact matrix using Mistry et al. approach
}
\examples{
params <- read_params("PHAC_testify.csv")
age_cat <- mk_agecats(min = 0, max = 80, da = 10)
expand_params_mistry(params = params, province = "Ontario", age_cat = age_cat)
}
