% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{expand_params_vax}
\alias{expand_params_vax}
\title{Expand parameter list to include vaccination}
\usage{
expand_params_vax(
  params,
  model_type = c("onedose", "twodose"),
  vax_doses_per_day = 1e+05,
  vax_prop_first_dose = 1,
  vax_efficacy_dose1 = 0.6,
  vax_efficacy_dose2 = 0.9,
  vax_avg_response_time = 14,
  vax_avg_response_time_R = 7,
  vax_alpha_dose1 = 0.5,
  vax_alpha_dose2 = 0.5,
  vax_mu_dose1 = 1,
  vax_mu_dose2 = 1
)
}
\arguments{
\item{params}{parameter list (e.g. read in with `read_params()`)}

\item{model_type}{choose either the one-dose or the two-dose model}

\item{vax_doses_per_day}{total number of vaccine doses administered per day}

\item{vax_prop_first_dose}{the proportion of vaccine doses administered per day that are first doses}

\item{vax_efficacy_dose1}{infection-blocking vaccine efficacy for dose 1}

\item{vax_efficacy_dose2}{infection-blocking vaccine efficacy for dose 2}

\item{vax_avg_response_time}{average number of days it takes for a vaccine dose to confer protection for an individual that has never been infected}

\item{vax_avg_response_time_R}{average number of days it takes for a vaccine dose to confer protection for an individual that has previously been infected (e.g. between being dosed and eliciting the protective immune response)}

\item{vax_alpha_dose1}{proportion of infections that are asymptomatic in individuals that have received one dose of the vaccine}

\item{vax_alpha_dose2}{proportion of infections that are asymptomatic in individuals that have received two doses of the vaccine}

\item{vax_mu_dose1}{proportion of infections that are mild (vs. severe) in individuals that have received one dose of the vaccine}

\item{vax_mu_dose2}{proportion of infections that are mild (vs. severe) in individuals that have received two doses of the vaccine}
}
\description{
Expand parameter list to include vaccination
}
\examples{
params <- read_params("PHAC.csv")
params_vax <- expand_params_vax(params)
}
\seealso{
Other classic_macpan: 
\code{\link{add_d_log}()},
\code{\link{add_updated_vaxrate}()},
\code{\link{aggregate_agecats}()},
\code{\link{calibrate_comb}()},
\code{\link{calibrate}()},
\code{\link{check_age_cat_compatibility}()},
\code{\link{check_contact_rate_setting}()},
\code{\link{col_multiply}()},
\code{\link{condense_age}()},
\code{\link{condense_params_vax}()},
\code{\link{condense_state}()},
\code{\link{condense_vax}()},
\code{\link{dev_is_tikz}()},
\code{\link{do_step}()},
\code{\link{expand_params_age}()},
\code{\link{expand_params_desc_age}()},
\code{\link{expand_params_desc_variant}()},
\code{\link{expand_params_desc_vax}()},
\code{\link{expand_params_mistry}()},
\code{\link{expand_params_variant}()},
\code{\link{expand_state_age}()},
\code{\link{expand_state_vax}()},
\code{\link{expand_stateval_testing}()},
\code{\link{fix_pars}()},
\code{\link{fix_stored}()},
\code{\link{forecast_ensemble}()},
\code{\link{forecast_sim}()},
\code{\link{getData}()},
\code{\link{get_GI_moments}()},
\code{\link{get_Gbar}()},
\code{\link{get_R0}()},
\code{\link{get_doses_per_day}()},
\code{\link{get_evec}()},
\code{\link{get_kernel_moments}()},
\code{\link{get_opt_pars}()},
\code{\link{get_r}()},
\code{\link{invlink_trans}()},
\code{\link{make_betavec}()},
\code{\link{make_beta}()},
\code{\link{make_jac}()},
\code{\link{make_ratemat}()},
\code{\link{make_state}()},
\code{\link{make_test_wtsvec}()},
\code{\link{make_vaxrate}()},
\code{\link{mk_Nvec}()},
\code{\link{mk_agecats}()},
\code{\link{mk_contact_rate_setting}()},
\code{\link{mk_mistry_Nvec}()},
\code{\link{mk_pmat}()},
\code{\link{mk_vaxcats}()},
\code{\link{mle_fun}()},
\code{\link{non_expanded_states}},
\code{\link{rExp}()},
\code{\link{read_params}()},
\code{\link{repair_names_age}()},
\code{\link{restore}()},
\code{\link{run_sim_ageify}()},
\code{\link{run_sim_break}()},
\code{\link{run_sim_loglin}()},
\code{\link{run_sim_mobility}()},
\code{\link{run_sim_range}()},
\code{\link{run_sim}()},
\code{\link{show_ratemat}()},
\code{\link{testify}()},
\code{\link{texify}()},
\code{\link{trans_state_vars}()},
\code{\link{update_contact_rate_setting}()},
\code{\link{update_foi}()},
\code{\link{update_params_mistry}()},
\code{\link{vis_model}()},
\code{\link{write_params}()}
}
\concept{classic_macpan}
