% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{expand_state_vax}
\alias{expand_state_vax}
\title{expand state vector by vaccination status}
\usage{
expand_state_vax(
  x,
  model_type = c("onedose", "twodose", "twodosewane"),
  unif = FALSE
)
}
\arguments{
\item{x}{state vector}

\item{model_type}{choose the one-dose, the two-dose model, or the two-dose model with waning}

\item{unif}{should individuals in each epidemic category be distributed evenly among vaccination strata? (if FALSE, put everyone in the first vaccination category, i.e., the unvaccinated category)}
}
\description{
by default, everyone starts with an unvaccinated status
}
\examples{
params <- read_params("PHAC.csv")
ss <- make_state(params=params)
ss2 <- expand_state_vax(ss)
}
