% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testify.R
\name{expand_stateval}
\alias{expand_stateval}
\title{expand states and values to include}
\usage{
expand_stateval(
  x,
  method = c("eigvec", "untested", "spread"),
  params = NULL,
  add_accum = TRUE
)
}
\arguments{
\item{x}{state vector}

\item{method}{method for distributing values across new (expanded) states}

\item{params}{parameters}

\item{add_accum}{add N and P (neg/pos test) accumulator categories?}
}
\description{
expand states and values to include
}
\examples{
pp <- read_params("PHAC_testify.csv")
s <- make_state(params=pp)
expand_stateval(s, params=pp)
expand_stateval(s, method="untested")
}
