% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{fix_pars}
\alias{fix_pars}
\title{adjust pars to match targets}
\usage{
fix_pars(
  params,
  target = c(r = 0.23, Gbar = 6),
  pars_adj = list("beta0", c("sigma", "gamma_s", "gamma_m", "gamma_a")),
  r_method = c("expsim", "rmult"),
  u_interval = c(-3, 3)
)
}
\arguments{
\item{params}{a parameter vector}

\item{target}{target values for one or more epidemic moments}

\item{pars_adj}{list of sets of parameters to adjust}

\item{r_method}{method for fixing r (brute-force exponential simulation or JD's kernel-based approach?)}

\item{u_interval}{interval for uniroot adjustment}
}
\description{
adjust pars to match targets
}
