% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{fix_pars}
\alias{fix_pars}
\title{adjust pars to match targets}
\usage{
fix_pars(
  params,
  target = c(r = 0.23, Gbar = 6),
  pars_adj = list("beta0", c("gamma", "lambda_s", "lambda_m", "lambda_a"))
)
}
\description{
adjust pars to match targets
}
