% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{forecast_ensemble}
\alias{forecast_ensemble}
\title{find confidence envelopes by simulation}
\usage{
forecast_ensemble(
  fit,
  nsim = 200,
  forecast_args = fit$forecast_args,
  qvec = c(0.05, 0.5, 0.95),
  qnames = c("lwr", "value", "upr"),
  seed = NULL,
  .progress = if (interactive()) "text" else "none"
)
}
\arguments{
\item{fit}{output from \code{calibrate}}

\item{nsim}{number of simulations}

\item{forecast_args}{arguments to pass to \code{forecast_sim}}

\item{qvec}{vector of quantiles}

\item{qnames}{quantile names}

\item{seed}{random-number seed}

\item{.progress}{progress bar?}
}
\description{
find confidence envelopes by simulation
}
