% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{forecast_ensemble}
\alias{forecast_ensemble}
\title{find confidence envelopes by simulation}
\usage{
forecast_ensemble(
  fit,
  nsim = 200,
  forecast_args = fit$forecast_args,
  qvec = c(0.05, 0.5, 0.95),
  qnames = c("lwr", "value", "upr"),
  seed = NULL,
  imp_wts = FALSE,
  Sigma = bbmle::vcov(fit$mle2),
  scale_Sigma = 1,
  calc_Rt = FALSE,
  fix_pars_re = "nb_disp",
  raw_ensembles = FALSE,
  parallel = FALSE,
  n_cores = 4,
  .progress = if (interactive()) "text" else "none"
)
}
\arguments{
\item{fit}{output from \code{calibrate}}

\item{nsim}{number of simulations}

\item{forecast_args}{arguments to pass to \code{forecast_sim}}

\item{qvec}{vector of quantiles: NULL to return an array (nt x nvars x nsims) instead of a tibble with date/var+ quantiles}

\item{qnames}{quantile names}

\item{seed}{random-number seed}

\item{imp_wts}{use importance weighting, i.e. weight ensemble based on log-likelihood?}

\item{Sigma}{covariance matrix to pass to \code{pop_pred_samp}}

\item{scale_Sigma}{multiplier for covariance matrix}

\item{calc_Rt}{calculate and include R(t) in prediction/forecast?}

\item{fix_pars_re}{a regular expression specifying the names of parameters that should be treated as fixed when constructing the parameter ensemble}

\item{raw_ensembles}{(logical) return ensembles (FIXME: should implement return_type=c("array","quantiles","raw") (but not really raw, "dataframe"??))}

\item{parallel}{whether to attempt parallel processing}

\item{n_cores}{number of cores for parallel forecasting (ignored if not parallel) (??)}

\item{.progress}{progress bar?}
}
\description{
find confidence envelopes by simulation
}
\seealso{
Other classic_macpan: 
\code{\link{add_d_log}()},
\code{\link{add_updated_vaxrate}()},
\code{\link{aggregate_agecats}()},
\code{\link{calibrate_comb}()},
\code{\link{calibrate}()},
\code{\link{check_age_cat_compatibility}()},
\code{\link{check_contact_rate_setting}()},
\code{\link{col_multiply}()},
\code{\link{condense_age}()},
\code{\link{condense_params_vax}()},
\code{\link{condense_state}()},
\code{\link{condense_vax}()},
\code{\link{dev_is_tikz}()},
\code{\link{do_step}()},
\code{\link{expand_params_age}()},
\code{\link{expand_params_desc_age}()},
\code{\link{expand_params_desc_variant}()},
\code{\link{expand_params_desc_vax}()},
\code{\link{expand_params_mistry}()},
\code{\link{expand_params_variant}()},
\code{\link{expand_params_vax}()},
\code{\link{expand_state_age}()},
\code{\link{expand_state_vax}()},
\code{\link{expand_stateval_testing}()},
\code{\link{fix_pars}()},
\code{\link{fix_stored}()},
\code{\link{forecast_sim}()},
\code{\link{getData}()},
\code{\link{get_GI_moments}()},
\code{\link{get_Gbar}()},
\code{\link{get_R0}()},
\code{\link{get_doses_per_day}()},
\code{\link{get_evec}()},
\code{\link{get_kernel_moments}()},
\code{\link{get_opt_pars}()},
\code{\link{get_r}()},
\code{\link{invlink_trans}()},
\code{\link{make_betavec}()},
\code{\link{make_beta}()},
\code{\link{make_jac}()},
\code{\link{make_ratemat}()},
\code{\link{make_state}()},
\code{\link{make_test_wtsvec}()},
\code{\link{make_vaxrate}()},
\code{\link{mk_Nvec}()},
\code{\link{mk_agecats}()},
\code{\link{mk_contact_rate_setting}()},
\code{\link{mk_mistry_Nvec}()},
\code{\link{mk_pmat}()},
\code{\link{mk_vaxcats}()},
\code{\link{mle_fun}()},
\code{\link{non_expanded_states}},
\code{\link{rExp}()},
\code{\link{read_params}()},
\code{\link{repair_names_age}()},
\code{\link{restore}()},
\code{\link{run_sim_ageify}()},
\code{\link{run_sim_break}()},
\code{\link{run_sim_loglin}()},
\code{\link{run_sim_mobility}()},
\code{\link{run_sim_range}()},
\code{\link{run_sim}()},
\code{\link{show_ratemat}()},
\code{\link{testify}()},
\code{\link{texify}()},
\code{\link{trans_state_vars}()},
\code{\link{update_contact_rate_setting}()},
\code{\link{update_foi}()},
\code{\link{update_params_mistry}()},
\code{\link{vis_model}()},
\code{\link{write_params}()}
}
\concept{classic_macpan}
