% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{forecast_sim}
\alias{forecast_sim}
\title{simulate/forecast a single trajectory}
\usage{
forecast_sim(
  p,
  opt_pars,
  base_params,
  start_date,
  end_date,
  time_args = NULL,
  fixed_pars = NULL,
  stoch = NULL,
  stoch_start = NULL,
  sim_args = list(),
  aggregate_args = NULL,
  condense_args = NULL,
  return_val = c("aggsim", "vals_only"),
  sim_fun = run_sim_break,
  calc_Rt = FALSE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{p}{vector of parameters - on the link (log/logit) scale as appropriate; these are the parameters that will be adjusted during calibration}

\item{opt_pars}{starting parameters (and structure).  Parameters
that are part of the \code{params_pansim} parameter vector can
be specified within the \code{params} element (with prefixes
if they are transformed); other parameters can include
distributional parameters or time-varying parameters}

\item{base_params}{baseline parameters (an object (vector?) of
type \code{params_pansim} containing all of the parameters
needed for a simulation; some may be overwritten during the
calibration process)}

\item{start_date}{starting date for sims (far enough back to allow
states to sort themselves out)}

\item{end_date}{ending date}

\item{time_args}{arguments passed to \code{sim_fun}}

\item{fixed_pars}{parameters to fix}

\item{stoch}{stochastic settings (see \code{\link{run_sim}})}

\item{stoch_start}{dates on which to enable stochasticity (vector of dates with names 'proc' and 'obs')}

\item{sim_args}{additional arguments to pass to
\code{\link{run_sim}}}

\item{aggregate_args}{arguments passed to
\code{\link{aggregate.pansim}}}

\item{condense_args}{arguments to pass to \code{\link{condense}} (via \code{\link{run_sim}})}

\item{return_val}{specify values to return (aggregated simulation, or just the values?)}

\item{sim_fun}{function for simulating a single run
(e.g. \code{\link{run_sim_break}},
\code{\link{run_sim_mobility}})}

\item{calc_Rt}{calculate and include R(t) in prediction/forecast?}

\item{debug}{print debugging messages?}

\item{...}{extra args (ignored)}
}
\description{
simulate based on a vector of parameters (including both time-varying change parameters, initial conditions, and other dynamical parameters), for fitting or forecasting
}
\examples{
ff <- ont_cal1$forecast_args
op <- ff$opt_pars
p <- unlist(op)
params <- fix_pars(read_params("ICU1.csv"))
forecast_sim(p, op, base_params=params,ff$start_date, ff$end_date,
    time_args=ff$time_args)
}
