% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmodel_parameters.R
\name{get_and_set_model_parameters}
\alias{get_and_set_model_parameters}
\alias{pars_base_sim}
\alias{pars_base_init}
\alias{pars_base_opt}
\alias{pars_base_final}
\alias{pars_time_sim}
\alias{pars_time_spec}
\alias{pars_time_opt}
\alias{pars_time_series}
\alias{pars_time_hist}
\alias{pars_time_norm}
\alias{pars_infer_init}
\alias{pars_infer_opt}
\alias{pars_base_sim<-}
\title{Get and Set Model Parameters}
\usage{
pars_base_sim(model)

pars_base_init(model)

pars_base_opt(model)

pars_base_final(model)

pars_time_sim(model)

pars_time_spec(model)

pars_time_opt(model)

pars_time_series(model)

pars_time_hist(model)

pars_time_norm(model)

pars_infer_init(model)

pars_infer_opt(model)

pars_base_sim(model) <- value
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{value}{named vector of parameters}
}
\value{
parameter vector or data frame of time-variation of parameters
}
\description{
Get and Set Model Parameters
}
\section{Functions}{
\itemize{
\item \code{pars_base_sim()}: base parameters that would be used in a simulation.

\item \code{pars_base_init()}: base parameters that would be used as an initial value for an optimizer.

\item \code{pars_base_opt()}: optimized base parameters if they exist.

\item \code{pars_base_final()}: base simulation parameters with time-varying parameters at the end of the simulation

\item \code{pars_time_sim()}: time-varying parameter schedule that would be used in a simulation

\item \code{pars_time_spec()}: time-varying parameter schedule that was entered by the user at a model definition step.

\item \code{pars_time_opt()}: time-varying parameter schedule with optimized values that were entered as NA by the user in the model definition.

\item \code{pars_time_series()}: time-varying parameter schedule with resolved multiplication strategy so that all values are on the same scale as the associated base parameter.

\item \code{pars_time_hist()}: time-series for each time-varying parameter over the entire simulation range with resolved multiplication strategy.

\item \code{pars_time_norm()}: normalized time-series for each time-varying parameter over the entire simulation range with resolved multiplication strategy, with normalization obtained by dividing all values by the initial value.

\item \code{pars_infer_init()}: the initial value of the parameter vector involved in inference (e.g. passed to an objective function during calibration).

\item \code{pars_infer_opt()}: the optimized value of the parameter vector involved in inference.

}}
\note{
\code{pars_infer_*} are on the transformed scale that was chosen by the
user (e.g. log, logit).
}
