% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_old.R
\name{get_break}
\alias{get_break}
\title{estimate magnitude of drop in beta0 at known dates}
\usage{
get_break(
  start_date = anydate("1-Mar-2020"),
  end_date = anydate("8-Apr-2020"),
  break_dates = c("20-Mar-2020"),
  params,
  data,
  var = "H",
  sim_args = NULL,
  optim_args = list()
)
}
\arguments{
\item{start_date}{starting date}

\item{end_date}{ending date of sim}

\item{break_dates}{specified breakpoints in beta0}

\item{params}{parameters}

\item{data}{data to match}

\item{var}{variable to match}

\item{sim_args}{additional arguments to pass to \code{\link{run_sim}}}

\item{optim_args}{additional args to \code{optim}}
}
\description{
estimate magnitude of drop in beta0 at known dates
}
