% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{init_model}
\alias{init_model}
\title{Initialize Compartmental Model}
\usage{
init_model(
  params,
  state,
  start_date = NULL,
  end_date = NULL,
  timevar_piece_wise = NULL
)
}
\arguments{
\item{params}{a \code{param_pansim} object}

\item{state}{a \code{state_pansim} object}
}
\value{
object representing a compartmental model
}
\description{
TODO: `recompute` argument for functions below
TODO: `overwrite` argument for add functions that force `recompute`
TODO: create versions of `rate/add_rate` functions that allow structure.
      maybe `rates/add_rates`??
}
