% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invlink_trans}
\alias{invlink_trans}
\title{apply inverse-link functions to parameter vector or list based on names}
\usage{
invlink_trans(p, unknown_link = "ignore", links = c("log", "log10", "logit"))
}
\arguments{
\item{p}{a named vector or list of parameter values with names of the form \code{linkfun_name}}

\item{unknown_link}{action to take when unknown link is specified (ignored)}

\item{links}{allowed link prefixes}
}
\description{
apply inverse-link functions to parameter vector or list based on names
}
\examples{
invlink_trans(c(log_p1=0,logit_p2=0))
invlink_trans(list(log_p1=c(0,0),logit_p2=c(0,0,0)))
invlink_trans(list(p1=c(log_a=0,log_b=0),p2=4))
tst <- list(params = c(log_beta0 = 0.693147180559945), log_nb_disp = 4.60517018598809)
invlink_trans(tst)
tst2 <- list(params=c(log_E0=1,log_beta0=1),log_nb_disp=c(H=0,report=0,death=0))
invlink_trans(tst2)
invlink_trans(list(time_beta=1,log_time_beta=0))  ## ignore non-link prefixes
invlink_trans(list(time_beta=c(2,3),log_time_beta=c(a=0,b=0)))  ##
}
