% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{make_base_model}
\alias{make_base_model}
\title{Represent a Standard Unstructured Model as a flexmodel}
\usage{
make_base_model(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=flexmodel]{flexmodel}}
  \describe{
    \item{\code{params}}{a \code{param_pansim} object}
    \item{\code{state}}{a \code{state_pansim} object}
    \item{\code{start_date}}{simulation start date}
    \item{\code{end_date}}{simulation end date}
    \item{\code{params_timevar}}{data frame with scheduling for piece-wise
constant parameter variation (TODO: direct to other help pages)}
    \item{\code{do_hazard}}{should hazard simulation steps be used?
(https://canmod.net/misc/flex_specs#v0.0.5) -- only used
if \code{spec_ver_gt('0.0.4')}}
    \item{\code{do_hazard_lin}}{like \code{do_hazard} but for the
linearized model that is used to construct the initial state
variable -- only used when \code{do_make_state == TRUE}}
    \item{\code{do_approx_hazard}}{like \code{do_approx_hazard} but for
the linearized model that is used to construct the initial
state (experimental)}
    \item{\code{do_make_state}}{should state be remade on the c++ size?
(https://canmod.net/misc/flex_specs#v0.1.1) -- only used
if \code{spec_ver_gt('0.1.0')}}
    \item{\code{max_iters_eig_pow_meth}}{maximum number of iterations
to use in computing the eigenvector for initial state
construction}
    \item{\code{tol_eig_pow_meth}}{tolerance for determining convergence
of the power method used in initial state construction}
  }}
}
\description{
Represent a Standard Unstructured Model as a flexmodel
}
\seealso{
Other flexmodels: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{flexmodel}()},
\code{\link{initial_population}()},
\code{\link{make_omicron_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{tmb_fun}()},
\code{\link{update_tmb_indices}()},
\code{\link{vec_rate}()}

Other canned_models: 
\code{\link{make_omicron_model}()},
\code{\link{make_vaccination_model}()}
}
\concept{canned_models}
\concept{flexmodels}
