% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{make_jac}
\alias{make_jac}
\title{construct Jacobian matrix for ICU model
(not quite complete: doesn't include flows to R)}
\usage{
make_jac(params, state = NULL)
}
\arguments{
\item{params}{parameter vector}

\item{state}{state vector (named)}
}
\description{
construct Jacobian matrix for ICU model
(not quite complete: doesn't include flows to R)
}
\examples{
params <- read_params("ICU1.csv")
state <- make_state(params[["N"]],E0=params[["E0"]], use_eigvec=FALSE)
## state[c("E","Ia","Ip","Im","Is")] <- 1
state[["E"]] <- 1
J <- make_jac(params,state)
J["S","S"]
Jr <- J[1:6,1:6]
round(Jr,3)
eigen(Jr)$values
make_jac(params)
}
