% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{make_nested_indices}
\alias{make_nested_indices}
\title{Nested Indices}
\usage{
make_nested_indices(x, patterns, invert = FALSE)
}
\arguments{
\item{x}{character vector}

\item{patterns}{named list or vector of regular expressions to be applied
sequentially to create a nested set of character vectors}
}
\value{
List with three elements.
\describe{
  \item{\code{x}}{Nested character vectors.}
  \item{\code{i}}{
    List of lists, with inner list giving the indices into character vectors
    that are higher in the hierarchy. For example, one may read this
    expression, \code{x$e == x$p\\[i$e$p\\]}, as "e equals p at the index
    that takes p to e".
  }
  \item{\code{j}}{Set difference version of \code{i}.}
}
}
\description{
Create and return a nested set of character vectors from a sequence
of regular expressions, and return indices into each vector for recovering
other shorter vectors that are lower in the hierarchy.
}
