% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_IP.R
\name{make_omicron_model}
\alias{make_omicron_model}
\title{Make an Omicron invasion model}
\usage{
make_omicron_model(..., do_variant = FALSE)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=init_model]{init_model}}
  \describe{
    \item{\code{}}{}
  }}

\item{do_variant}{allow for different variants (TODO: more info obviously required here)}
}
\description{
Make an Omicron invasion model
}
\seealso{
Other flexmodels: 
\code{\link{flexmodel}()},
\code{\link{make_ageified_model}()},
\code{\link{make_base_model}()},
\code{\link{make_hello_world_model}()},
\code{\link{make_sir_model}()},
\code{\link{make_testified_model}()},
\code{\link{make_vaccination_model}()}

Other canned_models: 
\code{\link{make_ageified_model}()},
\code{\link{make_base_model}()},
\code{\link{make_hello_world_model}()},
\code{\link{make_sir_model}()},
\code{\link{make_testified_model}()},
\code{\link{make_vaccination_model}()}
}
\concept{canned_models}
\concept{flexmodels}
