% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{make_ratemat}
\alias{make_ratemat}
\title{Create transition matrix: defines rates (per day) of flow
\emph{from} compartment i (row) \emph{to} compartment j (column)
base version matches structure of Stanford/Georgia models
\itemize{
\item flow diagram: see https://covid-measures.github.io/ 'model details' tab or ../pix/model_schematic.png
\item parameter definitions: see params_CI_base.csv, params_ICU_diffs.csv
}}
\usage{
make_ratemat(state, params, do_ICU = TRUE)
}
\arguments{
\item{state}{named vector of states}

\item{params}{named vector of parameters}

\item{do_ICU}{include additional health utilization compartments}
}
\value{
matrix of (daily) transition rates
}
\description{
Create transition matrix: defines rates (per day) of flow
\emph{from} compartment i (row) \emph{to} compartment j (column)
base version matches structure of Stanford/Georgia models
\itemize{
\item flow diagram: see https://covid-measures.github.io/ 'model details' tab or ../pix/model_schematic.png
\item parameter definitions: see params_CI_base.csv, params_ICU_diffs.csv
}
}
\examples{
params <- read_params(system.file("params","ICU1.csv",package="McMasterPandemic"))
state <- make_state(params[["N"]],E=params[["E0"]])
if (require(Matrix)) {
  image(Matrix(make_ratemat(state,params)))
}
}
