% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{make_ratemat}
\alias{make_ratemat}
\title{Create transition matrix}
\usage{
make_ratemat(state, params, do_ICU = TRUE, sparse = FALSE, symbols = FALSE)
}
\arguments{
\item{state}{named vector of states}

\item{params}{named vector of parameters}

\item{do_ICU}{include additional health utilization compartments}

\item{sparse}{return sparse matrix?}

\item{symbols}{return character (symbol) form? (FIXME: call adjust_symbols here rather than in show_ratemat()?)}
}
\value{
matrix of (daily) transition rates
}
\description{
Defines rates (per day) of flow \emph{from} compartment \code{i}
(row) \emph{to} compartment \code{j} (column).
}
\details{
The rates are as follows:

\eqn{ S to E:  - (\beta_0 / N) S (C_a I_a + C_p I_p + (1-iso_m)C_m I_m + (1-iso_s)C_s I_s) }
\eqn{ E to I_a: }
\eqn{ E to I_p: }
\eqn{ ... }

See \code{\link{read_params}} for parameter definitions.
}
\note{
Base version matches structure of Stanford/Georgia models
\itemize{
  \item flow diagram: see \url{http://covid-measures.stanford.edu/} 'model details' tab
        or \code{../pix/model_schematic.png}
  \item parameter definitions: see \code{params_CI_base.csv}, \code{params_ICU_diffs.csv}
}
}
\examples{
params <- read_params("ICU1.csv")
state <- make_state(params[["N"]],E0=params[["E0"]])
M <- make_ratemat(state,params)
if (require(Matrix)) {
   image(Matrix(M))
}
make_ratemat(state,params,symbols=TRUE)
}
