% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{make_state}
\alias{make_state}
\title{retrieve parameters from a CSV file}
\usage{
make_state(
  N = params[["N"]],
  E0 = params[["E0"]],
  type = "ICU1h",
  state_names = NULL,
  use_eigvec = NULL,
  params = NULL,
  x = NULL,
  normalize = FALSE,
  ageify = NULL,
  vaxify = NULL,
  testify = NULL
)
}
\arguments{
\item{N}{population size}

\item{E0}{initial number exposed}

\item{type}{(character) specify what model type this is intended
for (e.g., \code{"ICU1"}, \code{"CI"}); determines state names}

\item{state_names}{vector of state names, must include S and E}

\item{use_eigvec}{use dominant eigenvector to distribute non-Susc values
to zero: default is to set this to \code{TRUE} if \code{params} is non-NULL}

\item{params}{parameter vector (looked in for N and E0)}

\item{x}{proposed (named) state vector; missing values will be set}

\item{normalize}{(logical) should the state vector be normalized to sum to 1?}

\item{ageify}{expand state vector to include different age groups (??)}

\item{vaxify}{expand state vector to include groups that have 1-2 vaccine doses (??)}

\item{testify}{expand state vector to include testing compartments (untested, neg waiting, pos waiting, pos received) ?}
}
\description{
retrieve parameters from a CSV file
}
\details{
The parameters that must be set are:

\eqn{   N:  }  population size

\eqn{   \beta_0:  }  transmission rate

\eqn{   1/\sigma:  }  mean \emph{latent} period

\eqn{   1/\gamma_a:  }  mean \emph{infectious} period for asymptomatic individuals

\eqn{   ... }

generate initial state vector
}
\note{
\code{"CI"} refers to the Stanford group's
    "covid intervention" model.
}
\examples{
p <- read_params("ICU1.csv")
make_state(N=1e6,E0=1)
make_state(params=p)
}
