% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testify.R
\name{make_test_wtsvec}
\alias{make_test_wtsvec}
\alias{make_test_posvec}
\title{make weights vector for tests}
\usage{
make_test_wtsvec(params, var_names = NULL)

make_test_posvec(params)
}
\arguments{
\item{params}{parameter vector}

\item{var_names}{variables names, \emph{in matching order to state vector/rate matrix}}
}
\description{
make weights vector for tests

make vector of test positivity
}
\examples{
pp <- read_params("PHAC_testify.csv")
state1 <- state0 <- make_state(params=pp)   ## unexpanded
state1[] <- 1  ## occupy all states
state <- expand_stateval(state0)
wtsvec <- make_test_wtsvec(pp, var_names=setdiff(names(state0),non_expanded_states))
posvec <- make_test_posvec(pp)
## need to make_ratemat() with *unexpanded* state, then
##  expand it
ratemat <- testify(make_ratemat(state1,pp), pp)
betavec <- make_betavec(state,pp)
tt2 <- ratemat
tt2[tt2>0] <- 1 ## make all edges == 1
show_ratemat(tt2)
show_ratemat(ratemat,aspect="fill")
heatmap(ratemat, Rowv=NA, Colv=NA, scale="none")
if (require(igraph)) {
   g <- igraph::graph_from_adjacency_matrix(tt2)
   plot(g, layout=igraph::layout_nicely)
}
}
