% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testify.R
\name{make_test_wtsvec}
\alias{make_test_wtsvec}
\title{make weights vector for tests}
\usage{
make_test_wtsvec(params)
}
\arguments{
\item{params}{parameter vector}
}
\description{
make weights vector for tests
}
\examples{
pp <- read_params("PHAC_testify.csv")
state0 <- make_state(params=pp)   ## unexpanded
state <- expand_stateval(state0)
wtsvec <- make_test_wtsvec(pp)
posvec <- make_test_posvec(pp)
## need to make_ratemate() with *unexpanded* state, then
##  expand it
ratemat <- testify(make_ratemat(state0,pp), pp)
betavec <- make_betavec(state,pp)
tt2 <- ratemat
tt2[tt2>0] <- 1 ## make all edges == 1 
if (require(igraph)) {
   g <- igraph::graph_from_adjacency_matrix(tt2)
   plot(g, layout=igraph::layout_nicely)
}
}
