% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{mat_rate}
\alias{mat_rate}
\title{Specify Matrix of Rates}
\usage{
mat_rate()
}
\description{
Not implemented
}
\seealso{
Other flexmodels: 
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_tmb_indices}()},
\code{\link{init_model}()},
\code{\link{initial_population}()},
\code{\link{make_base_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{state_param_sum}()},
\code{\link{tmb_fun}()},
\code{\link{vec_rate}()}
}
\concept{flexmodels}
