% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{McMasterPandemic-package}
\alias{McMasterPandemic}
\alias{McMasterPandemic-package}
\title{\pkg{McMasterPandemic}}
\description{
This R package provides Compartmental epidemic models for
forecasting and analysis of infectious disease pandemics.
}
\details{
See the vignette \code{McMasterPandemic.pdf} for now...
Eventually \code{vignette("McMasterPandemic",package="McMasterPandemic")}
should work...

\deqn{ dS/dt = - (\beta_0 / N) S (C_a I_a + C_p I_p + (1-iso_m)C_m I_m + (1-iso_s)C_s I_s) }
\deqn{ dE/dt =  }
\deqn{ ... }
\deqn{ dR/dt =  }

Use \code{browseVignettes("McMasterPandemic")} to find vignettes
for this package.
}
\author{
\strong{Maintainer}: Ben Bolker \email{bolker@mcmaster.ca}

Authors:
\itemize{
  \item David Earn
  \item Morgan Kain
  \item Jonathan Dushoff
}

}
