% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandify.R
\name{mk_vaxcats}
\alias{mk_vaxcats}
\title{Construct vector of vaccine category labels}
\usage{
mk_vaxcats(use_doses = FALSE)
}
\arguments{
\item{use_doses}{(logical) should we make separate categories for two doses?}
}
\description{
Construct vector of vaccine category labels
}
