% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{mle_fun}
\alias{mle_fun}
\title{negative log-likelihood function}
\usage{
mle_fun(
  p,
  data,
  debug = FALSE,
  debug_plot = FALSE,
  opt_pars,
  base_params,
  start_date,
  end_date,
  time_args = NULL,
  sim_args = NULL,
  sim_fun = run_sim_break,
  checkpoint = FALSE,
  aggregate_args = NULL,
  priors = NULL,
  ...
)
}
\arguments{
\item{p}{parameter vector (in unlisted form)}

\item{data}{a data set to compare to, containing date/var/value (current version assumes that only a single state var is included)}

\item{debug}{print debugging messages?}

\item{debug_plot}{plot debugging curves? (doesn't work with parallel DEoptim)}

\item{opt_pars}{starting parameters (and structure).  Parameters that are part of the \code{params_pansim} parameter vector can be specified within the \code{params} element (with prefixes if they are transformed); other parameters can include distributional parameters or time-varying parameters}

\item{base_params}{baseline parameters (an object (vector?) of type \code{params_pansim} containing all of the parameters needed for a simulation; some may be overwritten during the calibration process)}

\item{start_date}{starting date for sims (far enough back to allow states to sort themselves out)}

\item{end_date}{ending date}

\item{time_args}{arguments passed to \code{sim_fun}}

\item{sim_args}{additional arguments to pass to \code{\link{run_sim}}}

\item{sim_fun}{function for simulating a single run (e.g. \code{\link{run_sim_break}}, \code{\link{run_sim_mobility}})}

\item{aggregate_args}{arguments passed to \code{\link{aggregate.pansim}}}

\item{priors}{a list of tilde-delimited expressions giving prior distributions expressed in terms of the elements of \code{opt_pars}, e.g. \code{list(~dlnorm(rel_beta0[1],meanlog=-1,sd=0.5))}}

\item{...}{unused (but useful in case junk needs to be discarded)}
}
\description{
negative log-likelihood function
}
