% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ont_all_sub}
\alias{ont_all_sub}
\alias{ont_cal1}
\title{Sample data and fit from Ontario (up to 18 April 2020)}
\format{
a data frame ...

a calibrated object ...
Sample COVID19 calibration
}
\source{
mike's repo
}
\usage{
ont_all_sub

ont_cal1
}
\description{
Sample data and fit from Ontario (up to 18 April 2020)
}
\keyword{datasets}
