% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ont_all}
\alias{ont_all}
\alias{comb_sub}
\title{Sample data and fit from Ontario (up to 18 April 2020)}
\format{
a data frame ...

a data frame containing date rel_activity
Averaged mobility data (Apple: Toronto driving, Google: Ontario retail_and_recreation, workplaces; Google data up to 22 May)
}
\source{
mike's repo
}
\usage{
ont_all

comb_sub
}
\description{
Sample data and fit from Ontario (up to 18 April 2020)
}
\keyword{datasets}
