% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ont_cal1}
\alias{ont_cal1}
\alias{ont_cal_2brks}
\alias{ont_cal_mob1}
\alias{bd
break dates}
\alias{bd}
\title{Fitted models to Ontario data}
\format{
a calibrated object ...
Sample COVID19 calibration

a calibrated object ...
Sample COVID19 calibration - only two breakpoints

a calibrated object ...
Calibration to mobility data

An object of class \code{character} of length 3.
}
\usage{
ont_cal1

ont_cal_2brks

ont_cal_mob1

bd
}
\description{
Fitted models to Ontario data
}
\keyword{datasets}
