% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{optim_flexmodel}
\alias{optim_flexmodel}
\alias{nlminb_flexmodel}
\alias{bbmle_flexmodel}
\title{Wrapped Optimizers for Flexmodels}
\usage{
optim_flexmodel(model, ...)

nlminb_flexmodel(model, ...)

bbmle_flexmodel(model, ...)
}
\arguments{
\item{model}{a \code{\link{flexmodel}} object}

\item{update_default_params}{should the default parameters be updated
to their calibrated values?}
}
\description{
Currently there are \code{optim_flexmodel}, \code{nlminb_flexmodel},
and \code{bbmle_flexmodel}
}
