% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{rateform_as_char}
\alias{rateform_as_char}
\alias{parse_formula}
\title{Parse a Flexmodel Formula}
\usage{
rateform_as_char(x)

parse_formula(x)
}
\arguments{
\item{x}{one-sided formula, character vector, or struc object describing
the dependence of rate matrix elements on parameter and/or state variables}
}
\value{
depends on the spec version (TODO: add detail once we converge)
}
\description{
Parse a Flexmodel Formula
}
