% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{plot.pansim}
\alias{plot.pansim}
\title{plot method for simulations}
\usage{
\method{plot}{pansim}(
  x,
  drop_states = c("S", "R", "E", "I"),
  keep_states = NULL,
  aggregate = TRUE,
  log = FALSE,
  ...
)
}
\arguments{
\item{x}{fitted \code{pansim} object}

\item{drop_states}{states to \emph{exclude} from plot}

\item{keep_states}{states to \emph{include} in plot (overrides \code{drop_states})}

\item{aggregate}{collapse states (e.g. all ICU states -> "ICU") before plotting?  See \code{\link{aggregate.pansim}}}

\item{log}{plot y-axis on log scale?}
}
\description{
plot method for simulations
}
