% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_res_by_age}
\alias{plot_res_by_age}
\title{Plot age-structured simulation result faceted by age categories}
\usage{
plot_res_by_age(res, drop_states = NULL, condense_I = FALSE, show_times = TRUE)
}
\arguments{
\item{res}{age-structured simulation result}

\item{drop_states}{epidemiological states to \emph{exclude} from plot}

\item{condense_I}{condense all the different I (infectious) states to a single I state? (??)}

\item{show_times}{indicate times when parameters changed?}
}
\description{
Plot age-structured simulation result faceted by age categories
}
