% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.fit_pansim}
\alias{predict.fit_pansim}
\title{make forecasts from sim}
\usage{
\method{predict}{fit_pansim}(
  object,
  end_date = NULL,
  stoch = NULL,
  keep_vars = c("H", "ICU", "d", "incidence", "report", "newTests/1000"),
  ensemble = FALSE,
  ...
)
}
\arguments{
\item{object}{a fitted object}

\item{end_date}{ending date for sim}

\item{stoch}{stochasticity}

\item{keep_vars}{...}

\item{ensemble}{run ensemble?}

\item{...}{extra args (passed to forecast_ensemble)}
}
\description{
make forecasts from sim
}
\examples{
predict(ont_cal1)
predict(ont_cal_2brks,ensemble=TRUE)
}
