% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.fit_pansim}
\alias{predict.fit_pansim}
\alias{plot.fit_pansim}
\title{make forecasts from sim}
\usage{
\method{predict}{fit_pansim}(
  object,
  end_date = NULL,
  stoch = NULL,
  stoch_start = NULL,
  keep_vars = c("H", "ICU", "death", "incidence", "report", "cumRep", "newTests/1000"),
  ensemble = FALSE,
  new_params = NULL,
  ...
)

\method{plot}{fit_pansim}(
  x,
  predict_args = NULL,
  data = NULL,
  break_dates = NULL,
  dlspace = 1,
  limspace = 10,
  add_tests = FALSE,
  add_ICU_cap = FALSE,
  mult_var = NULL,
  directlabels = TRUE,
  log = TRUE,
  ...
)
}
\arguments{
\item{object}{a fitted object}

\item{end_date}{ending date for sim}

\item{stoch}{stochasticity}

\item{stoch_start}{stoch starting date}

\item{keep_vars}{...}

\item{ensemble}{run ensemble?}

\item{new_params}{parameters to update in base parameters (e.g. adding stochastic parameters)}

\item{...}{extra args (passed to forecast_ensemble)}

\item{x}{a calibrated object (result from \code{\link{calibrate}}) or a prediction (from \code{\link{predict.fit_pansim}})}

\item{predict_args}{additional arguments to pass to predict}

\item{data}{original time series data}

\item{break_dates}{breakpoints}

\item{dlspace}{spacing for direct labels (not working)}

\item{limspace}{extra space (in days) to add to make room for direct labels}

\item{add_tests}{plot newTests/1000?}

\item{add_ICU_cap}{include horizontal lines showing ICU capacity?}

\item{mult_var}{variable in data set indicating multiple forecast types to compare}

\item{directlabels}{use direct labels?}

\item{log}{use a log10 scale for the y axis?}
}
\description{
make forecasts from sim
}
\examples{
##' predict(ont_cal1)
\dontrun{
## non-pos-def vcov ... ???
predict(ont_cal_2brks,ensemble=TRUE)
}
}
