% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.params_pansim}
\alias{print.params_pansim}
\title{print parameters, possibly with detailed description}
\usage{
\method{print}{params_pansim}(x, describe = FALSE, silent = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{params_pansim} (parameters for pandemic simulation)}

\item{describe}{print full description?}

\item{silent}{if \code{describe} and \code{silent}, do not print (just return data frame)}

\item{...}{(unused, for generic consistency)}
}
\description{
If the detailed description is requested, it is returned as a data frame.
}
