% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{rate}
\alias{rate}
\title{Define Rate for Single Element of Rate Matrix}
\usage{
rate(from, to, formula, state, params, sums, factrs, ratemat)
}
\arguments{
\item{from}{from state}

\item{to}{to state}

\item{formula}{one-sided formula defining the rate with reference
to the parameters and state variables}

\item{state}{state_pansim object}

\item{params}{param_pansim object}

\item{sums}{vector of sums of state variables and parameters}

\item{ratemat}{rate matrix}
}
\description{
Define Rate for Single Element of Rate Matrix
}
\seealso{
Other flexmodel_definition_functions: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{flexmodel}()},
\code{\link{initial_population}()},
\code{\link{mat_rate}()},
\code{\link{rep_rate}()},
\code{\link{update_disease_free_state}()},
\code{\link{update_linearized_params}()},
\code{\link{vec_rate}()}
}
\concept{flexmodel_definition_functions}
