% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{rate_summary}
\alias{rate_summary}
\title{Rate Summary}
\usage{
rate_summary(
  model,
  include_parse_info = TRUE,
  include_formula = FALSE,
  include_latex = FALSE
)
}
\arguments{
\item{model}{a \code{\link{flexmodel}} object}

\item{include_parse_info}{include information of numbers of variables,
products, and factors associated with each rate}

\item{include_formula}{include a column for the expanded rate formula}

\item{include_latex}{include a column with latex code for the rate formula
(experimental)}
}
\description{
Summarize the properties of the rates of transition amongst the
compartments in the model.
}
