% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{read_params}
\alias{read_params}
\title{retrieve parameters from a CSV file}
\usage{
read_params(fn, value_col = "Value", symbol_col = "Symbol")
}
\arguments{
\item{fn}{file name (CSV file containing at least value and symbol columns}

\item{value_col}{name of column containing values}

\item{symbol_col}{name of column containing symbols}
}
\description{
retrieve parameters from a CSV file
}
\details{
The parameters that must be set are:

\eqn{   N:  }  population size

\eqn{   \beta_0:  }  transmission rate

\eqn{   1/\sigma:  }  mean \emph{latent} period

\eqn{   1/\gamma_a:  }  mean \emph{infectious} period for asymptomatic individuals

\eqn{   ... }
}
