% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{rep_rate}
\alias{rep_rate}
\title{Repeat a Rate for Several Rate Matrix Elements}
\usage{
rep_rate(model, from, to, formula, mapping = c("pairwise", "blockwise"))
}
\arguments{
\item{model}{flexmodel}

\item{formula}{formula or length-1 character vector}

\item{indices}{two-column matrix of indices with column names "from_pos"
and "to_pos", locaing elements of the rate matrix in model}
}
\description{
Repeat a Rate for Several Rate Matrix Elements
}
\seealso{
Other flexmodels: 
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_tmb_indices}()},
\code{\link{init_model}()},
\code{\link{initial_population}()},
\code{\link{make_base_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{state_param_sum}()},
\code{\link{tmb_fun}()},
\code{\link{vec_rate}()}
}
\concept{flexmodels}
