% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{rep_rate}
\alias{rep_rate}
\title{Repeat a Rate for Several Rate Matrix Elements}
\usage{
rep_rate(to, from, formula, state, params, ratemat)
}
\arguments{
\item{to}{to state}

\item{from}{from state}

\item{formula}{one-sided formula defining the rate with reference
to the parameters and state variables}

\item{state}{state_pansim object}

\item{params}{param_pansim object}

\item{ratemat}{rate matrix}
}
\description{
Repeat a Rate for Several Rate Matrix Elements
}
