% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restore}
\alias{restore}
\title{convert a parameter vector back to a structured list
Like \code{\link{relist}}, but adds the capability to fill in
parameters from a "fixed-parameter" vector}
\usage{
restore(flesh, skeleton, fixed = NULL)
}
\arguments{
\item{flesh}{a vector to be restored}

\item{skeleton}{a list, the structure of which determines the structure of the result}

\item{fixed}{a list which determines extra components to fill in}
}
\description{
convert a parameter vector back to a structured list
Like \code{\link{relist}}, but adds the capability to fill in
parameters from a "fixed-parameter" vector
}
\note{
Depends at present on the names of the unlisted object; may be fragile
}
