% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfuns.R
\name{rExp}
\alias{rExp}
\title{run a pure-exponential sim;
uses run_sim_range with a population of 1 (proportions) and a very small starting value,
 run for 100 steps (by default)
used to calculate either r (technically r0) or eigenvector (for distributing initial exposed across states)}
\usage{
rExp(
  params,
  steps = 100,
  ndt = 1,
  do_hazard = FALSE,
  testify = has_testing(params = params),
  return_val = c("r0", "eigenvector", "sim"),
  type = "ICU1"
)
}
\arguments{
\item{params}{parameters}

\item{steps}{number of steps to run}

\item{ndt}{sub-time steps}

\item{do_hazard}{run with hazard}

\item{testify}{testing compartments}

\item{return_val}{return growth rate or eigenvector?}

\item{type}{model type (passed to \code{\link{make_state}})}
}
\description{
run a pure-exponential sim;
uses run_sim_range with a population of 1 (proportions) and a very small starting value,
 run for 100 steps (by default)
used to calculate either r (technically r0) or eigenvector (for distributing initial exposed across states)
}
\examples{
pp <- read_params("PHAC_testify.csv")
rExp(pp)
rExp(pp,return_val="eigenvector")
rExp(pp,return_val="eigenvector",testify=TRUE)
}
