% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfuns.R
\name{rExp}
\alias{rExp}
\title{run a pure-exponential sim;
uses run_sim_range with a population of 1 (proportions) and a very small starting value,
 run for 100 steps (by default)
used to calculate either r (technically r0) or eigenvector (for distributing initial exposed across states)}
\usage{
rExp(
  params,
  steps = getOption("MP_rexp_steps_default"),
  ndt = 1,
  do_hazard = FALSE,
  state = NULL,
  testify = has_testing(params = params),
  return_val = c("r0", "eigenvector", "sim"),
  type = "ICU1"
)
}
\arguments{
\item{params}{parameters}

\item{steps}{number of steps to run}

\item{ndt}{sub-time steps}

\item{do_hazard}{run with hazard}

\item{state}{initial state as created by make_state. Leave as NULL to create state from params (??)}

\item{testify}{testing compartments}

\item{return_val}{return growth rate or eigenvector?}

\item{type}{model type (passed to \code{\link{make_state}})}
}
\description{
run a pure-exponential sim;
uses run_sim_range with a population of 1 (proportions) and a very small starting value,
 run for 100 steps (by default)
used to calculate either r (technically r0) or eigenvector (for distributing initial exposed across states)
}
\examples{
pp <- read_params("PHAC_testify.csv")
rExp(pp)
rExp(pp,return_val="eigenvector")
rExp(pp,return_val="eigenvector",testify=TRUE)
}
\seealso{
Other classic_macpan: 
\code{\link{add_d_log}()},
\code{\link{add_updated_vaxrate}()},
\code{\link{aggregate_agecats}()},
\code{\link{calibrate_comb}()},
\code{\link{calibrate}()},
\code{\link{check_age_cat_compatibility}()},
\code{\link{check_contact_rate_setting}()},
\code{\link{col_multiply}()},
\code{\link{condense_age}()},
\code{\link{condense_params_vax}()},
\code{\link{condense_state}()},
\code{\link{condense_vax}()},
\code{\link{dev_is_tikz}()},
\code{\link{do_step}()},
\code{\link{expand_params_age}()},
\code{\link{expand_params_desc_age}()},
\code{\link{expand_params_desc_variant}()},
\code{\link{expand_params_desc_vax}()},
\code{\link{expand_params_mistry}()},
\code{\link{expand_params_variant}()},
\code{\link{expand_params_vax}()},
\code{\link{expand_state_age}()},
\code{\link{expand_state_vax}()},
\code{\link{expand_stateval_testing}()},
\code{\link{fix_pars}()},
\code{\link{fix_stored}()},
\code{\link{forecast_ensemble}()},
\code{\link{forecast_sim}()},
\code{\link{getData}()},
\code{\link{get_GI_moments}()},
\code{\link{get_Gbar}()},
\code{\link{get_R0}()},
\code{\link{get_doses_per_day}()},
\code{\link{get_evec}()},
\code{\link{get_kernel_moments}()},
\code{\link{get_opt_pars}()},
\code{\link{get_r}()},
\code{\link{invlink_trans}()},
\code{\link{make_betavec}()},
\code{\link{make_beta}()},
\code{\link{make_jac}()},
\code{\link{make_ratemat}()},
\code{\link{make_state}()},
\code{\link{make_test_wtsvec}()},
\code{\link{make_vaxrate}()},
\code{\link{mk_Nvec}()},
\code{\link{mk_agecats}()},
\code{\link{mk_contact_rate_setting}()},
\code{\link{mk_mistry_Nvec}()},
\code{\link{mk_pmat}()},
\code{\link{mk_vaxcats}()},
\code{\link{mle_fun}()},
\code{\link{non_expanded_states}},
\code{\link{read_params}()},
\code{\link{repair_names_age}()},
\code{\link{restore}()},
\code{\link{run_sim_ageify}()},
\code{\link{run_sim_break}()},
\code{\link{run_sim_loglin}()},
\code{\link{run_sim_mobility}()},
\code{\link{run_sim_range}()},
\code{\link{run_sim}()},
\code{\link{show_ratemat}()},
\code{\link{testify}()},
\code{\link{texify}()},
\code{\link{trans_state_vars}()},
\code{\link{update_contact_rate_setting}()},
\code{\link{update_foi}()},
\code{\link{update_params_mistry}()},
\code{\link{vis_model}()},
\code{\link{write_params}()}
}
\concept{classic_macpan}
