% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{row_col_mult}
\alias{row_col_mult}
\alias{row_mult}
\alias{col_mult}
\alias{row_mult,struc,struc-method}
\alias{col_mult,struc,struc-method}
\title{Symbolic Row and Column Multiplication}
\usage{
row_mult(x, y)

col_mult(x, y)

\S4method{row_mult}{struc,struc}(x, y)

\S4method{col_mult}{struc,struc}(x, y)
}
\arguments{
\item{x}{\code{\link{struc}} object playing the role of a matrix}

\item{y}{\code{\link{struc}} object playing the role of a vector}
}
\description{
Element-wise multiplication of the rows/columns of a matrix by
a vector of the same length as each row/column.
}
