% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{run_sim}
\alias{run_sim}
\title{Run pandemic simulation}
\usage{
run_sim(
  params,
  state = NULL,
  start_date = "2020-03-20",
  end_date = "2020-05-1",
  params_timevar = NULL,
  dt = 1,
  ndt = 1,
  stoch = c(obs = FALSE, proc = FALSE),
  stoch_start = NULL,
  ratemat_args = NULL,
  step_args = list(),
  ode_args = list(),
  use_ode = FALSE,
  condense = TRUE,
  condense_args = NULL,
  verbose = FALSE,
  use_flex = FALSE,
  flexmodel = NULL,
  obj_fun = NULL
)
}
\arguments{
\item{params}{named vector of parameters}

\item{state}{named vector of states}

\item{start_date}{starting date (Date or character, any sensible D-M-Y format)}

\item{end_date}{ending date (ditto)}

\item{params_timevar}{four-column data frame containing columns 'Date'; 'Symbol' (parameter name/symbol); 'Value'; and 'Type' (\code{rel_orig} (relative to value at time zero); \code{rel_prev} (relative to previous value); or \code{abs} (absolute value at specified time). If a 'Relative_value' column is present, it is renamed to 'Value' and 'Type' is set to \code{rel_orig} for back-compatibility.}

\item{dt}{time step for \code{\link{do_step}}}

\item{ndt}{number of internal time steps per time step}

\item{stoch}{a logical vector with elements "obs" (add obs error?) and "proc" (add process noise?)}

\item{stoch_start}{dates on which to enable stochasticity (vector of dates with names 'proc' and 'obs')}

\item{ratemat_args}{additional arguments to pass to \code{\link{make_ratemat}}}

\item{step_args}{additional arguments to pass to \code{\link{do_step}}}

\item{ode_args}{additional arguments to \code{\link[deSolve]{ode}}}

\item{use_ode}{integrate via ODE rather than discrete step?}

\item{condense}{if \code{TRUE}, use \code{\link{condense.pansim}} to reduce the number of variables in the output (in particular, collapse subclasses and return only one \code{I}, \code{H}, and \code{ICU} variable)}

\item{condense_args}{arguments to pass to \code{\link{condense}} (before adding observation error)}

\item{verbose}{print messages (e.g. about time-varying parameters)?}

\item{use_flex}{use \code{flexmodel} approach}

\item{flexmodel}{optional \code{\link{flexmodel}} object}

\item{obj_fun}{optional TMB objective function computed with
\code{\link{tmb_fun}}}
}
\description{
Run pandemic simulation
}
\examples{
params <- read_params("ICU1.csv")
paramsS <- update(params,c(proc_disp=0.1,obs_disp=100))
paramsSz <- update(paramsS, zeta=5)
state <- make_state(params=params)
time_pars <- data.frame(Date=c("2020-03-20","2020-03-25"),
                       Symbol=c("beta0","beta0"),
                       Relative_value=c(0.7,0.1),
                       stringsAsFactors=FALSE)
res1 <- run_sim(params,state,start_date="2020-02-01",end_date="2020-06-01")
res1X <- run_sim(params,state,start_date="2020-02-01",end_date="2020-06-01",
                 condense_args=list(keep_all=TRUE))
res1_S <- update(res1, params=paramsS, stoch=c(obs=TRUE, proc=TRUE))
res1_t <- update(res1, params_timevar=time_pars)
res1_S_t <- update(res1_S, params_timevar=time_pars)
res2_S_t <- update(res1_S_t,params=update(paramsS, proc_disp=0.5))
res3_S_t <- update(res2_S_t,stoch_start="2020-04-01")
res3_Sz <- update(res1_S, params=paramsSz)
plot(res3_Sz,log=TRUE,log_lwr=1e-4)
}
