% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{run_sim}
\alias{run_sim}
\title{Run pandemic simulation}
\usage{
run_sim(
  params,
  state = make_state(params[["N"]], params[["E0"]]),
  start_date = "20-Mar-2020",
  end_date = "1-May-2020",
  params_timevar = NULL,
  dt = 1,
  ndt = 1,
  stoch = c(obs = FALSE, proc = FALSE),
  ratemat_args = NULL,
  step_args = NULL
)
}
\arguments{
\item{params}{named vector of parameters}

\item{state}{named vector of states}

\item{start_date}{starting date (Date or character, any sensible D-M-Y format)}

\item{end_date}{ending date (ditto)}

\item{params_timevar}{three-column data frame containing columns 'Date'; 'Symbol' (parameter name/symbol); 'Relative_value' (value \emph{relative to baseline})}

\item{dt}{time step for do_step}

\item{ndt}{number of internal time steps per time step}

\item{stoch}{stochastic simulation? logical vector for observation and process noise}

\item{ratemat_args}{additional arguments to pass to \code{make_ratemat}}

\item{step_args}{additional arguments to pass to \code{do_step}}
}
\description{
Run pandemic simulation
}
\examples{
params <- read_params(system.file("params","ICU1.csv",package="McMasterPandemic"))
state <- make_state(params=params)
sdate <- "10-Feb-2020" ## arbitrary!
time_pars <- data.frame(Date=c("20-Mar-2020","25-Mar-2020"),
                       Symbol=c("beta0","beta0"),
                       Relative_value=c(0.7,0.1))
res <- run_sim(params,state,start_date=sdate,end_date="1-Jun-2020",
                   params_timevar=time_pars)
summary(res)
}
