% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{run_sim_break}
\alias{run_sim_break}
\title{run simulation with one or more breakpoints}
\usage{
run_sim_break(params, break_dates, rel_beta0, ...)
}
\arguments{
\item{params}{parameters}

\item{break_dates}{dates of breakpoints in transmission}

\item{rel_beta0}{numeric vector (same length as \code{break_dates}): transmission relative to original value after each breakpoint}

\item{...}{additional arguments to \code{run_sim}}
}
\description{
run simulation with one or more breakpoints
}
