% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{run_sim_mobility}
\alias{run_sim_mobility}
\title{run with transmission propto relative mobility}
\usage{
run_sim_mobility(
  params,
  extra_pars = NULL,
  time_args = NULL,
  sim_args = list(),
  return_timevar = FALSE,
  ...
)
}
\arguments{
\item{params}{parameters}

\item{extra_pars}{parameters that are used to set up time-varying parameters, etc., but \emph{not} used by \code{run_sim}:
**FIXME** should contain rel_<parname> values matching the break dates.  MAYBE rel_ or abs_ determines whether the value
is relative to baseline *or* is absolute? (will require corresponding change in run_sim to allow this)}

\item{time_args}{list containing \code{break_dates}, **FIXME and Symbol!}

\item{sim_args}{parameters to pass to \code{\link{run_sim}}}

\item{return_timevar}{return data frame of params by time?}

\item{...}{extra args (why??)}
}
\description{
run with transmission propto relative mobility
}
