% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{run_sim_mobility}
\alias{run_sim_mobility}
\title{run with transmission propto relative mobility}
\usage{
run_sim_mobility(
  params,
  extra_pars = NULL,
  time_args = NULL,
  sim_args = list(),
  return_timevar = FALSE,
  ...
)
}
\arguments{
\item{params}{parameters}

\item{extra_pars}{parameters that are used to set up time-varying parameters, etc., but \emph{not} used by \code{run_sim}}

\item{time_args}{list containing \code{break_dates}}

\item{sim_args}{parameters to pass to \code{\link{run_sim}}}

\item{return_timevar}{return data frame of beta by time?}

\item{...}{additional arguments to \code{run_sim}}
}
\description{
run with transmission propto relative mobility
}
