% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{run_sim_range}
\alias{run_sim_range}
\title{Run simulation across a range of times}
\usage{
run_sim_range(
  params,
  state = make_state(params[["N"]], params[["E0"]]),
  nt = 100,
  dt = 1,
  M = NULL,
  ratemat_args = NULL,
  step_args = NULL,
  use_ode = FALSE,
  ode_args = list()
)
}
\arguments{
\item{params}{named vector of parameters}

\item{state}{named vector of states}

\item{nt}{number of steps to take}

\item{dt}{time step (days)}

\item{M}{rate matrix}

\item{ratemat_args}{additional arguments to pass to \code{make_ratemat}}

\item{step_args}{additional arguments to pass to \code{do_step}}

\item{use_ode}{solve as differential equation?}

\item{ode_args}{additional arguments to ode()}
}
\description{
Run simulation across a range of times
}
\examples{
params <- read_params("ICU1.csv")
r1 <- run_sim_range(params)
r2 <- run_sim_range(params,use_ode=TRUE)
matplot(r1[,"t"],r1[,-1],type="l",lty=1,log="y")
matlines(r2[,"t"],r2[,-1],lty=2)
}
