% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{run_sim_range}
\alias{run_sim_range}
\title{Run simulation across a range of times}
\usage{
run_sim_range(
  params,
  state = make_state(params[["N"]], params[["E0"]]),
  nt = 100,
  dt = 1,
  M = NULL,
  stoch = c(obs = FALSE, proc = FALSE),
  ratemat_args = NULL,
  step_args = NULL
)
}
\arguments{
\item{params}{named vector of parameters}

\item{state}{named vector of states}

\item{nt}{number of steps to take}

\item{dt}{time step (days)}

\item{stoch}{stochastic simulation? logical vector for observation and process noise}

\item{ratemat_args}{additional arguments to pass to \code{make_ratemat}}

\item{step_args}{additional arguments to pass to \code{do_step}}
}
\description{
Run simulation across a range of times
}
\examples{
params <- read_params(system.file("params","ICU1.csv",package="McMasterPandemic"))
run_sim_range(params)
}
