% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{same_dims}
\alias{same_dims}
\title{Test if struc Objects have the Same Dimensions}
\usage{
same_dims(x, y)
}
\arguments{
\item{x, y}{struc objects}
}
\value{
TRUE or FALSE
}
\description{
Test if struc Objects have the Same Dimensions
}
