% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{set_spec_version}
\alias{set_spec_version}
\title{Set Spec Version}
\usage{
set_spec_version(v, cpp_path = NULL)
}
\arguments{
\item{v}{character string with spec version}

\item{cpp_path}{string containing path
to a directory containing a file called
\code{macpan.cpp}, which is to be used
to construct the objective function --
the default is \code{NULL}, which will
make use of the packaged source file
\code{McMasterPandemic.cpp}}
}
\description{
Set the spec version and optionally compile a
specific c++ file associated with this version.
}
\details{
The user need to have write permissions to
\code{cpp_path}, because object and shared
object files will be created in this
directory.
}
