% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_ratemat}
\alias{show_ratemat}
\title{visualize rate (per-capita flow) matrix}
\usage{
show_ratemat(
  M,
  method = c("Matrix", "diagram", "igraph"),
  aspect = "iso",
  add_blocks = NULL,
  blocksize = NULL,
  block_col = 2,
  axlabs = TRUE,
  const_width = (method == "igraph"),
  do_symbols = NULL,
  box.size = 0.02,
  ...
)
}
\arguments{
\item{M}{rate matrix}

\item{method}{visualization method}

\item{aspect}{aspect ratio ("iso", "fill" are the sensible options)}

\item{add_blocks}{add lines showing blocks for testified matrices?}

\item{axlabs}{for flow matrices, show axis tick labels?}

\item{const_width}{set flows to constant value of 1?}

\item{do_symbols}{plot symbolic values for flows?}

\item{box.size}{box size for diagram}

\item{...}{arguments to pass to lower level functions (plotmat::diagram/image/igraph)}
}
\description{
visualize rate (per-capita flow) matrix
}
\examples{
params <- read_params("ICU1.csv")
state <- make_state(params[["N"]],E0=params[["E0"]], use_eigvec=FALSE)
M <- make_ratemat(state, params)
show_ratemat(M)
## silly but shows we can do multiple block types in different colours
show_ratemat(M, add_blocks=TRUE, blocksize=c(3,5), block_col=c(2,4))
}
