% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_ratemat}
\alias{show_ratemat}
\title{visualize rate (per-capita flow) matrix}
\usage{
show_ratemat(
  M,
  method = c("Matrix", "diagram", "igraph"),
  subset = NULL,
  xlab = "to",
  ylab = "from",
  sub = "",
  zlim = c(0, 1),
  aspect = "iso",
  block_size = NULL,
  block_col = 2,
  axlabs = TRUE,
  const_width = (method == "igraph"),
  colour_palette = viridis::magma(n = 50, direction = -1),
  do_symbols = NULL,
  box.size = 0.02,
  ...
)
}
\arguments{
\item{M}{rate matrix}

\item{method}{visualization method}

\item{subset}{list of two regular expressions, the first to subset rate matrix rows (based on rownames) and the second to subset columns (based on colnames)}

\item{xlab}{label on x axis}

\item{ylab}{label on y axis}

\item{sub}{subtitle for plot}

\item{zlim}{appears to be unused (!!)}

\item{aspect}{aspect ratio ("iso", "fill" are the sensible options)}

\item{block_size}{numeric vector of number of compartments per block; if NA, try to guess from number of epidemiological compartments}

\item{block_col}{each element in block_col controls the color of the corresponding grid overlay added by block_size (if add_blocks==TRUE)}

\item{axlabs}{for flow matrices, show axis tick labels?}

\item{const_width}{set flows to constant value of 1?}

\item{colour_palette}{vector of colours for rate matrix heatmap}

\item{do_symbols}{plot symbolic values for flows?}

\item{box.size}{box size for diagram}

\item{...}{arguments to pass to lower level functions (plotmat::diagram/image/igraph)}
}
\description{
visualize rate (per-capita flow) matrix
}
\examples{
params <- read_params("ICU1.csv")
state <- make_state(params[["N"]],E0=params[["E0"]], use_eigvec=FALSE)
M <- make_ratemat(state, params)
show_ratemat(M)
## silly but shows we can do multiple block types in different colours
show_ratemat(M, block_size=c(3,5), block_col=c(2,4))
}
\seealso{
Other classic_macpan: 
\code{\link{add_d_log}()},
\code{\link{add_updated_vaxrate}()},
\code{\link{aggregate_agecats}()},
\code{\link{calibrate_comb}()},
\code{\link{calibrate}()},
\code{\link{check_age_cat_compatibility}()},
\code{\link{check_contact_rate_setting}()},
\code{\link{col_multiply}()},
\code{\link{condense_age}()},
\code{\link{condense_params_vax}()},
\code{\link{condense_state}()},
\code{\link{condense_vax}()},
\code{\link{dev_is_tikz}()},
\code{\link{do_step}()},
\code{\link{expand_params_age}()},
\code{\link{expand_params_desc_age}()},
\code{\link{expand_params_desc_variant}()},
\code{\link{expand_params_desc_vax}()},
\code{\link{expand_params_mistry}()},
\code{\link{expand_params_variant}()},
\code{\link{expand_params_vax}()},
\code{\link{expand_state_age}()},
\code{\link{expand_state_vax}()},
\code{\link{expand_stateval_testing}()},
\code{\link{fix_pars}()},
\code{\link{fix_stored}()},
\code{\link{forecast_ensemble}()},
\code{\link{forecast_sim}()},
\code{\link{getData}()},
\code{\link{get_GI_moments}()},
\code{\link{get_Gbar}()},
\code{\link{get_R0}()},
\code{\link{get_doses_per_day}()},
\code{\link{get_evec}()},
\code{\link{get_kernel_moments}()},
\code{\link{get_opt_pars}()},
\code{\link{get_r}()},
\code{\link{invlink_trans}()},
\code{\link{make_betavec}()},
\code{\link{make_beta}()},
\code{\link{make_jac}()},
\code{\link{make_ratemat}()},
\code{\link{make_state}()},
\code{\link{make_test_wtsvec}()},
\code{\link{make_vaxrate}()},
\code{\link{mk_Nvec}()},
\code{\link{mk_agecats}()},
\code{\link{mk_contact_rate_setting}()},
\code{\link{mk_mistry_Nvec}()},
\code{\link{mk_pmat}()},
\code{\link{mk_vaxcats}()},
\code{\link{mle_fun}()},
\code{\link{non_expanded_states}},
\code{\link{rExp}()},
\code{\link{read_params}()},
\code{\link{repair_names_age}()},
\code{\link{restore}()},
\code{\link{run_sim_ageify}()},
\code{\link{run_sim_break}()},
\code{\link{run_sim_loglin}()},
\code{\link{run_sim_mobility}()},
\code{\link{run_sim_range}()},
\code{\link{run_sim}()},
\code{\link{testify}()},
\code{\link{texify}()},
\code{\link{trans_state_vars}()},
\code{\link{update_contact_rate_setting}()},
\code{\link{update_foi}()},
\code{\link{update_params_mistry}()},
\code{\link{vis_model}()},
\code{\link{write_params}()}
}
\concept{classic_macpan}
