% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_ratemat}
\alias{show_ratemat}
\title{visualize rate (per-capita flow) matrix}
\usage{
show_ratemat(
  M,
  method = c("Matrix", "diagram", "igraph"),
  subset = NULL,
  xlab = "to",
  ylab = "from",
  sub = "",
  zlim = c(0, 1),
  aspect = "iso",
  add_blocks = NULL,
  blocksize = NULL,
  block_col = 2,
  axlabs = TRUE,
  const_width = (method == "igraph"),
  colour_palette = viridis::magma(n = 50, direction = -1),
  do_symbols = NULL,
  box.size = 0.02,
  ...
)
}
\arguments{
\item{M}{rate matrix}

\item{method}{visualization method}

\item{subset}{list of two regular expressions, the first to subset rate matrix rows (based on rownames) and the second to subset columns (based on colnames)}

\item{aspect}{aspect ratio ("iso", "fill" are the sensible options)}

\item{add_blocks}{add lines showing blocks for testified matrices?}

\item{axlabs}{for flow matrices, show axis tick labels?}

\item{const_width}{set flows to constant value of 1?}

\item{colour_palette}{vector of colours for rate matrix heatmap}

\item{do_symbols}{plot symbolic values for flows?}

\item{box.size}{box size for diagram}

\item{...}{arguments to pass to lower level functions (plotmat::diagram/image/igraph)}
}
\description{
visualize rate (per-capita flow) matrix
}
