% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_ratemat}
\alias{show_ratemat}
\title{visualize rate (per-capita flow) matrix}
\usage{
show_ratemat(M, aspect = "iso", add_blocks = NULL)
}
\arguments{
\item{M}{rate matrix}

\item{aspect}{aspect ratio ("iso", "fill" are the sensible options)}

\item{add_blocks}{add lines showing blocks for testified matrices?}
}
\description{
visualize rate (per-capita flow) matrix
}
