% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbutils.R
\name{simulate_ensemble}
\alias{simulate_ensemble}
\title{Ensemble Simulation}
\usage{
simulate_ensemble(
  model,
  sim_params_matrix = NULL,
  covmat = NULL,
  qvec = c(value = 0.5, lwr = 0.025, upr = 0.975),
  use_progress_bar = TRUE,
  ...
)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object, preferably a
\code{flexmodel_calibrated} object that was fitted with
\code{\link{bbmle_flexmodel}} or \code{\link{calibrate_flexmodel}}}

\item{sim_params_matrix}{numberic matrix with rows giving simulation
iterations and columns giving model parameters}

\item{covmat}{not used}

\item{qvec}{named numeric vector giving quantiles to compute}

\item{...}{arguments to pass on the \code{pop_pred_samp}, such as
\code{PDify = TRUE} that can solve issues with non-positive definite
matrices}
}
\description{
Given a sample of parameter vectors, simulate the condensed state simulation
history for each ... TODO
}
\seealso{
Other simulation: 
\code{\link{simulation_history}()}
}
\concept{simulation}
