% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\docType{class}
\name{struc-class}
\alias{struc-class}
\title{Class to Represent Matrix Structure}
\description{
Class to Represent Matrix Structure
}
\section{Slots}{

\describe{
\item{\code{v}}{character vector giving the expressions for each element
of the matrix structure object}

\item{\code{dims}}{numeric vector giving the dimensions of the matrix
structure object}
}}

