% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstruc.R
\name{struc_expanded}
\alias{struc_expanded}
\title{alter.EpiMatrixListDerived = function(mat_list, ...) {
  stop("calls to _alter_ must come before all calls to _derive_")
}}
\usage{
struc_expanded(l, d)
}
\description{
alter.EpiMatrixList = function(mat_list, ...) {
  valid_funcs = getOption("MP_valid_alter_funcs")
  new_mats = list(...)
  for (i in seq_along(new_mats)) {
    mat_i = eval(
      new_mats[[i]][[2L]],
      c(mat_list, valid_funcs),
      enclos = emptyenv()
    )
    new_mats[[i]] = epi_mat(
      names(new_mats[i]),
      as.matrix(mat_i)
    )
  }
  structure(
    c(mat_list, new_mats),
    class = c("EpiMatrixList")
  )
}
}
