% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testify.R
\name{testify}
\alias{testify}
\title{expand rate matrix for testing status}
\usage{
testify(
  ratemat,
  params,
  debug = FALSE,
  non_expand_set = c("D", "R", "X"),
  testing_time = NULL
)
}
\arguments{
\item{ratemat}{original rate matrix}

\item{params}{parameters}

\item{debug}{what it sounds like}

\item{non_expand_set}{states \emph{not} to expand}

\item{testing_time:}{"report" (N and P are counted at the time when individuals move from _n, _p to _u, _t compartments) or "sample" (N and P are counted when individuals move from _u to _n or _p)}
}
\description{
expand rate matrix for testing status
}
