% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testify.R
\name{testify}
\alias{testify}
\title{expand rate matrix for testing status}
\usage{
testify(ratemat, params, debug = FALSE, non_expand_set = c("D", "R", "X"))
}
\arguments{
\item{ratemat}{original rate matrix}

\item{params}{parameters}

\item{debug}{what it sounds like}

\item{non_expand_set}{states \emph{not} to expand}
}
\description{
expand rate matrix for testing status
}
