% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{tmb_fun}
\alias{tmb_fun}
\title{Make Objective Function with TMB}
\usage{
tmb_fun(model)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}
}
\description{
Construct an objective function in TMB from a \code{flexmodel}
object. The behaviour of \code{tmb_fun} depends on \code{spec_version()}
}
