% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{tmb_fun}
\alias{tmb_fun}
\title{Make Objective Function with TMB}
\usage{
tmb_fun(model)
}
\arguments{
\item{model}{object of class \code{flexmodel}}
}
\description{
Construct an objective function in TMB from a \code{flexmodel}
object. The behaviour of \code{tmb_fun} depends on \code{spec_version()}
}
\seealso{
Other flexmodels: 
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_tmb_indices}()},
\code{\link{init_model}()},
\code{\link{initial_population}()},
\code{\link{make_base_model}()},
\code{\link{make_vaccination_model}()},
\code{\link{mat_rate}()},
\code{\link{rate}()},
\code{\link{rep_rate}()},
\code{\link{state_param_sum}()},
\code{\link{vec_rate}()}
}
\concept{flexmodels}
