% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_params}
\alias{update_params}
\title{Update Default Parameters}
\usage{
update_params(model, ...)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{...}{named vectors of parameter names}
}
\description{
Update Default Parameters
}
