% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_piece_wise}
\alias{update_piece_wise}
\title{Update the Piece-Wise Parameter Time-Variation Schedule}
\usage{
update_piece_wise(model, params_timevar, regenerate_rates = TRUE)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{params_timevar}{data frame with scheduling for piece-wise
constant parameter variation (TODO: direct to other help pages)}

\item{regenerate_rates}{should the rates in the flexmodel be
regenerated? If you don't know what you are doing, you should
select \code{TRUE}, which is the default.}
}
\description{
Update the Piece-Wise Parameter Time-Variation Schedule
}
