% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_tmb_indices}
\alias{update_tmb_indices}
\alias{add_tmb_indices}
\alias{tmb_indices}
\title{Update TMB Indices}
\usage{
update_tmb_indices(model)

add_tmb_indices(model)

tmb_indices(model)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}
}
\description{
Add or update indices used to access appropriate values
during simulation and calibration using TMB
}
\details{
\describe{
  \item{\code{make_ratemat_indices$from}}{}
  \item{\code{make_ratemat_indices$to}}{}
  \item{\code{make_ratemat_indices$count}}{}
  \item{\code{updateidx}}{indices into the \code{to}, \code{from}, and \code{count}
  vectors, identifying rates that need (or at least will) be updated at
  every simulation step}
}
}
\seealso{
Other flexmodel_definition_functions: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{initial_population}()},
\code{\link{mat_rate}()},
\code{\link{rep_rate}()},
\code{\link{update_disease_free_state}()},
\code{\link{update_linearized_params}()},
\code{\link{vec_rate}()}
}
\concept{flexmodel_definition_functions}
