% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{update_tmb_indices}
\alias{update_tmb_indices}
\alias{add_tmb_indices}
\title{Update TMB Indices}
\usage{
update_tmb_indices(model)

add_tmb_indices(model)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}
}
\description{
Add or update indices used to access appropriate values
during simulation and calibration using TMB
}
\details{
\describe{
  \item{\code{make_ratemat_indices$from}}{}
  \item{\code{make_ratemat_indices$to}}{}
  \item{\code{make_ratemat_indices$count}}{}
  \item{\code{updateidx}}{indices into the \code{to}, \code{from}, and \code{count}
  vectors, identifying rates that need (or at least will) be updated at
  every simulation step}
}
}
