% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{vec_factr}
\alias{vec_factr}
\title{Vectors of Intermediate Factors}
\usage{
vec_factr(model, factr_nms, formula)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{factr_nms}{vector of names of the new intermediate factors}

\item{formula}{\code{\link{struc-class}} object describing the
vector of intermediate factors. See \code{\link{avail_for_factr}} for
a function that will return the names of all variables that are available
for use in these \code{\link{struc-class}} objects.}
}
\value{
updated \code{\link{flexmodel}} object
}
\description{
Vectors of Intermediate Factors
}
\seealso{
See \code{\link{add_factr}} to add a single
scalar-valued intermediate factor and \code{\link{vec}}
to create a vector-valued \code{\link{struc}} object.
}
