% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb.R
\name{vec_rate}
\alias{vec_rate}
\title{Specify Vector of Rates}
\usage{
vec_rate(model, from, to, formula)
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{from}{character vector defining states from which flow is coming}

\item{to}{character vector defining states from which flow is going}

\item{formula}{\code{\link{struc-class}} object defining a vector of flows
for each \code{from-to} pair. See \code{\link{avail_for_rate}} for a
function that will print out the names of variables that are available for
use in these \code{\link{struc-class}} objects.}

\item{mapping}{experimental -- please choose default for now}
}
\description{
Specify Vector of Rates
}
\seealso{
Other flexmodel_definition_functions: 
\code{\link{add_factr}()},
\code{\link{add_linearized_outflow}()},
\code{\link{add_parallel_accumulators}()},
\code{\link{add_rate}()},
\code{\link{add_state_mappings}()},
\code{\link{add_state_param_sum}()},
\code{\link{initial_population}()},
\code{\link{mat_rate}()},
\code{\link{rep_rate}()},
\code{\link{update_disease_free_state}()},
\code{\link{update_linearized_params}()},
\code{\link{update_tmb_indices}()}

Other rate_functions: 
\code{\link{add_rate}()},
\code{\link{mat_rate}()},
\code{\link{rep_rate}()}
}
\concept{flexmodel_definition_functions}
\concept{rate_functions}
